+++ .gitignore
@@ -96,3 +96,4 @@ ENV/
 
 # Content generated.
 portal/pages/*
+portal/menus/*
