+++ portal/portal/sitemap_helper.py
@@ -74,6 +74,10 @@ def get_menu_path_cache(content_id, lang, version):
     return menu_path
 
 
+def get_production_menu_path(content_id, lang, version):
+    return os.path.join(settings.MENUS_DIR, '%s/%s/%s/menu.json' % (content_id, lang, version))
+
+
 def find_in_top_level_navigation(path):
     for i in settings.TOP_LEVEL_NAVIGATION:
         if i['path'] == path:
@@ -96,84 +100,69 @@ def get_top_level_navigation(version, language):
     """
     Returns a list of categories available for this version & language.
     """
+    if settings.ENV in ['production', 'staging']:
+        return [i for i in settings.TOP_LEVEL_NAVIGATION if os.path.exists(
+            get_production_menu_path(i['path'][1:], language, version)
+        )]
+
     return filter(lambda i: i['dir'] is not None, settings.TOP_LEVEL_NAVIGATION)
 
 
-def get_sitemap(version, language, content_id):
+def get_sitemap(content_id, lang, version):
     """
     Given a version and language, fetch the sitemap for all contents from the
     cache, if available, or load them from the pre-compiled sitemap.
+
+    [For now] Returns a freshly generated sitemap file, given a version and language.
     """
     # cache_key = 'menu.%s.%s' % (version, language)
     # sitemap_cache = cache.get(cache_key, None)
 
+    # if sitemap_cache:
+    #     cache.set(cache_key, sitemap_cache)
+    # else:
+    #     raise Exception('Cannot generate sitemap for version %s' % version)
+
+    sitemap = None
+
     # These are the repos without menus.
     if content_id in ['models', 'mobile']:
         # NOTE(varunarora): This is a hack because there is no menu.json here.
-        return None, _get_sitemap_path('', content_id)
+        sitemap_path = None if settings.ENV in ['production', 'staging'] else _get_sitemap_path('', content_id)
 
     else:
-        # if not sitemap_cache:
-        return _load_sitemap_from_file(version, language, content_id)
+        if settings.ENV in ['production', 'staging']:
+            sitemap_path = get_production_menu_path(content_id, lang, version)
+        else:
+            sitemap_path = _get_sitemap_path('menu.json', content_id)
 
-        # if sitemap_cache:
-        #     cache.set(cache_key, sitemap_cache)
-        # else:
-        #     raise Exception('Cannot generate sitemap for version %s' % version)
+        if not sitemap_path:
+            raise Exception('Cannot find a sitemap file with the name %s in the directory for: %s' % (sitemap_path, content_id))
 
+        if os.path.isfile(sitemap_path):
+            set_menu_path_cache(content_id, lang, version, sitemap_path)
 
-def _load_sitemap_from_file(version, language, content_id):
-    """
-    [For now] Returns a freshly generated sitemap file, given a version and language.
-    """
-    sitemap = None
+            # Sitemap file exists, lets load it
+            try:
+                with open(sitemap_path) as json_data:
+                    sitemap = json.loads(json_data.read())
 
-    #sitemap_filename = ('sitemap.%s.json' % language)
-    sitemap_path = _get_sitemap_path('menu.json', content_id)
+                    # cache.set(get_all_links_cache_key(version, language), sitemap['all_links_cache'], None)
 
-    if not sitemap_path:
-        raise Exception('Cannot find a sitemap file with the name %s in the directory for: %s' % (sitemap_path, content_id))
-
-    if os.path.isfile(sitemap_path):
-        set_menu_path_cache(content_id, language, version, sitemap_path)
-
-        # Sitemap file exists, lets load it
-        try:
-            with open(sitemap_path) as json_data:
-                print 'Loading sitemap from %s' % sitemap_path
-                sitemap = json.loads(json_data.read())
-
-                # sitemap = json.loads(json_data.read(), object_pairs_hook=collections.OrderedDict)
-                # cache.set(get_all_links_cache_key(version, language), sitemap['all_links_cache'], None)
-
-        except Exception as e:
-            print 'Cannot load sitemap from file %s: %s' % (sitemap_path, e.message)
-
-    # if not sitemap:
-    #     # We couldn't load sitemap.<version>.json file, lets generate it
-    #     sitemap = generate_sitemap(version, language)
+            except Exception as e:
+                print 'Cannot load sitemap from file %s: %s' % (sitemap_path, e.message)
 
     return sitemap, sitemap_path
 
 
-
 def _transform_section_urls(section, prefix, content_id):
-    # """
-    # Since paths defined in assets/sitemaps/<version>.json are defined relative to the folder structure of the content
-    # directories, we will need to append the URL path prefix so our URL router knows how to resolve the URLs.
-    #
-    # ex:
-    # /documentation/en/getstarted/index_en.html -> /docs/<version>/documentation/en/gettingstarted/index_en.html
-    # /book/01.fit_a_line/index.html -> /docs/<version>/book/01.fit_a_line/index.html
-    # """
-    # all_links_cache = {}
-
+    """
+    Since paths defined in assets/sitemaps/<version>.json are defined relative to the folder structure of the content
+    directories, we will need to append the URL path prefix so our URL router knows how to resolve the URLs.
+    """
     # Make a copy that we shall mutate.
-
     new_transformed_sections = []
     for subsection in section['sections']:
-        # if book and DEFAULT_BRANCH in book and version != 'doc_test':
-        #     version = book[DEFAULT_BRANCH]
         new_subsection = {}
 
         for key, value in subsection.items():
@@ -195,50 +184,22 @@ def _transform_section_urls(section, prefix, content_id):
             else:
                 new_subsection[key] = value
 
-            # for category_data in book['categories'].itervalues():
-            #_transform_urls(version, sitemap, category_data, all_links_cache, language)
-
         new_transformed_sections.append(new_subsection)
 
     return new_transformed_sections
-    # sitemap['all_links_cache'] = all_links_cache
-    # cache.set(get_all_links_cache_key(version, language), all_links_cache, None)
 
 
-
-def get_content_navigation(request, content_id, version, language):
+def get_content_navigation(request, content_id, language, version):
     """
     Get the navigation sitemap for a particular content service.
     """
-    # from portal import portal_helper
-
-    # category_data = None
     navigation = {'sections': _transform_section_urls(
-        get_sitemap(version, language, content_id)[0],
+        get_sitemap(content_id, language, version)[0],
         url_helper.get_page_url_prefix(content_id, language, version),
         content_id
     )}
 
-    # book = root_nav.get(content_id, None)
-    # Go through all links and give them relative paths from the root.
-
     return navigation
-
-    # if book:
-    #     category = book['default-category']
-    #
-    #     if content_id == portal_helper.Content.DOCUMENTATION or \
-    #                     content_id == portal_helper.Content.API:
-    #         # For Documentation or API, we also filter by category
-    #         api_category = portal_helper.get_preferred_api_version(request)
-    #         if api_category:
-    #             category = api_category
-    #
-    #     category_data = book['categories'][category]
-    #
-    # return category_data, category
-
-
 
 
 def _get_sitemap_path(sitemap_filename, content_id):
@@ -248,8 +209,6 @@ def _get_sitemap_path(sitemap_filename, content_id):
     repo_path = find_in_top_level_navigation('/' + content_id)
 
     if os.path.exists(repo_path['dir']):
-        # os.makedirs(settings.RESOLVED_SITEMAP_DIR)
-        # os.chmod(settings.RESOLVED_SITEMAP_DIR, 0775)
         if content_id == 'api':
             sitemap_filename = 'api/' + sitemap_filename
 
@@ -257,8 +216,6 @@ def _get_sitemap_path(sitemap_filename, content_id):
 
     raise Exception('Cannot find the directory for %s: %s' % (
         content_id, repo_path['dir']))
-
-    #return '%s/sitemap.%s.%s.json' % (settings.RESOLVED_SITEMAP_DIR, version, language)
 
 
 def get_available_versions(content_id=None):
@@ -283,13 +240,6 @@ def get_available_versions(content_id=None):
 
     if versions:
         for version in versions:
-            # if content_id:
-            #     folder_path = '%s/%s/%s' % (path, version, content_id)
-            #     if not os.path.isdir(folder_path):
-            #         # If content_id folder does not exists in versioned directory,
-            #         # then don't add it to list of available versions
-            #         continue
-
             normalized_version = version.split('.')
             if len(normalized_version) > 1:
                 number_based_version.append(version)
@@ -304,169 +254,5 @@ def get_available_versions(content_id=None):
     return string_based_version + number_based_version
 
 
-def is_version_greater_eq(v1, v2):
-    f = lambda s: list(map(int, s.split('.')))
-    if v1 == 'develop':
-        return True
-
-    try:
-        return f(v1) >= f(v2)
-    except:
-        return False
-
-
 def get_external_file_path(sub_path):
-    #return '%s/%s' % (settings.EXTERNAL_TEMPLATE_DIR, sub_path)
     return os.path.join(settings.WORKSPACE_DIR, sub_path)
-
-
-##################################################################
-
-
-DEFAULT_BRANCH = 'default-branch'
-
-def generate_sitemap(version, language):
-    """
-    Using a sitemap template, generated a full sitemap using individual content
-    sitemaps.
-    """
-    sitemap = None
-    sitemap_template_path = '%s/assets/sitemaps/sitemap_tmpl.json' % settings.PROJECT_ROOT
-
-    try:
-        # Read the sitemap template.
-        with open(sitemap_template_path) as json_data:
-            sitemap = json.loads(json_data.read(), object_pairs_hook=collections.OrderedDict)
-
-            # Resolve JSON references with contents' individual sitemaps.
-            sitemap = _resolve_references(sitemap, version, language)
-
-            # Change URLs to represent accurate URL paths and not references to repo directory structures.
-            _transform_sitemap_urls(version, sitemap, language)
-
-            sitemap_path = _get_sitemap_path(version, language)
-
-        # Write the built sitemaps to the main sitemap file the app reads.
-        with open(sitemap_path, 'w') as fp:
-            json.dump(sitemap, fp)
-            # Enable the write permissions so the deploy_docs scripts can delete the sitemaps to force updates.
-            os.chmod(sitemap_path, 0664)
-
-    except Exception as e:
-        print 'Cannot generate sitemap from %s: %s' % (sitemap_template_path, e.message)
-        traceback.print_exc()
-
-    return sitemap
-
-
-# def _transform_urls(version, sitemap, node, all_links_cache, language):
-#     all_node_links = []
-#
-#     if sitemap and node:
-#         if 'link' in node and language in node['link']:
-#             transformed_path = node['link'][language]
-#
-#             if not transformed_path.startswith('http'):
-#                 # We only append the document root/version if this is not an absolute URL
-#                 path_with_prefix = url_helper.append_prefix_to_path(version, transformed_path)
-#                 if path_with_prefix:
-#                     transformed_path = path_with_prefix
-#
-#             if transformed_path:
-#                 node['link'][language] = transformed_path
-#
-#             all_node_links.append(transformed_path)
-#
-#             if all_links_cache != None:
-#                 key = url_helper.link_cache_key(transformed_path)
-#                 all_links_cache[key] = transformed_path
-#
-#         if 'sections' in node:
-#             for child_node in node['sections']:
-#                 child_node_links = _transform_urls(version, sitemap, child_node, all_links_cache, language)
-#                 all_node_links.extend(child_node_links)
-#
-#         node['links'] = all_node_links
-#         if ('link' not in node or language not in node['link']):
-#             # After we process the node's children, we check if the node has a default link.
-#             # If not, then we set the node's first link
-#             if len(all_node_links) > 0:
-#                 node['link'] = {
-#                     language: all_node_links[0]
-#                 }
-#
-#     return all_node_links
-
-
-def load_json_and_resolve_references(path, version, language):
-    """
-    Loads any sitemap file (content root or site's root sitemap), and resolves
-    references to generate a combined sitemap dictionary.
-    """
-    sitemap = None
-    sitemap_path = '%s/docs/%s/%s' % (settings.EXTERNAL_TEMPLATE_DIR, version, path)
-
-    try:
-        with open(sitemap_path) as json_data:
-            sitemap = json.loads(json_data.read(), object_pairs_hook=collections.OrderedDict)
-
-        # Resolve any reference in inner sitemap files.
-        sitemap = _resolve_references(sitemap, version, language)
-    except Exception as e:
-        print 'Cannot resolve sitemap from %s: %s' % (sitemap_path, e.message)
-
-    return sitemap
-
-
-def _resolve_references(navigation, version, language):
-    """
-    Iterates through an object (could be a dict, list, str, int, float, unicode, etc.)
-    and if it finds a dict with `$ref`, resolves the reference by loading it from
-    the respective JSON file.
-    """
-    if isinstance(navigation, list):
-        # navigation is type list, resolved_navigation should also be type list
-        resolved_navigation = []
-
-        for item in navigation:
-            resolved_navigation.append(_resolve_references(item, version, language))
-
-        return resolved_navigation
-
-    elif isinstance(navigation, dict):
-        # navigation is type dict, resolved_navigation should also be type dict
-        resolved_navigation = collections.OrderedDict()
-
-        if DEFAULT_BRANCH in navigation and version != 'doc_test':
-            version = navigation[DEFAULT_BRANCH]
-
-        for key, value in navigation.items():
-            if key == '$ref' and language in value:
-                # The value is the relative path to the associated json file
-                referenced_json = load_json_and_resolve_references(value[language], version, language)
-                if referenced_json:
-                    resolved_navigation = referenced_json
-            else:
-                resolved_navigation[key] = _resolve_references(value, version, language)
-
-        return resolved_navigation
-
-    else:
-        # leaf node: The type of navigation should be [string, int, float, unicode]
-        return navigation
-
-
-def get_doc_subpath(version):
-    return 'docs/%s/' % version
-
-
-def get_all_links_cache_key(version, lang):
-    return 'links.%s.%s' % (version, lang)
-
-
-def remove_all_resolved_sitemaps():
-    try:
-        if os.path.exists(settings.RESOLVED_SITEMAP_DIR):
-            shutil.rmtree(settings.RESOLVED_SITEMAP_DIR)
-    except os.error as e:
-        print 'Cannot remove resolved sitemaps: %s' % e
