+++ portal/portal/static/indexes/develop/index.js
@@ -1,1 +1,1 @@
-var index = [{"content": ["\u4e0b\u9762\u7684\u4e00\u4e2a\u5b50\u76ee\u5f55\u91cc\u3002\u5f53\u8fd9\u4e00\u7ae0\u7a33\u5b9a\u4e0b\u6765\uff0c\u6211\u4eec\u4e00\u8d77\u628a\u60a8\u7684\u76ee\u5f55\u632a\u5230\u6839\u76ee\u5f55\u3002", "\u201c", "\u6211\u4eec\u628aJupyter\u3001PaddlePaddle\u3001\u4ee5\u53ca\u5404\u79cd\u88ab\u4f9d\u8d56\u7684\u8f6f\u4ef6\u90fd\u6253\u5305\u8fdb\u4e00\u4e2aDocker", "We", "\u89c6\u9891\u8bfe\u5802", "use_gpu", "\u5e76\u53ef\u4ee5\u7528", "image\u3002\u9605\u8bfb\u548c\u5728\u7ebf\u7f16\u8f91\u672c\u4e66\u8bf7\u5728\u6d4f\u89c8\u5668\u91cc\u8bbf\u95ee", "init", "pending", "\u6765\u751f\u6210\u65b0\u7684Docker", "image\u4e86\u3002\u6240\u4ee5\u60a8\u4e0d\u9700\u8981\u81ea\u5df1\u6765\u5b89\u88c5\u5404\u79cd\u8f6f\u4ef6\uff0c\u53ea\u9700\u8981\u5b89\u88c5Docker\u5373\u53ef\u3002\u5bf9\u4e8e\u5404\u79cdLinux\u53d1\u884c\u7248\uff0c\u8bf7\u53c2\u8003", "\u4f7f\u7528GPU\u8bad\u7ec3", "\u60a8\u73b0\u5728\u5728\u770b\u7684\u8fd9\u672c\u4e66\u662f\u4e00\u672c", "endverbatim", "False", "d", "\u4e2a\u6027\u5316\u63a8\u8350", "p", "latest-gpu"], "path": "/book/en/develop/index.cn.html", "title": "\u6df1\u5ea6\u5b66\u4e60\u5165\u95e8"}, {"content": ["Web", "all", "code", "Fit", "contributed", "move", "chapters", "GPU", "follow", "CPU", "Running", "write", "We", "slow", "somewhere", "use_gpu", "Pull", "except", "to", "init"], "path": "/book/en/develop/index.html", "title": "Deep Learning with PaddlePaddle"}, {"content": ["8:2", "Hyperparameter", "AGE", "global", "\u7684\u6d4b\u8bd5\u96c6\uff0c", "layers", "feeding", "use_gpu", "TAX", "\u8bad\u7ec3\u8fc7\u7a0b", "to", "init", "\u3002\u8fd9\u91cc\u6211\u4eec\u4f7f\u7528\u4e00\u79cd\u5f88\u5e38\u89c1\u7684\u64cd\u4f5c\u65b9\u6cd5\uff1a\u51cf\u6389\u5747\u503c\uff0c\u7136\u540e\u9664\u4ee5\u539f\u53d6\u503c\u8303\u56f4\u3002", "\u3002\u56e0\u6b64\uff0c\u6211\u4eec\u7684\u6a21\u578b\u5c31\u53ef\u4ee5\u8868\u793a\u6210\uff1a", "\uff09\u6a21\u578b\u5f00\u59cb\u8fd9\u4efd\u6559\u7a0b\u3002\u5728\u8fd9\u4e00\u7ae0\u91cc\uff0c\u4f60\u5c06\u4f7f\u7528\u771f\u5b9e\u7684\u6570\u636e\u96c6\u5efa\u7acb\u8d77\u4e00\u4e2a\u623f\u4ef7\u9884\u6d4b\u6a21\u578b\uff0c\u5e76\u4e14\u4e86\u89e3\u5230\u673a\u5668\u5b66\u4e60\u4e2d\u7684\u82e5\u5e72\u91cd\u8981\u6982\u5ff5\u3002", "activation", "print", "every", "fit_a_line/trainer.py", "endverbatim"], "path": "/book/en/develop/01.fit_a_line/index.cn.html", "title": "\u7ebf\u6027\u56de\u5f52"}, {"content": ["represent", "all", "code", "consider", "8:2", "AGE", "global", "results", "highways", "concept", "comments", "computation", "go", "relationships", "feeding", "use_gpu", "1940", "technique", "TAX", "0"], "path": "/book/en/develop/01.fit_a_line/index.html", "title": "Linear Regression"}, {"content": ["AISTATS", "global", "results", "1+e", "0.361920", "computation", "\u4ee3\u8868\u56fe\u7247\u5206\u7c7b\u4e3a\u7b2c", "'pixel", "Convolutional", "layer.data", "\u662f\u8f93\u51fa\uff1a\u5206\u7c7b\u5668\u7684\u8f93\u51fa\u662f10\u7c7b\u6570\u5b57\uff080-9\uff09\uff0c\u5373", "\u7ecf\u8fc7\u7b2c\u4e8c\u4e2a\u9690\u85cf\u5c42\uff0c\u53ef\u4ee5\u5f97\u5230", "use_gpu", "Dan", "\u5206\u522b\u505a\u8bad\u7ec3\u548c\u6d4b\u8bd5\u6570\u636e\u96c6\u3002\u8fd9\u4e24\u4e2a\u51fd\u6570\u5404\u81ea\u8fd4\u56de\u4e00\u4e2areader\u2014\u2014PaddlePaddle\u4e2d\u7684reader\u662f\u4e00\u4e2aPython\u51fd\u6570\uff0c\u6bcf\u6b21\u8c03\u7528\u7684\u65f6\u5019\u8fd4\u56de\u4e00\u4e2aPython", "Platt", "0", "\u591a\u5c42\u611f\u77e5\u5668\u7f51\u7edc\u7ed3\u6784\u56fe", "W_2H_1", "\u5377\u79ef\u5c42\u56fe\u7247"], "path": "/book/en/develop/02.recognize_digits/index.cn.html", "title": "\u8bc6\u522b\u6570\u5b57"}, {"content": ["limited", "all", "code", "consider", "illustrates", "During", "2278-2324", "global", "dynamic", "results", "1+e", "0.361920", "Bureau", "Fig", "computation", "go", "'pixel", "rectangles", "magnifying", "two-dimensional"], "path": "/book/en/develop/02.recognize_digits/index.html", "title": "Recognize Digits"}, {"content": ["\u76f4\u8fde", "nStages", "Van", "Olshausen", "Jia", "\u3002\u964d\u91c7\u6837\u4e5f\u662f\u56fe\u50cf\u5904\u7406\u4e2d\u5e38\u89c1\u7684\u4e00\u79cd\u64cd\u4f5c\uff0c\u53ef\u4ee5\u8fc7\u6ee4\u6389\u4e00\u4e9b\u4e0d\u91cd\u8981\u7684\u9ad8\u9891\u4fe1\u606f\u3002", "\u901a\u8fc7", "results", "R.R", "ICLR", "Krizhevsky", "\u603b\u7ed3", "feeding", "\u5206\u7c7b\u5668\u8bbe\u8ba1", "use_gpu", "Object", "Residual", "GoogleNet", "0", "init"], "path": "/book/en/develop/03.image_classification/index.cn.html", "title": "\u56fe\u50cf\u5206\u7c7b"}, {"content": ["consists", "Van", "feeding", "patches", "Residual", "fine-grained", "swap", "under", "preprocess", "activation", "connects", "updates", "0.6875", "Classification", "vector", "enhance", "direct", "second", "max-pooling", "contributed"], "path": "/book/en/develop/03.image_classification/index.html", "title": "Image Classification"}, {"content": ["dict_size", "similarity", "0.0678709", "1,1", "\u4e3a\u6240\u6709\u8bcd\u8bed\u7684\u8bcd\u5411\u91cf\u8fde\u63a5\u6210\u7684\u5927\u5411\u91cf\uff0c\u8868\u793a\u6587\u672c\u5386\u53f2\u7279\u5f81\uff1b", "\u7136\u540e\uff0c\u6307\u5b9a\u8bad\u7ec3\u76f8\u5173\u7684\u53c2\u6570\uff1a", "Testing", "\uff1a", "\u7531\u4e8e\u5f88\u591a\u8bcd\u6ca1\u6709\u51fa\u73b0\uff0c\u5bfc\u81f4\u77e9\u9635\u6781\u5176\u7a00\u758f\uff0c\u56e0\u6b64\u9700\u8981\u5bf9\u8bcd\u9891\u505a\u989d\u5916\u5904\u7406\u6765\u8fbe\u5230\u597d\u7684\u77e9\u9635\u5206\u89e3\u6548\u679c\uff1b", "\u603b\u7ed3", "0.999591588973999", "\uff0c\u7531\u4e8e\u5176\u5411\u91cf\u6b63\u4ea4\uff0c\u90fd\u8ba4\u4e3a\u8fd9\u4e24\u4e2a\u8bcd\u6beb\u65e0\u76f8\u5173\u6027\u3002", "imikolov", "use_gpu", "\u4e2a\u8bcd\u7684\u6982\u7387\u3002", "0", "init", "\u4e2d\u4ecb\u7ecd\u5982\u4f55\u5b66\u4e60\u4e00\u4e2a\u795e\u7ecf\u5143\u7f51\u7edc\u8868\u793a\u7684\u8bcd\u5411\u91cf\u6a21\u578b\u3002\u6587\u4e2d\u7684\u795e\u7ecf\u6982\u7387\u8bed\u8a00\u6a21\u578b\uff08Neural", "paddle.batch", "0.2069038"], "path": "/book/en/develop/04.word2vec/index.cn.html", "title": "\u8bcd\u5411\u91cf"}, {"content": ["limited", "all", "code", "semantic", "two-dimensional", "iterator", "similarity", "0.0678709", "caused", "results", "1,1", "illustrated", "concept", "Take", "computation", "initial_std", "go", "query", "saved", "dict_size"], "path": "/book/en/develop/04.word2vec/index.html", "title": "Word2Vec"}, {"content": ["copy", "global", "\u81f3\u6b64\uff0c\u6211\u4eec\u7684\u4f18\u5316\u76ee\u6807\u5c31\u662f\u8fd9\u4e2a\u7f51\u7edc\u914d\u7f6e\u4e2d\u7684", "'gender_id", "\u3001\u57fa\u4e8e\u7269\u54c1\uff08Item-Based\uff09\u7684\u63a8\u8350", "\u603b\u7ed3", "\u57fa\u4e8e\u5185\u5bb9\u8fc7\u6ee4\u63a8\u8350", "use_gpu", "Collaborative", "0", "init", "\u6700\u8fd1\u7684", "\u6559\u7a0b\uff09\u5904\u7406\u3002\u6700\u540e\uff0c\u8bad\u7ec3\u65f6\u5c06MLP\u7684\u8f93\u51fa\u7ed9softmax\u505a\u5206\u7c7b\uff0c\u9884\u6d4b\u65f6\u8ba1\u7b97\u7528\u6237\u7684\u7efc\u5408\u7279\u5f81\uff08MLP\u7684\u8f93\u51fa\uff09\u4e0e\u6240\u6709\u89c6\u9891\u7684\u76f8\u4f3c\u5ea6\uff0c\u53d6\u5f97\u5206\u6700\u9ad8\u7684", "Feed", "\u5176\u6b21\uff0c\u8fdb\u884c\u5377\u79ef\u64cd\u4f5c\uff1a\u628a\u5377\u79ef\u6838", "movie_id", "activation", "u'___embedding_layer_1__.w0", "P", "every"], "path": "/book/en/develop/05.recommender_system/index.cn.html", "title": "\u4e2a\u6027\u5316\u63a8\u8350"}, {"content": ["represent", "all", "code", "semantic", "two-dimensional", "Web", "shuffle", "results", "four", "'gender_id", "during", "go", "Personalized", "consider", "Convolutional", "dimensionality", "calculate", "use_gpu", "fixed-length", "to"], "path": "/book/en/develop/05.recommender_system/index.html", "title": "Personalized Recommendation"}, {"content": ["\u3002\u81ea\u7136\u8bed\u8a00\u662f\u4e00\u79cd\u5178\u578b\u7684\u5e8f\u5217\u6570\u636e\uff08\u8bcd\u5e8f\u5217\uff09\uff0c\u8fd1\u5e74\u6765\uff0c\u5faa\u73af\u795e\u7ecf\u7f51\u7edc\u53ca\u5176\u53d8\u4f53\uff08\u5982long", "Van", "forget", "0.294321", "global", "\u63a8\u8350\u7cfb\u7edf", "computation", "\u7f51\u7edc\u53c2\u6570", "\u5f88\u4e0d\u597d\u770b\uff0c\u597d\u50cf\u4e00\u4e2a\u5730\u65b9\u53f0\u7684\u7535\u89c6\u5267", "semantic", "LSTM", "\uff1a", "layers", "\u4e00\u8282\u4ecb\u7ecd\u8fc7\u7684\u6587\u672c\u5377\u79ef\u795e\u7ecf\u7f51\u7edc\uff0c\u4ee5\u53ca", "\u53ca\u5176\u6269\u5c55\u3002\u4e0b\u9762\u4f9d\u6b21\u4ecb\u7ecd\u8fd9\u51e0\u4e2a\u6a21\u578b\u3002", "use_gpu", "\uff08\u5411\u91cf\u8868\u793a\uff09\u53ca\u524d\u4e00\u65f6\u523b\u9690\u5c42\u7684\u72b6\u6001\u503c", "to", "init", "train_reader"], "path": "/book/en/develop/06.understand_sentiment/index.cn.html", "title": "\u60c5\u611f\u5206\u6790"}, {"content": ["represent", "all", "code", "semantic", "forget", "illustrates", "leads", "computation", "consists", "Van", "LSTM", "\uff1a", "issues", "Convolutional", "Turing-complete", "feeding", "use_gpu", "similarity", "to", "init"], "path": "/book/en/develop/06.understand_sentiment/index.html", "title": "Sentiment Analysis"}, {"content": ["semantic", "\u7136\u800c\uff0c\u5b8c\u5168\u53e5\u6cd5\u5206\u6790\u9700\u8981\u786e\u5b9a\u53e5\u5b50\u6240\u5305\u542b\u7684\u5168\u90e8\u53e5\u6cd5\u4fe1\u606f\uff0c\u5e76\u786e\u5b9a\u53e5\u5b50\u5404\u6210\u5206\u4e4b\u95f4\u7684\u5173\u7cfb\uff0c\u662f\u4e00\u4e2a\u975e\u5e38\u56f0\u96be\u7684\u4efb\u52a1\uff0c\u76ee\u524d\u6280\u672f\u4e0b\u7684\u53e5\u6cd5\u5206\u6790\u51c6\u786e\u7387\u5e76\u4e0d\u9ad8\uff0c\u53e5\u6cd5\u5206\u6790\u7684\u7ec6\u5fae\u9519\u8bef\u90fd\u4f1a\u5bfc\u81f4SRL\u7684\u9519\u8bef\u3002\u4e3a\u4e86\u964d\u4f4e\u95ee\u9898\u7684\u590d\u6742\u5ea6\uff0c\u540c\u65f6\u83b7\u5f97\u4e00\u5b9a\u7684\u53e5\u6cd5\u7ed3\u6784\u4fe1\u606f\uff0c", "\u901a\u8fc7", "feature_out", "proposition", "Van", "\uff1a", "\u603b\u7ed3", "feeding", "use_gpu", "test.wsj", "book/label_semantic_roles", "0", "init", "\u56fe3\u662f\u6700\u7ec8\u5f97\u5230\u7684\u6808\u5f0f\u5faa\u73af\u795e\u7ecf\u7f51\u7edc\u7ed3\u6784\u793a\u610f\u56fe\u3002", "CRF\u4ee5\u7b2c3\u6b65\u4e2dLSTM\u5b66\u4e60\u5230\u7684\u7279\u5f81\u4e3a\u8f93\u5165\uff0c\u4ee5\u6807\u8bb0\u5e8f\u5217\u4e3a\u76d1\u7763\u4fe1\u53f7\uff0c\u5b8c\u6210\u5e8f\u5217\u6807\u6ce8\uff1b", "\u8f93\u5165\u6587\u672c\u5e8f\u5217", "predicate", "\u5927\u5bb6\u53ef\u4ee5\u5c1d\u8bd5\u4e0a\u9762\u8fd9\u79cd\u65b9\u6cd5\u3002\u8fd9\u91cc\uff0c\u6211\u4eec\u63d0\u51fa\u4e00\u4e9b\u6539\u8fdb\uff0c\u5f15\u5165\u4e24\u4e2a\u7b80\u5355\u4f46\u5bf9\u63d0\u9ad8\u7cfb\u7edf\u6027\u80fd\u975e\u5e38\u6709\u6548\u7684\u7279\u5f81\uff1a", "role"], "path": "/book/en/develop/07.label_semantic_roles/index.cn.html", "title": "\u8bed\u4e49\u89d2\u8272\u6807\u6ce8"}, {"content": ["proposition", "all", "code", "semantic", "consider", "chain", "illustrates", "skip", "2012", "hidden-to-hidden", "concern", "results", "test_creator", "illustrated", "paddle.layer.fc", "\u4e86", "Fig", "computation", "initial_std", "consists"], "path": "/book/en/develop/07.label_semantic_roles/index.html", "title": "Semantic Role Labeling"}, {"content": ["Van", "dict_size", "\u901a\u8fc7", "name=decoder_group_name", "\u76ee\u6807\u8bed\u8a00\u5b57\u5178\u7ef4\u5ea6", "\u7684\u8ba1\u7b97\u516c\u5f0f\u5982\u4e0b\uff1a", "ICLR", "\uff0c\u5373\u9488\u5bf9\u6bcf\u4e00\u4e2a\u76ee\u6807\u8bed\u8a00\u4e2d\u7684\u8bcd", "\u603b\u7ed3", "\u673a\u5668\u7ffb\u8bd1\u4efb\u52a1\u7684\u751f\u6210\u8fc7\u7a0b\uff0c\u901a\u4fd7\u6765\u8bb2\u5c31\u662f\u6839\u636e\u9884\u5148\u8bad\u7ec3\u7684\u6a21\u578b\u6765\u7ffb\u8bd1\u6e90\u8bed\u8a00\u53e5\u5b50\u3002\u751f\u6210\u8fc7\u7a0b\u4e2d\u7684\u89e3\u7801\u9636\u6bb5\u548c\u4e0a\u8ff0\u8bad\u7ec3\u8fc7\u7a0b\u7684\u6709\u6240\u5dee\u5f02\uff0c\u5177\u4f53\u4ecb\u7ecd\u8bf7\u89c1", "backward_first", "use_gpu", "\u76ee\u6807\u5b57\u5178", "148.444983", "\u884c\u8fde\u63a5\uff0c\u7528't'\u5206\u9694\u3002", "0", "init", "vector\u8868\u793a\u5b58\u5728\u4e24\u4e2a\u95ee\u9898\uff0c1\uff09\u751f\u6210\u7684\u5411\u91cf\u7ef4\u5ea6\u5f80\u5f80\u5f88\u5927\uff0c\u5bb9\u6613\u9020\u6210\u7ef4\u6570\u707e\u96be\uff1b2\uff09\u96be\u4ee5\u523b\u753b\u8bcd\u4e0e\u8bcd\u4e4b\u95f4\u7684\u5173\u7cfb\uff08\u5982\u8bed\u4e49\u76f8\u4f3c\u6027\uff0c\u4e5f\u5c31\u662f\u65e0\u6cd5\u5f88\u597d\u5730\u8868\u8fbe\u8bed\u4e49\uff09\u3002\u56e0\u6b64\uff0c\u9700\u518done-hot", "z_i", "x=\\left"], "path": "/book/en/develop/08.machine_translation/index.cn.html", "title": "\u673a\u5668\u7ffb\u8bd1"}, {"content": ["limited", "all", "code", "semantic", "consider", "breadth", "global", "focus", "computation", "embeding", "query", "LSTM", "research", "CPU", "forget", "layers", "dimensionality", "calculate", "backward_first", "Attention-based"], "path": "/book/en/develop/08.machine_translation/index.html", "title": "Machine Translation"}, {"content": ["code", "just", "github.com/sugyan/tensorflow-mnist", "front", "Usage", "its", "We", "parameters", "Javascript", "0", "Front", "build", "them", "localhost:5000", "familiar", "stop", "topylogy", "s3.us-east-2.amazonaws.com/models.paddlepaddle/end-to-end-mnist/param.tar", "not", "name"], "path": "/book/en/develop/mnist-client/index.html", "title": "MNIST classification by PaddlePaddle"}, {"content": ["load", "all", "code", "stores", "0", "just", "show", "numbers", "text", "supported", "obtain", "save_parameter_to_tar", "already", "0.10060056298971176", "integer", "mentioned", "139", "cases", "layers", "content"], "path": "/book/en/develop/serve/index.html", "title": "Inference Server Example"}, {"content": ["and", "DataSets", "Image", "Dataset", "\u00c2\u00b6", "Reader", "Interface", "Data"], "path": "/documentation/en/develop/data.html", "title": "Data Reader Interface and DataSets"}, {"content": ["program=None", "data_feeder", "DataFeeder", "feed_list", "place", "\u00b6", "class", "paddle.fluid.data_feeder"], "path": "/documentation/en/develop/data_feeder.html", "title": "data_feeder"}, {"content": ["Index"], "path": "/documentation/en/develop/genindex.html", "title": "Index"}, {"content": ["gradient_clip=None", "set", "dim=None", "weight", "extra", "param_attr", "trainable=True", "an", "do_model_average=None", "in", "indicate", "regularizer=None", "normalize", "ParamAttr", "for", "also", "except", "field", "Besides", "kwargs"], "path": "/documentation/en/develop/param_attr.html", "title": "param_attr"}, {"content": ["results", "not", "sorted", "\u2018", "\u201c", "previous", "Key-Value", "output_file", "state", "should", "0", "add", "program", "Although", "has", "By", "string", "format", "records", "telling"], "path": "/documentation/en/develop/profiler.html", "title": "profiler"}, {"content": ["Index", "Python", "paddle", "p", "Module", "paddle.v2.image"], "path": "/documentation/en/develop/py-modindex.html", "title": "Python Module Index"}, {"content": ["all", "protocol", "square_error_cost", "avg_cost", "shape", "Lookup", "computation", "Evaluates", "Scope", "CPU", "layers", "Fluid", "paddle/fluid/framework/tensor.h", "Base", "Initialize", "input", "has", "variables", "handle", "fluid"], "path": "/documentation/en/develop/read_source.html", "title": "PaddlePaddle Fluid Source Code Overview"}, {"content": ["and", "A", "Regularization", "set", "Returns", "weight", "regularizer", "Unknown", "Create", "global", "It", "regularized", "as", "type", "in", "need", "gradients", "append_regularization_ops", "BlockDesc", "L2DecayRegularizer"], "path": "/documentation/en/develop/regularizer.html", "title": "regularizer"}, {"content": ["all", "code", "consider", "BLAS", "saves", "Does", "go", "follow", "disk", "saved", "CPU", "build/python/dist/*.whl", "environment", "to", "Step", "under", "Not", "Frequently", "include", "paddle/paddle/scripts/docker/build.sh"], "path": "/documentation/en/develop/build_and_install/build_from_source_en.html", "title": "Build from Sources"}, {"content": ["all", "code", "proc/cpuinfo", "DockerHub", "edit", "BLAS", "move", "python", "libs", "serving", "work/train.py", "using", "go", "CPU", "OpenBLAS", "before", "one", "We", "ll", "train.py"], "path": "/documentation/en/develop/build_and_install/docker_install_en.html", "title": "Run in Docker Containers"}, {"content": ["all", "code", "Bottom", "Does", "during", "still", "We", "\u521b\u5efaissue", "environment", "to", "choose", "has", "python", "not", "using", "instructions", "approach", "Warning", "Learning", "require"], "path": "/documentation/en/develop/build_and_install/index_en.html", "title": "install and Compile"}, {"content": ["MacOS", "code", "just", "supported", "Access", "using", "still", "\u201c", "cuda8.0_cudnn7_avx_mkl", "suffix", "libraries", "We", "trying", "ll", "2.7.x", "seems", "libc.so", "environment", "to", "only"], "path": "/documentation/en/develop/build_and_install/pip_install_en.html", "title": "Install using pip"}, {"content": ["code", "chain", "nested", "whose", "compose", "iterator", "tp", "created", "to", "program", "batched", "InputType", "supported", "string", "every", "entries", "dim", "False", "get_line", "list"], "path": "/documentation/en/develop/data/data_reader.html", "title": "Data Reader Interface"}, {"content": ["semantic", "dict_size", "consists", "categoriges", "imikolov", "to", "preprocess", "indicating", "Training", "string", "verb", "categories", "entire", "False", "gender", "large", "Movielens", "Sima", "vector", "small"], "path": "/documentation/en/develop/data/dataset.html", "title": "Dataset"}, {"content": ["operations", "\u2018", "filename", "1", "load_image_bytes", "layout", "horizontal", "param", "0", "usage", "input", "include", "image", "them", "return", "string", "format", "num_per_batch", "flipping", "2,0,1"], "path": "/documentation/en/develop/data/image.html", "title": "Image Interface"}, {"content": ["and", "Management", "Network", "Gradient", "Multi-Device", "Distributed", "Complex", "Memory", "Type", "Core", "Multi-Language", "Update", "Concurrent", "Important", "Goals", "Interface", "Execution", "Data", "Support", "Structure"], "path": "/documentation/en/develop/design/index_en.html", "title": "Design"}, {"content": ["Algorithm", "Gradient", "Averaging", "PaddlePaddle", "Update", "\u00c2\u00b6", "in", "Parameter"], "path": "/documentation/en/develop/design/algorithm/index_en.html", "title": "Gradient Update Algorithm"}, {"content": ["operations", "represent", "all", "code", "Based", "proposed", "Juditsky", "over", "desirable", "obtain", "re-writing", "through", "computation", "using", "mentioned", "speed", "its", "before", "layers", "1"], "path": "/documentation/en/develop/design/algorithm/parameter_average.html", "title": "Averaging Parameter in PaddlePaddle"}, {"content": ["default", "all", "global", "unique", "scalar", "attrs", "computation", "Var", "if-else", "layers", "const", "graph", "description", "languages", "0", "init", "program", "ifelse", "string", "void"], "path": "/documentation/en/develop/design/concepts/block.html", "title": "Design Doc: Block and Scope"}, {"content": ["all", "Feeding", "EOF", "MultiPassReader", "leads", "prefix", "code", "feeding", "const", "ReadNext", "graph", "Fluid", "environment", "to", "program", "friendly", "And", "very", "void", "HasNextOp"], "path": "/documentation/en/develop/design/concepts/cpp_data_feeding.html", "title": "C++ Data Feeding"}, {"content": ["operations", "code", "Also", "layer.fc", "using", "Var", "fine", "O2", "O1", "layers", "We", "compose", "implementation", "with", "explain", "should", "to", "add", "going", "illustrative"], "path": "/documentation/en/develop/design/concepts/functions_operators_layers.html", "title": "Design Doc: Functions, Operators, and Layers"}, {"content": ["and", "Layers", "Functions", "Feeding", "Tensor", "Executor", "PaddlePaddle", "Paddle", "ParallelExecutor", "in", "Scope", "Type", "Function", "Core", "for", "Operators", "Level-of-Detail", "Variable", "Data", "TensorArray"], "path": "/documentation/en/develop/design/concepts/index_en.html", "title": "Core Concepts"}, {"content": ["represent", "What", "just", "show", "numbers", "indeed", "Images", "all", "ending", "underlying", "through", "consists", "revisit", "image", "adopt", "convolution", "Level-of-Detail", "We", "2-level", "640x480"], "path": "/documentation/en/develop/design/concepts/lod_tensor.html", "title": "Design Doc: LoD (Level-of-Detail) Tensor"}, {"content": ["represent", "all", "concept", "results", "computation", "CPU", "duplicated", "graph", "Fluid", "to", "program", "helps", "nodes", "Parallel", "string", "synchronized", "very", "MultiGPUs", "multi-GPUs", "endverbatim"], "path": "/documentation/en/develop/design/concepts/parallel_executor.html", "title": "ParallelExecutor"}, {"content": ["concept", "vars", "global", "true_block", "IfElseOp", "shape", "stored", "attrs", "consists", "20", "parameters", "with", "should", "0", "program", "has", "local", "into", "int32", "return"], "path": "/documentation/en/develop/design/concepts/program.html", "title": "Design Doc: PaddlePaddle Programs"}, {"content": ["all", "code", "global", "cudnn", "go", "lgtest", "depend", "graph", "cc_test", "to", "under", "nodes", "Now", "very", "ops.h", "choice", "cc_binary", "GO", "difference", "nv_library"], "path": "/documentation/en/develop/design/concepts/README.html", "title": "Required CMake Function"}, {"content": ["default", "all", "concept", "just", "but", "global", "Also", "entity", "stored", "relation", "same", "computation", "Var", "Scope", "pointer", "its", "some", "unique", "const", "vars"], "path": "/documentation/en/develop/design/concepts/scope.html", "title": "Design of Scope in Paddle"}, {"content": ["represent", "all", "code", "consider", "iterator", "caused", "global", "dynamic", "four", "manages", "consists", "CPU", "row", "const", "pointing", "to", "tail", "include", "friendly", "resources"], "path": "/documentation/en/develop/design/concepts/tensor.html", "title": "Tensor: An Unified Data Type in PaddlePaddle"}, {"content": ["all", "concept", "dynamic", "splitting", "existing", "scalar", "code", "Var", "Supports", "layers", "whose", "batches", "const", "segments", "presents", "to", "init", "th", "preprocess", "ta"], "path": "/documentation/en/develop/design/concepts/tensor_array.html", "title": "Design for TensorArray"}, {"content": ["operations", "represent", "executed", "from", "TensorDesc", "supported", "concept", "looks", "computation", "consists", "saved", "workers", "distributed", "should", "to", "only", "program", "4", "5", "interprets"], "path": "/documentation/en/develop/design/concepts/var_desc.html", "title": "Design Doc: Var_desc"}, {"content": ["code", "Requirements", "Similarly", "able", "saves", "anyway", "syntax", "held", "pointer", "its", "utilizes", "Lazy", "implementation", "graph", "technique", "should", "Allocation", "to", "only", "blob"], "path": "/documentation/en/develop/design/concepts/variable.html", "title": "Design Doc: Variable"}, {"content": ["limited", "they", "Closes", "being", "supported", "synchronous", "shape", "ch", "through", "to/from", "sent/received", "still", "before", "note", "We", "on", "Channels", "send", "should", "0"], "path": "/documentation/en/develop/design/concurrent/channel.html", "title": "Channel Design"}, {"content": ["represent", "all", "executed", "managed", "0", "just", "parallel_for", "process/node", "saves", "control-flow", "splitting", "syntax", "listen_and_do", "attrs", "questions", "integer", "its", "Executor.Run", "Tensor", "layers"], "path": "/documentation/en/develop/design/concurrent/concurrent_programming.html", "title": "Design Doc: Concurrent Programming with Fluid"}, {"content": ["all", "concept", "illustrates", "control-flow", "focus", "four", "code", "computation", "go", "if-else", "layers", "kqueue", "compose", "unblocked", "graph", "Fluid", "send", "to", "program", "Now"], "path": "/documentation/en/develop/design/concurrent/csp.html", "title": "Design Doc: CSP in PaddlePaddle Fluid"}, {"content": ["all", "executed", "creation", "vars", "less", "being", "captures", "global", "executing", "shape", "Take", "X", "go", "only", "before", "whose", "We", "Backward", "3", "should"], "path": "/documentation/en/develop/design/concurrent/go_op.html", "title": "go_op Design"}, {"content": ["Parallel_Do", "Doc", "Programming", "Fluid", "Concurrent", "\u00c2\u00b6", "Design", "PaddlePaddle", "in", "with"], "path": "/documentation/en/develop/design/concurrent/index_en.html", "title": "Concurrent Programming"}, {"content": ["represent", "all", "executed", "vars", "Imporvement", "over", "AddInput", "kPlaces", "looks", "CPU", "duplicated", "layers", "Inputs", "We", "copy", "20", "decay", "implementation", "use_gpu", "param"], "path": "/documentation/en/develop/design/concurrent/parallel_do.html", "title": "Design Doc: Parallel_Do in PaddlePaddle"}, {"content": ["operations", "fluid.Select.case", "all", "executed", "set", "true", "random", "while_cond", "elementwise_add", "shape", "simplier", "executes", "attrs", "X", "Scope", "channel_102", "its", "channel_101", "fill_constant_109.tmp_0", "whose"], "path": "/documentation/en/develop/design/concurrent/select_op.html", "title": "select_op Design"}, {"content": ["all", "code", "global", "However", "executing", "computation", "consists", "CPU", "layers", "whose", "feeding", "const", "Fluid", "send", "intensive", "to", "finally", "program", "norm", "numpy.uint16"], "path": "/documentation/en/develop/design/data_type/float16.html", "title": "Design Doc: float16"}, {"content": ["float16", "Data", "\u00c2\u00b6", "Design", "Doc", "Type"], "path": "/documentation/en/develop/design/data_type/index_en.html", "title": "Data Type"}, {"content": ["all", "global", "synchronous", "competitive", "Threadpool", "immediately", "find", "before", "whose", "We", "Optimize", "parameters", "Schedule", "distributed", "Update", "send", "than", "better", "to", "finally"], "path": "/documentation/en/develop/design/dist_train/async_update.html", "title": "Design Doc: Asynchronous Update With Distributed Training"}, {"content": ["all", "code", "Take", "computation", "'pixel", "if-else", "layers", "mem_per_pserver", "Kubernetes", "Executors", "graph", "Fluid", "send", "environment", "to", "fluid.layer.fc", "program", "train_reader", "nodes", "Executor.run"], "path": "/documentation/en/develop/design/dist_train/distributed_architecture.html", "title": "Design Doc: Fluid Distributed Training Architecture"}, {"content": ["Service", "illustrates", "when", "both", "symbols", "performance", "Does", "Lookup", "solutions", "Forward", "its", "non-zero", "We", "one", "distributed", "except", "former", "better", "to", "Distributed"], "path": "/documentation/en/develop/design/dist_train/distributed_lookup_table_design.html", "title": "Design Doc: Distributed Lookup Table Operator"}, {"content": ["all", "Norouzi", "results", "computation", "LSTM", "issues", "Schuster", "layers", "depend", "technique", "to", "stable", "worth", "resources", "case", "Training", "worse", "advantage", "very", "N+M"], "path": "/documentation/en/develop/design/dist_train/distributed_traing_review.html", "title": "Parallelism, Asynchronous,  Synchronous, Codistillation"}, {"content": ["Training", "Doc", "Distributed", "Fluid", "Server", "\u00c2\u00b6", "Design", "Architecture", "Table", "Operator", "Lookup", "Parameter"], "path": "/documentation/en/develop/design/dist_train/index_en.html", "title": "Distributed Training"}, {"content": ["operations", "all", "code", "distribute", "find", "mpirun", "bring", "identify", "using", "compiling.Add", "CPU", "low-level", "before", "one", "write", "We", "copy", "group", "system", "integer"], "path": "/documentation/en/develop/design/dist_train/mpi_enabled_design.html", "title": "MPI-enabled PaddlePaddle Design doc"}, {"content": ["all", "executed", "thread_pool", "atomic", "CPU", "parameters", "graph", "0", "Initialize", "input", "Wait", "then", "Execute", "propose", "endverbatim", "TODO", "Op", "execute", "user-defined", "Multi-CPU"], "path": "/documentation/en/develop/design/dist_train/multi_cpu.html", "title": "Design Doc: Execute the Program with Multi CPU"}, {"content": ["represent", "now", "code", "scheduler", "both", "Heterogeneous", "bandwidth", "Systems", "computation", "Async", "mentioned", "converting", "its", "previous", "layers", "heuristic", "We", "slow", "graph", "workers"], "path": "/documentation/en/develop/design/dist_train/parameter_server.html", "title": "Design Doc: Parameter Server"}, {"content": ["BATCH_BARRIER_MESSAGE", "all", "consider", "just", "synchronous", "go", "speed", "its", "parameters", "distributed", "Fluid", "send", "should", "0", "Using", "Now", "local", "then", "them", "handle"], "path": "/documentation/en/develop/design/dist_train/README.html", "title": "Distributed training overview doc"}, {"content": ["RNNOp", "RNN", "\u53d8\u957f\u8f93\u5165\u8bbe\u8ba1", "Dynamic", "design", "\u00b6"], "path": "/documentation/en/develop/design/dynamic_rnn/index_en.html", "title": "Dynamic RNN"}, {"content": ["default", "all", "step-scopes", "illustrates", "text", "128-dim", "dynamic", "128", "shape", "including", "using", "mentioned", "its", "previous", "whose", "We", "compose", "20", "parameters", "also"], "path": "/documentation/en/develop/design/dynamic_rnn/rnn.html", "title": "RNNOp design"}, {"content": ["\uff0c\u5e76\u4e14\u8c03\u7528", "\u6765\u4fdd\u8bc1\u4f20\u9012\u53ea\u8fdb\u884c\u4e00\u6b21", "producer", "\u5373\u5c06\u4e00\u4e2amini-batch\u5185\u4e0d\u540c\u957f\u5ea6\u7684\u5e8f\u5217\u88650\u5230\u56fa\u5b9a\u957f\u5ea6\u53c2\u4e0e\u8ba1\u7b97\u3002", "Bucketing", "not", "\u800c\u6240\u6709\u4f9d\u8d56\u5176", "\u7684\u4f20\u9012", "ShareConstLODFrom", "\u7684\u5b9e\u73b0\u4e2d\uff0c\u64cd\u4f5c\u66f4\u65b0\u81ea\u5df1\u7684", "size_t", "\u9700\u8981\u4fee\u6539", "sould", "boot_memories", "\u5fc5\u987b\u7528zero-padding\u7684\u65b9\u5f0f\u5c06\u53d8\u957f\u5e8f\u5217\u8865\u5168\u4e3a\u56fa\u5b9ashape\u7684tensor\u3002", "Tensor", "\u7b49\u5747\u4f7f\u7528\u4e86padding\u7684\u65b9\u5f0f\uff0c", "member", "typedef", "\u7684\u4fe1\u606f\u7684\u4f20\u9012\u5bf9\u975eOLD\u7684Op\u7684\u5b9e\u73b0\u662f\u5b8c\u5168\u900f\u660e\u7684\u3002"], "path": "/documentation/en/develop/design/dynamic_rnn/rnn_design.html", "title": "RNN \u53d8\u957f\u8f93\u5165\u8bbe\u8ba1"}, {"content": ["all", "chain", "existing", "sould", "const", "to", "Argument.subSequenceStartPositions", "mainstream", "void", "replication", "endverbatim", "level", "vector", "prevent", "consistent", "PaddlePaddle", "cost", "design", "pass", "implemented"], "path": "/documentation/en/develop/design/dynamic_rnn/rnn_design_en.html", "title": "Varient Length supported RNN Design"}, {"content": ["true_block", "shape", "identify", "differs", "20", "true", "either", "to", "add", "program", "usage", "bool", "mini-batch", "15", "Example", "condition", "endverbatim", "d", "follows", "scalar"], "path": "/documentation/en/develop/design/execution/if_else_op.html", "title": "The"}, {"content": ["Process", "Doc", "Switch", "\u00c2\u00b6", "Design", "Operator", "IfElse", "The", "Execution"], "path": "/documentation/en/develop/design/execution/index_en.html", "title": "Execution Process"}, {"content": ["control-flow", "scalar", "differs", "if-else", "We", "implementation", "Fluid", "should", "languages", "0", "add", "program", "usage", "matched", "easier", "Semantics", "which", "matches", "Many", "fluid"], "path": "/documentation/en/develop/design/execution/switch.html", "title": "Design Doc: Switch"}, {"content": ["Multi-Language", "Interface", "\u00c2\u00b6", "TBD"], "path": "/documentation/en/develop/design/interface/index_en.html", "title": "Multi-Language Interface"}, {"content": ["\u00c2\u00b6", "Management", "Optimization", "Memory"], "path": "/documentation/en/develop/design/memory/index_en.html", "title": "Memory Management"}, {"content": ["limited", "all", "go", "follow", "expressions", "graph", "technique", "to", "init", "program", "Not", "nodes", "very", "activation", "every", "formula", "endverbatim", "Reference", "large", "solved"], "path": "/documentation/en/develop/design/memory/memory_optimization.html", "title": "Memory Optimization"}, {"content": ["initial_pool_size", "cout", "max_pool_size", "both", "gpu_id", "Region-based", "not", "looks", "Usage", "CPU", "to", "fallback", "hold", "buddy", "parameters", "meta-data", "Free", "0", "only", "extends"], "path": "/documentation/en/develop/design/memory/README.html", "title": "Region-based Heterogeneous Memory Management"}, {"content": ["all", "code", "chain", "scalar", "block=s_block", "nested", "whose", "Backward", "calculate", "graph", "Args", "father", "Fluid", "to", "program", "marks", "Cases", "Now", "uninitialized", "updated"], "path": "/documentation/en/develop/design/modules/backward.html", "title": "Backward Building"}, {"content": ["Kernels", "What", "inferencing", "0", "Once", "suggests", "all", "var_cache", "assigned", "cudnn", "batch_var", "through", "contains", "computation", "Attributes", "go", "estimated_mean", "its", "previous", "As"], "path": "/documentation/en/develop/design/modules/batch_norm_op.html", "title": "Batch Normalization"}, {"content": ["and", "Regularization", "Network", "Gradient", "PaddlePaddle", "in", "Optimizer", "Python", "Operators", "Important", "Structure", "Building", "Prune", "InferVarType", "Doc", "Modules", "Backward", "Code", "\u00c2\u00b6", "API"], "path": "/documentation/en/develop/design/modules/index_en.html", "title": "Code Structure and Important Modules"}, {"content": ["just", "follow", "its", "We", "const", "struct", "implementation", "ease", "should", "to", "input", "them", "return", "void", "Register", "not", "using", "some_op", "replace", "endverbatim"], "path": "/documentation/en/develop/design/modules/infer_var_type.html", "title": "Design Doc: InferVarType"}, {"content": ["default", "all", "code", "Network.Run", "just", "neural", "hierarchy", "AddOptimization", "RecurrentNet", "stored", "attrs", "OpIndex", "managed", "label", "networks", "BaseNet", "member", "typedef", "unique", "const"], "path": "/documentation/en/develop/design/modules/net_op_design.html", "title": "Network Design"}, {"content": ["code", "0", "being", "cost_np", "op_desc.is_traget", "session", "attrs", "Var", "its", "count", "We", "const", "depended", "also", "situations", "foward", "add", "4", "bool", "into"], "path": "/documentation/en/develop/design/modules/prune.html", "title": "Prune"}, {"content": ["all", "code", "uniform_random", "global", "dynamic", "attrs", "follow", "layers", "param.stop_gradient=True", "to", "init", "program", "manipulates", "string", "activation", "endverbatim", "InferShape", "Global", "list", "fewer"], "path": "/documentation/en/develop/design/modules/python_api.html", "title": "Design Doc: Python API"}, {"content": ["now", "code", "just", "existing", "stored", "virtual", "computation", "consists", "minus_grad", "Developers", "its", "whose", "We", "rapid", "const", "struct", "field", "CreateGradientOperator", "should", "to"], "path": "/documentation/en/develop/design/modules/register_grad_op.html", "title": "Design Doc: Gradient Operators Registration"}, {"content": ["operations", "represent", "code", "background", "just", "spaces", "computation", "using", "implement", "CPU", "before", "layers", "adding", "We", "based", "parameters", "also", "implementation", "graph", "should"], "path": "/documentation/en/develop/design/modules/regularization.html", "title": "Regularization in PaddlePaddle"}, {"content": ["represent", "SelectedRow", "Selected", "shape", "So", "satisfies", "saved", "its", "row", "non-zero", "3,4", "written", "should", "to", "only", "4", "5", "has", "them", "string"], "path": "/documentation/en/develop/design/modules/selected_rows.html", "title": "Design Doc: Selected Rows"}, {"content": ["all", "code", "semantic", "global", "existing", "computation", "ParameterSet", "layers", "compose", "Kubernetes", "environment", "to", "program", "Not", "include", "sent", "Training", "updater", "decide", "read_from_realistic_images"], "path": "/documentation/en/develop/design/motivation/api.html", "title": "PaddlePaddle Design Doc"}, {"content": ["Existed", "all", "concept", "Based", "four", "layers", "Turing-complete", "graph", "Fluid", "removal", "languages", "to", "program", "very", "difference", "condition", "look", "endverbatim", "False", "PaddlePaddle"], "path": "/documentation/en/develop/design/motivation/fluid.html", "title": "Design Doc: PaddlePaddle Fluid"}, {"content": ["code", "vars", "global", "rest", "looks", "consists", "interpreter", "its", "before", "We", "const", "Generator", "generator", "instance", "distributed", "Fluid", "should", "0", "program", "input"], "path": "/documentation/en/develop/design/motivation/fluid_compiler.html", "title": "PaddlePaddle Fluid: Towards a Compiled Programming Language"}, {"content": ["Motivations", "and", "Towards", "Language", "Doc", "PaddlePaddle", "Fluid", "Refactorization", "\u00c2\u00b6", "Design", "Goals", "a", "Overview", "Programming", "Compiled"], "path": "/documentation/en/develop/design/motivation/index_en.html", "title": "Design Motivations and Goals"}, {"content": ["represent", "code", "migrated", "particular", "Macros", "Take", "computation", "consists", "CPU", "layers", "whose", "OpKernel", "OpWithKernel/Kernel", "Kubernetes", "graph", "send", "languages", "to", "program", "Registry"], "path": "/documentation/en/develop/design/motivation/refactorization.html", "title": "Design Doc: Refactorization Overview"}, {"content": ["Kernel", "Selection", "Hint", "Keys", "Doc", "Multi-Device", "\u00c2\u00b6", "Design", "of", "Operator", "The", "Type", "Support"], "path": "/documentation/en/develop/design/multi_devices/index_en.html", "title": "Multi-Device Support"}, {"content": ["all", "code", "global", "kernel_hint", "same", "kForceCPU", "CPU", "layer_helper", "layout", "struct", "fit", "also", "personal", "send", "should", "to", "add", "choose", "input", "GetExpectedKernelType"], "path": "/documentation/en/develop/design/multi_devices/kernel_hint_design.html", "title": "Kernel Hint Design"}, {"content": ["global", "find", "rest", "Take", "Var", "follow", "its", "library_type", "Inputs", "We", "layout", "struct", "situations", "new_scope", "Kernel", "CPU", "Fluid", "should", "better", "to"], "path": "/documentation/en/develop/design/multi_devices/kernel_selection.html", "title": "Kernel Selection"}, {"content": ["represent", "all", "supported", "mainly", "four", "course", "shape", "identify", "computation", "ifdef", "follow", "stride", "pointer", "CPU", "based", "member", "We", "now", "layout", "struct"], "path": "/documentation/en/develop/design/multi_devices/operator_kernel_type.html", "title": "Design Doc: The Keys of Operator Kernel Type"}, {"content": ["all", "WER/CER", "Improvement", "prefix", "computation", "Department", "row", "devoted", "batches", "fixed-length", "Standard", "to", "CTC-LM", "Not", "supported", "fully-connected", "DNNs", "very", "CTC-loss", "word"], "path": "/documentation/en/develop/design/network/deep_speech_2.html", "title": "DeepSpeech2 on PaddlePaddle: Design Doc"}, {"content": ["Generating", "Network", "Sequence", "LoDTensors", "\u00c2\u00b6", "Complex", "Design", "Decoder"], "path": "/documentation/en/develop/design/network/index_en.html", "title": "Complex Network Design"}, {"content": ["limited", "all", "semantic", "dict_size", "represent", "results", "prefix", "last_seq", "lookup", "Introducing", "Encoder", "prefixes", "to", "init", "Changing", "updated", "very", "activation", "translations", "source_dict_dim"], "path": "/documentation/en/develop/design/network/sequence_decoder.html", "title": "Design: Sequence Decoder Generating LoDTensors"}, {"content": ["all", "code", "fluid_model", "global", "dynamic", "focus", "four", "whose", "inverse", "graph", "Fluid", "to", "finally", "program", "nodes", "Now", "case", "mainstream", "very", "two-way"], "path": "/documentation/en/develop/design/onnx/onnx_convertor.html", "title": "Background"}, {"content": ["restore", "absolute", "accurate", "supported", "backward_op", "stability", "shape", "computation", "GPU", "CPP", "check_grad", "suffer", "CPU", "Calculate", "whose", "Theory", "instance", "cpu", "shoule", "should"], "path": "/documentation/en/develop/design/others/auto_gradient_check.html", "title": "Auto Gradient Check Design"}, {"content": ["all", "0", "github.com/rajathkmp/DCGAN", "if-else", "Convolutional", "graph", "to", "init", "tf", "001", "decide", "summarized", "G_h1_bn", "Declare", "endverbatim", "paddle.v2", "list", "load_mnist", "small", "the"], "path": "/documentation/en/develop/design/others/gan_api.html", "title": "Design for GAN"}, {"content": ["all", "code", "chain", "results", "1", "four", "looks", "computation", "Forward", "followed", "saved", "its", "fifth", "before", "layers", "layer.data", "ConstructOptimizationGraph", "vars", "parameters", "also"], "path": "/documentation/en/develop/design/others/graph.html", "title": "Design Doc: Computations as a Graph"}, {"content": ["represent", "all", "concept", "help", "produces", "name='fc1", "cg.print_graphviz", "global", "focus", "flatten0", "produced", "128", "session", "including", "same", "Expression", "follow", "label", "pointer", "find"], "path": "/documentation/en/develop/design/others/graph_survey.html", "title": "Survey on Graph"}, {"content": ["concept", "stores", "just", "from", "ParameterUpdater", "Multi-GPU", "multi-GPU/CPU", "code", "CPP", "its", "MultiGradientMachine", "before", "We", "const", "parameters", "implementation", "should", "to", "only", "Make"], "path": "/documentation/en/develop/design/others/parameters_in_cpp.html", "title": "Design Doc: The C++ Class"}, {"content": ["What", "code", "sigmoid", "all", "existing", "opdesc", "not", "elif", "AddInput", "OpProtoMaker", "attrs", "implement", "desc", "CosineOpProto", "\u201c", "low-level", "register", "layers", "1", "We"], "path": "/documentation/en/develop/design/others/simple_op_design.html", "title": "Interaction between C++ and Python"}, {"content": ["coresponding", "N-1", "less", "layer_norm", "results", "ParamAttr|list", "shape", "num_flatten_dims=1", "dynamic_lstm", "Take", "Under", "LSTM", "act=None", "its", "Tensor", "whose", "parameters", "omit", "Args", "should"], "path": "/documentation/en/develop/dev/api_doc_std_en.html", "title": "API Doc Standard"}, {"content": ["Git", "reviewers", "Fix", "code", "follow", "contributions", "Skipped", "Workflow", "Standard", "environment", "0", "helps", "7", "Fixes", "every", "upstream", "forked", "endverbatim", "level", "button"], "path": "/documentation/en/develop/dev/contribute_to_paddle_en.html", "title": "Contribute Code"}, {"content": ["Kernels", "\u2019", "Device/Library", "Convention", "PaddlePaddle", "Paddle", "How", "Add", "in", "operator", "New", "Parameter", "Development", "Supporting", "use", "for", "Format", "Operators", "Releasing", "Standard"], "path": "/documentation/en/develop/dev/index_en.html", "title": "Development"}, {"content": ["unified", "all", "show", "Comments", "AddInput", "Attributes", "follow", "Tensor", "We", "op_checker", "explain", "OpProtoAndCheckerMaker", "Here", "should", "examples", "to", "only", "listing", "usage", "input"], "path": "/documentation/en/develop/dev/name_convention.html", "title": "Operator's Parameter Name Convention"}, {"content": ["code", "device_context", "Describes", "attrs", "computation", "consists", "CPU", "OpKernel", "const", "constructor\u00ef\u00bc\u0161", "InferShapeContext", "0", "MulOp", "include", "string", "void", "VariableNameMap", "Where", "entire", "endverbatim"], "path": "/documentation/en/develop/dev/new_op_en.html", "title": "How to write a new operator"}, {"content": ["Kernels", "code", "GemmConvKernel", "your", "virtual", "using", "CPU", "before", "library_type", "resource", "We", "kCUDNN", "const", "struct", "implementation", "Kernel", "Fluid", "should", "to", "add"], "path": "/documentation/en/develop/dev/new_op_kernel.html", "title": "Add Kernels for a New Device"}, {"content": ["all", "words", "proposed", "rest", "relevant", "including", "mentioned", "follow", "paper/article", "Inputs", "should", "to", "other", "array_to_lod_tensor", "covered", "has", "them", "above", "get", "overall"], "path": "/documentation/en/develop/dev/op_markdown_format.html", "title": "Standard Markdown Format for Operators"}, {"content": ["all", "dist", "Artifacts", "major", "chapters", "feature", "not", "manages", "including", "merge", "Test", "follow", "Advise", "each", "its", "previous", "We", "Personalized", "Linear", "V1"], "path": "/documentation/en/develop/dev/releasing_process_en.html", "title": "PaddlePaddle Releasing Process"}, {"content": ["represent", "all", "consider", "Crop", "REGISTER_OP_CUDA_KERNEL", "demand", "mainly", "obtain", "However", "performance", "shape", "CropGradKernel", "cudaStream_t", "file", "using", "Integrate", "pointer", "ptr", "low-level", "CPUDeviceContext"], "path": "/documentation/en/develop/dev/support_new_device.html", "title": "Design Doc: Supporting new Device/Library"}, {"content": ["all", "code", "Using", "show", "being", "developed", "bring", "shape", "stored", "computation", "using", "Usage", "still", "pointer", "its", "note", "OpKernel", "input0/input1/output", "const", "Computation"], "path": "/documentation/en/develop/dev/use_eigen_en.html", "title": "How to use Eigen in Paddle"}, {"content": ["WITH_DOC=ON", "documentation\uff0cPlease", "code", "set", "produces", "find", "results", "processors", "installed", "indicates", "Release", "stored", "through", "PRs", "affect", "\u2018", "facilitates", "once", "its", "developer"], "path": "/documentation/en/develop/dev/write_docs_en.html", "title": "Contribute Documentation"}, {"content": ["\u00c2\u00b6", "FAQ"], "path": "/documentation/en/develop/faq/index_en.html", "title": "FAQ"}, {"content": ["and", "is", "PaddlePaddle", "an", "Start", "including", "want", "regression", "linear", "use", "GET", "STARTED", "format", "please", "Here", "quickly", "how", "Running", "program", "build"], "path": "/documentation/en/develop/getstarted/index_en.html", "title": "GET STARTED"}, {"content": ["operations", "limited", "all", "code", "just", "increasingly", "over", "move", "syntax", "through", "research", "networks", "before", "layers", "Think", "how", "implementation", "graph", "Fluid", "should"], "path": "/documentation/en/develop/getstarted/overview_en.html", "title": "Overview"}, {"content": ["feed", "MacOS", "code", "fluid_model", "installed", "housing.py", "paddle.dataset.uci_housing", "Scope", "should", "0", "voila", "details", "format", "python", "paddle.fluid", "fluid", "new", "Quick", "GPU", "CentOS"], "path": "/documentation/en/develop/getstarted/quickstart_en.html", "title": "Quick Start"}, {"content": ["all", "executed", "subsections", "syntax", "code", "through", "same", "concurrency", "Scope", "yes", "\u201c", "feedforward", "Aspects", "interesting", "epoch", "actually", "to", "program", "under", "input"], "path": "/documentation/en/develop/getstarted/running_a_program_en.html", "title": "Running a program"}, {"content": ["\u00c2\u00b6", "TBD", "Concepts"], "path": "/documentation/en/develop/getstarted/concepts/index_en.html", "title": "Concepts"}, {"content": ["code", "skip", "existing", "compose", "Returning", "to", "paddle.batch", "supported", "true_label", "advantage", "very", "true_image", "entries", "endverbatim", "False", "list", "try", "item", "PaddlePaddle", "second"], "path": "/documentation/en/develop/getstarted/concepts/reader/README.html", "title": "Python Data Reader Design Doc"}, {"content": ["all", "code", "stores", "0", "text", "computation", "consists", "TensorDesc.dims", "disk", "saved", "its", "bunch", "We", "parameters", "to", "version", "has", "Optional", "By", "string"], "path": "/documentation/en/develop/getstarted/concepts/save_model/model_format.html", "title": "Design Doc: Model Format"}, {"content": ["Performance", "HOW", "\u00c2\u00b6", "Optimization", "TO"], "path": "/documentation/en/develop/howto/index_en.html", "title": "HOW TO"}, {"content": ["all", "code", "DWITH_SWIG_PY", "layers", "Introducing", "Fluid", "0", "program", "train_reader", "exit", "nodes", "Now", "Training", "fully-connected", "paddle.fluid", "endverbatim", "list", "try", "PaddlePaddle", "192"], "path": "/documentation/en/develop/howto/cluster/fluid_cluster_train_en.html", "title": "Fluid Distributed Training"}, {"content": ["0", "InfiniBand", "ping", "supported", "installed", "libmlx4.so", "including", "7.4", "GPU", "on", "PADDLE_WORKERS", "find", "before", "Files", "1", "We", "Multiple", "www.mellanox.com/page/products_dyn", "everything", "somewhere"], "path": "/documentation/en/develop/howto/cluster/nccl2_rdma_training.html", "title": "Distributed Training with NCCL2 and RDMA"}, {"content": ["code", "focus", "go", "follow", "CPU", "ncalls", "home/yuyang/perf_test/.env/lib/python2.7/site-packages/paddle/fluid/framework.py:219", "to", "init", "program", "include", "Identifying", "very", "results", "libgoogle-perftools-dev", "Where", "endverbatim", "Directing", "p", "Go"], "path": "/documentation/en/develop/howto/optimization/cpu_profiling_en.html", "title": "Profiling the Python Code"}, {"content": ["profile", "do", "use", "Profiling", "how", "Python", "timeline", "tool", "Benchmark", "to", "\u00c2\u00b6", "Optimization", "Performance", "the", "Code"], "path": "/documentation/en/develop/howto/optimization/index_en.html", "title": "Performance Optimization"}, {"content": ["load", "code", "batches", "should", "to", "train_reader", "main", "details", "do", "python", "fluid", "not", "endverbatim", "like", "button", "cmd", "Warning", "'All", "timeline.py", "Run"], "path": "/documentation/en/develop/howto/optimization/timeline.html", "title": "how to use timeline tool to do profile"}, {"content": ["Training", "for", "vgg16", "Distributed", "Benchmark", "Cluster", "\u00c2\u00b6", "Performance"], "path": "/documentation/en/develop/howto/optimization/benchmark/index_en.html", "title": "Benchmark"}, {"content": ["5GB", "over", "trainer/pserver", "v1.6.2", "per", "CPU", "TN", "We", "20", "Kubernetes", "Benchmark", "Fluid", "Efficiency", "Cases", "By", "Training", "Setup", "mini-batch", "Batch", "Variable"], "path": "/documentation/en/develop/howto/optimization/benchmark/README.html", "title": "Cluster Training Benchmark"}, {"content": ["77.47", "59.20", "Fuild", "190.20", "detail", "Test", "CPU", "247.40", "GLOG_v=3", "20", "fix", "cache", "distributed", "Fluid", "should", "environment", "to", "add", "samples", "under"], "path": "/documentation/en/develop/howto/optimization/benchmark/vgg16/README.html", "title": "Performance for Distributed vgg16"}, {"content": ["represent", "all", "less", "being", "assigned", "has_key", "target", "employ", "Usage", "isinstance", "its", "before", "whose", "Notice", "myErrorClip", "instance", "Here", "should", "to", "only"], "path": "/documentation/en/develop/howto/performance/error_clip.html", "title": "Error Clip"}, {"content": ["xrange", "all", "executed", "object", "text", "accurate", "global", "synchronous", "speedup", "RecordEvent", "ending", "disabled", "timing", "code", "including", "during", "mentioned", "yep", "kCPU", "CPU"], "path": "/documentation/en/develop/howto/performance/profiler.html", "title": "Introduction"}, {"content": ["fwd_primitive", "all", "is_cpu_place", "prim", "help", "just", "dst", "device_context", "satisfied", "best", "four", "createMemory", "executed", "forward_training", "paddle", "GetLibrary", "Test", "follow", "Sulution", "fwd_PD"], "path": "/documentation/en/develop/howto/third_party/mkldnn_fluid.html", "title": "Design Doc: Add MKLDNN Kernel in Fluid Operator"}, {"content": ["operations", "all", "executed", "aggregate", "aggregating", "Transpiler", "over", "Multi-GPU", "synchronous", "rank0", "utilize", "assigned", "Cards", "bandwidth", "neighberhood", "during", "plug-in", "mentioned", "all-gather", "its"], "path": "/documentation/en/develop/howto/third_party/paddle_nccl.html", "title": "Design Doc: NCCL support in Paddle Fluid"}, {"content": ["\u81ea\u52a8\u95ee\u7b54\uff08Question", "\u662f\u4fe1\u606f\u68c0\u7d22\u548c\u641c\u7d22\u5f15\u64ce\u7814\u7a76\u7684\u6838\u5fc3\u95ee\u9898\u4e4b\u4e00\uff0c\u901a\u8fc7\u673a\u5668\u5b66\u4e60\u65b9\u6cd5\u5b66\u4e60\u4e00\u4e2a\u5206\u503c\u51fd\u6570\u5bf9\u5f85\u6392\u5e8f\u7684\u5019\u9009\u8fdb\u884c\u6253\u5206\uff0c\u518d\u6839\u636e\u5206\u503c\u7684\u9ad8\u4f4e\u786e\u5b9a\u5e8f\u5173\u7cfb\u3002\u6df1\u5ea6\u795e\u7ecf\u7f51\u7edc\u53ef\u4ee5\u7528\u6765\u5efa\u6a21\u5206\u503c\u51fd\u6570\uff0c\u6784\u6210\u5404\u7c7b\u57fa\u4e8e\u6df1\u5ea6\u5b66\u4e60\u7684LTR\u6a21\u578b\u3002", "\u6392\u5e8f\u5b66\u4e60", "CRF\uff09\u5728\u7279\u5f81\u57fa\u7840\u4e0a\u5b8c\u6210\u5e8f\u5217\u6807\u6ce8\u4efb\u52a1\uff0c\u9010\u6e10\u6210\u4e3a\u89e3\u51b3\u5e8f\u5217\u6807\u6ce8\u95ee\u9898\u7684\u6807\u914d\u89e3\u51b3\u65b9\u6848\u3002", "\u7684\u4ec5\u6709\u4e00\u4e9b\u503c\u4e3a0", "\u95ee\u9898-\u8bc1\u636e\uff08Evidence\uff09-\u7b54\u6848", "\u5728\u5e8f\u5217\u5230\u5e8f\u5217\u5b66\u4e60\u4efb\u52a1\u4e2d\uff0c\u6211\u4eec\u9996\u5148\u4ee5\u673a\u5668\u7ffb\u8bd1\u4efb\u52a1\u4e3a\u4f8b\uff0c\u63d0\u4f9b\u4e86\u591a\u79cd\u6539\u8fdb\u6a21\u578b\u4f9b\u5927\u5bb6\u5b66\u4e60\u548c\u4f7f\u7528\u3002\u5305\u62ec\uff1a\u4e0d\u5e26\u6ce8\u610f\u529b\u673a\u5236\u7684\u5e8f\u5217\u5230\u5e8f\u5217\u6620\u5c04\u6a21\u578b\uff0c\u8fd9\u4e00\u6a21\u578b\u662f\u6240\u6709\u5e8f\u5217\u5230\u5e8f\u5217\u5b66\u4e60\u6a21\u578b\u7684\u57fa\u7840\uff1b\u4f7f\u7528Scheduled", "\u4ecd\u6709\u4e00\u5b9a\u8ddd\u79bb\u3002\u4e3a\u4e86\u8861\u91cf\u73b0\u6709NLP\u6280\u672f\u5230", "\u63a8\u8350\u7cfb\u7edf", "\u57fa\u4e8ePairwise\u548cListwise\u7684\u6392\u5e8f\u5b66\u4e60", "11.2", "6.1", "11.3", "\u201c", "\u4e00\u8bfe\uff09\u3002", "10.1", "Sampling\u6539\u5584RNN\u6a21\u578b\u5728\u751f\u6210\u4efb\u52a1\u4e2d\u7684\u9519\u8bef\u7d2f\u79ef\u95ee\u9898\uff1b\u5e26\u5916\u90e8\u8bb0\u5fc6\u673a\u5236\u7684\u795e\u7ecf\u673a\u5668\u7ffb\u8bd1\uff0c\u901a\u8fc7\u589e\u5f3a\u795e\u7ecf\u7f51\u7edc\u7684\u8bb0\u5fc6\u80fd\u529b\uff0c\u6765\u5b8c\u6210\u590d\u6742\u7684\u5e8f\u5217\u5230\u5e8f\u5217\u5b66\u4e60\u4efb\u52a1\u3002\u9664\u673a\u5668\u7ffb\u8bd1\u4efb\u52a1\u4e4b\u5916\uff0c\u6211\u4eec\u4e5f\u63d0\u4f9b\u4e86\u4e00\u4e2a\u57fa\u4e8e\u6df1\u5c42LSTM\u7f51\u7edc\u751f\u6210\u53e4\u8bd7\u8bcd\uff0c\u5b9e\u73b0\u540c\u8bed\u8a00\u751f\u6210\u7684\u6a21\u578b\u3002", "\u57fa\u4e8e\u6df1\u5ea6\u56e0\u5b50\u5206\u89e3\u673a\u7684\u70b9\u51fb\u7387\u9884\u4f30\u6a21\u578b", "Residual", "to"], "path": "/models/en/develop/README.cn.html", "title": "models \u7b80\u4ecb"}, {"content": ["all", "semantic", "similarity", "focus", "poetry", "Convolutional", "word-vector", "Residual", "to", "finally", "under", "Training", "advantage", "bottleneck", "tagging", "endverbatim", "applicable", "Entity", "Listwise", "discrete"], "path": "/models/en/develop/README.html", "title": "Introduction to models"}, {"content": ["code", "0", "sentence", "paper", "consists", "pos_size", "data/src_dict", "Convolutional", "20", "use_gpu", "Modify", "train.py", "should", "to", "does", "src_dict_path", "2017", "Grangier", "test_data_path", "then"], "path": "/models/en/develop/conv_seq2seq/README.html", "title": "Convolutional Sequence to Sequence Learning"}, {"content": ["github.com/PaddlePaddle/Paddle/blob/develop/doc/api/v1/data_provider/pydataprovider2_en.rst", "all", "5", "14091123", "Generator", "Embedding\u5c06\u6bcf\u79cd\u503c\u6620\u5c04\u4e3a\u8fde\u7eed\u503c\u7684\u5411\u91cf\u3002", "\u4e2d", "app_id", "\u90e8\u5206\uff0c\u53ef\u4ee5\u8f93\u5165\u5f88", "not", "X", "\u9700\u8981\u5148\u626b\u63cf\u6570\u636e\u96c6\uff0c\u5f97\u5230\u8be5\u7c7b\u522b\u5bf9\u5e94\u7684\u9879\u96c6\u5408\uff0c\u4e4b\u540e\u624d\u80fd\u5f00\u59cb\u751f\u6210\u7279\u5f81\u3002", "Understanding", "\u786e\u5b9a\u8868\u793a\u7684\u6700\u5927\u7ef4\u5ea6", "features", "24", "concat_sparse_vectors", "21", "22", "\u6211\u4eec\u7684\u5b9e\u9a8c\u6570\u636e\u96c6"], "path": "/models/en/develop/ctr/dataset.html", "title": "\u6570\u636e\u53ca\u5904\u7406"}, {"content": ["\u6a21\u578b\u7684\u4f18\u52bf\u662f\u5bf9\u5927\u89c4\u6a21\u7a00\u758f\u7279\u5f81\u7684\u5bb9\u7eb3\u80fd\u529b\uff0c\u5305\u62ec\u5185\u5b58\u548c\u8ba1\u7b97\u91cf\u7b49\u65b9\u9762\uff0c\u5de5\u4e1a\u754c\u90fd\u6709\u975e\u5e38\u6210\u719f\u7684\u4f18\u5316\u65b9\u6cd5\uff1b", "\u800c\u6a21\u578b\u4e0b\u8fb9\u7684", "gz", "prefix", "\u5bf9\u5e94\u7684\u6570\u503c\u3002", "query", "\uff1a", "\u4e2a\u8f93\u5165\uff0c\u5206\u522b\u662f", "feeding", "use_gpu", "\u8bad\u597d\u7684\u6a21\u578b\u53ef\u4ee5\u7528\u6765\u9884\u6d4b\u65b0\u7684\u6570\u636e\uff0c", "test_data_path", "\u6570\u636e\u548c\u4efb\u52a1\u62bd\u8c61", "\u6a21\u578b\u7ed3\u6784\u5bf9\u6bd4", "0", "init", "\u8fdb\u884c\u5904\u7406\uff0c\u5177\u4f53\u4f7f\u7528\u65b9\u6cd5\u53c2\u8003\u5982\u4e0b\u8bf4\u660e\uff1a", "activation", "OUTPUT_DIR", "\u70b9\u51fb\u7387\u9884\u4f30"], "path": "/models/en/develop/ctr/README.cn.html", "title": "\u70b9\u51fb\u7387\u9884\u4f30"}, {"content": ["code", "results", "prefix", "go", "query", "\uff1a", "layers", "feeding", "Business", "use_gpu", "test_data_path", "Trees", "\u6570\u636e\u548c\u4efb\u52a1\u62bd\u8c61", "0", "init", "Training", "activation", "gz", "OUTPUT_DIR", "\uff1aClassification"], "path": "/models/en/develop/ctr/README.html", "title": "Click-Through Rate Prediction"}, {"content": ["code", "0", "proposed", "Ye", "both", "not", "paper", "paddle", "during", "90", "Criteo", "networks", "row", "raw", "1", "26", "data/raw", "based", "Linear", "parameters"], "path": "/models/en/develop/deep_fm/README.html", "title": "Deep Factorization Machine for Click-Through Rate prediction"}, {"content": ["semantic", "similarity", "\u888b", "prefix", "last_seq", "layers", "build_classification_or_regression_model", "test_data_path", "to", "\u4e4b\u95f4\u7684\u5927\u5c0f\u5173\u7cfb", "\u5206\u522b\u6253\u5206\u2014\u2014", "\u56de\u5f52", "P", "\u6574\u4f53\u4e0a\u7b97\u6761\u4ef6\u6982\u7387\u7528\u5bf9\u6570\u4f3c\u7136\u51fd\u6570\u4f5c\u4e3a\u635f\u5931\uff0c\u8fd9\u4e5f\u5c31\u662f\u56fe", "dim", "endverbatim", "0\u3002\u5b9e\u73b0\u8bf7\u53c2\u8003", "list", "Rank", "vector"], "path": "/models/en/develop/dssm/README.cn.html", "title": "\u6df1\u5ea6\u7ed3\u6784\u5316\u8bed\u4e49\u6a21\u578b (Deep Structured Semantic Models, DSSM)"}, {"content": ["code", "semantic", "similarity", "prefix", "last_seq", "query", "layers", "whose", "calculate", "test_data_path", "0", "finally", "Training", "activation", "P", "are\uff1a", "dim", "school", "level", "list"], "path": "/models/en/develop/dssm/README.html", "title": "Deep Structured Semantic Models (DSSM)"}, {"content": ["Currently", "Paddle", "are", "in", "still", "use", "codes", "APIs", "Models", "Fluid", "under", "development", "verbatim", "that", "models", "collection", "active", "The", "a", "endverbatim"], "path": "/models/en/develop/fluid/README.html", "title": "Paddle Fluid Models"}, {"content": ["code", "mnist_model.py", "results", "Tutorials", "ICLR", "contributions", "Keras", "Fluid", "to", "asterisks", "under", "adversary.py", "Moosavi-Dezfooli", "endverbatim", "Adversary", "Testing", "form", "small", "Szegedy", "MXNet"], "path": "/models/en/develop/fluid/adversarial/README.html", "title": "Advbox"}, {"content": ["precision:0.00474094141551", "recall:0.00727476217124", "\u5f00\u59cb\u9884\u6d4b\u3002", "precision:0", "\u7684", "\u8bad\u7ec3", "precision:0.0228873239437", "recall:0.00762112139358", "10-B", "precision:0.0120967741935", "endverbatim", "precision:0.000796178343949", "f1:0.00287769784173", "pass_id:2", "pass_id:0", "pass_id:1", "Test", "7793", "recall:0.0059880239521", "\u9884\u6d4b"], "path": "/models/en/develop/fluid/chinese_ner/README.html", "title": "\u4f7f\u7528ParallelExecutor\u7684\u4e2d\u6587\u547d\u540d\u5b9e\u4f53\u8bc6\u522b\u793a\u4f8b"}, {"content": ["and", "engine", "right", "tqdm", "one-dimensional", "Training", "is", "pass", "positioned", "back", "up", "How", "single", "Therefore", "in", "CartPole-v0", "The", "as", "mountain", "use"], "path": "/models/en/develop/fluid/DeepQNetwork/README.html", "title": "Reproduce DQN model"}, {"content": ["all", "code", "semantic", "paper", "pretrained_model", "still", "num_layers=50", "layers", "We", "password", "val_list.txt", "Untar", "train.py", "train/n04596742/n04596742_3032.jpeg", "under", "include", "into", "2012", "Training", "unzip"], "path": "/models/en/develop/fluid/image_classification/README.html", "title": "SE-ResNeXt for image classification"}, {"content": ["and", "code", "inference", "file", "Prepare", "show", "demo", "is", "models.caffe/alexnet", "caffemodel", "installed", "models.caffe/resnet", "How", "as", "stored", "Caffe", "trained", "in", "need", "converting"], "path": "/models/en/develop/fluid/image_classification/caffe2fluid/examples/imagenet/README.html", "title": "How to use"}, {"content": ["and", "show", "lenet.caffemodel", "demo", "eg", "How", "'mnist", "download", "converting", "models.caffe/lenet", "use", "prepare", "lenet.prototxt", "contains", "environment", "to", "caffe", "which", "lenet", "run"], "path": "/models/en/develop/fluid/image_classification/caffe2fluid/examples/mnist/README.html", "title": "How to use"}, {"content": ["ppl:335.877", "ppl:130.092", "ppl:211.507", "from", "ppl:203.581", "ppl:324.272", "hid_size", "model/epoch_6", "base_lr", "step:500", "ppl:364.404", "GPU", "\u5f53\u524d\u652f\u6301\u7684\u53c2\u6570\u53ef\u53c2\u89c1", "saved", "step:1400", "\u81ea\u5b9a\u4e49\u7f51\u7edc\u7ed3\u6784", "ppl:156.251", "20", "test_reader", "train.py"], "path": "/models/en/develop/fluid/language_model/README.html", "title": "\u8bed\u8a00\u6a21\u578b"}, {"content": ["all", "code", "Kaiser", "Systems", "still", "Jakob", "6000", "title", "implementation", "Fluid", "the", "Illia", "under", "2017", "L", "TODO", "6010", "endverbatim", "Noam", "arxiv"], "path": "/models/en/develop/fluid/neural_machine_translation/transformer/README.html", "title": "Attention is All You Need: A Paddle Fluid implementation"}, {"content": ["github.com/cocodataset/cocoapi.git", "Visualization", "code", "test/val", "'annotations/instances_minival2017.json", "skip", "global", "'coco2014", "own", "file", "eval_coco_map", "Test", "eval_coco_map.py", "pretrained_model", "VOC07+12", "ap_version", "We", "equation-0", "based", "'integral"], "path": "/models/en/develop/fluid/object_detection/README.html", "title": "SSD Object Detection"}, {"content": ["\u53ef\u67e5\u770b\u53c2\u6570\u8be6\u7ec6\u8bf4\u660e\u3002", "models/model_00052_15000", "help", "\u5728\u8be5\u811a\u672c\u4e2d\u5b9a\u4e49\u4e86\u8bad\u7ec3\u7f51\u7edc\u3001\u9884\u6d4b\u7f51\u7edc\u548cevaluate\u7f51\u7edc\u3002", "Character", "test_images", "\u8868\u793a\u4e2d\u6587\u5b57\u7b26\u9017\u53f7\u7684\u7d22\u5f15\u3002", "\u5bf9\u5e94\u7684label\u5982\u4e0b\u6240\u793a\uff1a", "\u5728\u8bc4\u4f30\u65f6\uff0c\u8bc4\u4f30\u96c6\u7684\u8def\u5f84\u901a\u8fc7eval.py\u7684\u9009\u9879", "\u2018", "data/test_images/00003.jpg", "input_images_dir", "\u4e0e\u8bad\u7ec3\u96c6\u7c7b\u4f3c", "\u4ece\u6807\u51c6\u8f93\u5165\u8bfb\u53d6\u4e00\u5f20\u56fe\u7247\u7684\u8def\u5f84\uff0c\u5e76\u5bf9\u9f50\u8fdb\u884c\u9884\u6d4b\uff1a", "\u4e24", "use_gpu", "train.py", "data/test.list", "\uff0c\u5176\u683c\u5f0f\u5982\u4e0b\u6240\u793a\uff1a", "\u4e2d\u5b9e\u73b0\u3002"], "path": "/models/en/develop/fluid/ocr_recognition/README.html", "title": "Optical Character Recognition"}, {"content": ["approx", "a_2", "a_1", "\u5927\u591a\u6570\u60c5\u51b5\uff0c\u6211\u4eec\u65e0\u6cd5\u7a77\u4e3e\u51fa\u6240\u6709\u7684", "r_t|s_t", "\u4e3a\u6b63\uff0c\u90a3\u4e48\u6211\u4eec\u5e0c\u671b\u5f53\u524d\u64cd\u4f5c\u7684\u51fa\u73b0\u7684\u6982\u7387", "\u5185\u5bb9\u5206\u4e3a", "entropy", "\u7684\u68af\u5ea6\u65b9\u5411\u66f4\u65b0", "\u2018", "\u6211\u4eec\u7684\u76ee\u6807\u51fd\u6570\u662f", "\u6211\u4eec\u600e\u4e48\u8bc4\u4f30\u4e00\u4e2aplayer", "1", "\u5f62\u5f0f\u5316\u8868\u793a\u5982\u4e0b\uff1a", "\u8868\u793a\u505a\u8fc7\u52a8\u4f5c", "\u4e3a\u540e\u9000\u7684\u6982\u7387\uff0c", "model\uff09", "42", "0", "4"], "path": "/models/en/develop/fluid/policy_gradient/README.html", "title": "Policy Gradient RL by PaddlePaddle"}, {"content": ["pass_id:127", "4.0867705", "being", "9585", "F1_score0.4473684", "caddick", "F1_score0.4097561", "3.6826708", "0.37846154", "\u201c", "TrainSet", "before", "4.1130385", "0.42424244", "pass_f1_score", "\u5728PaddlePaddle", "use_gpu", "discard", "train.py", "9560"], "path": "/models/en/develop/fluid/sequence_tagging_for_ner/README.html", "title": "\u547d\u540d\u5b9e\u4f53\u8bc6\u522b"}, {"content": ["0.850160", "0.026639", "0.137117", "bow_model/epoch11", "avg_cost", "\u4e2d\u81ea\u884c\u6dfb\u52a0\uff0c\u5e76\u8bbe\u7f6e", "0.951280", "\u5b66\u4e60\u7387\u5927\u5c0f", "lstm", "\u8bad\u7ec3\u6570\u636e", "24", "25", "26", "27", "20", "\u5728PaddlePaddle", "22", "23", "bow_model/epoch16", "28"], "path": "/models/en/develop/fluid/text_classification/README.html", "title": "\u6587\u672c\u5206\u7c7b"}, {"content": ["\u4e00", "\u4e0b\u8f7d\u539f\u59cb\u6570\u636e", "\u2502", "help", "\u674e\u767d", "10.4802", "data/dict.txt", "LSTM", "\u591a", "\u4e2d\u6536\u96c6\u7684\u5168\u5510\u8bd7\u4f5c\u4e3a\u8bad\u7ec3\u6570\u636e\uff0c\u5171\u6709\u7ea65.4\u4e07\u9996\u5510\u8bd7\u3002", "\u4e2d\u56fd\u53e4\u8bd7\u751f\u6210", "batch\u5927\u5c0f\uff0c\u9ed8\u8ba4\u4e3a1", "\u8bad\u7ec3\u6570\u636e\u8def\u5f84", "bi-directional", "data/raw", "use_gpu", "train.py", "to", "trainer\u6570\u76ee\uff0c\u9ed8\u8ba4\u4e3a1", "\u89c0\u7375"], "path": "/models/en/develop/generate_chinese_poetry/README.html", "title": "\u4e2d\u56fd\u53e4\u8bd7\u751f\u6210"}, {"content": ["\u2502", "words", "data/word_vocab.txt", "loss\uff1a\u5b9a\u4e49\u591a\u7c7b\u4ea4\u53c9\u71b5\u4f5c\u4e3a\u6a21\u578b\u7684\u635f\u5931\u51fd\u6570\u3002", "\u7b2c\u4e8c", "\u5047\u5982\u6307\u5b9a\u4e86", "\uff1a\u6307\u5b9a\u8f93\u51fa\u6587\u4ef6\u8def\u5f84\uff0c\u751f\u6210\u7ed3\u679c\u5c06\u5199\u5165\u6b64\u6587\u4ef6\u3002", "\u4e2d\u7684\u914d\u7f6e\uff0c\u7ec6\u8282\u548c\u9002\u914d\u5de5\u4f5c\u8be6\u60c5\u5982\u4e0b\uff1a", "\uff1a\u6307\u5b9a\u6bcf\u4e00\u53e5\u751f\u6210\u7684\u8bdd\u6700\u957f\u957f\u5ea6\uff0c\u5982\u679c\u6a21\u578b\u65e0\u6cd5\u751f\u6210\u51fa", "lstm", "\uff1a", "\u201c", "\u6784\u5efa\u8bcd\u5178\u65f6\uff0c\u4f1a\u81ea\u52a8\u52a0\u5165\u4e24\u4e2a\u7279\u6b8a\u7b26\u53f7\uff1a", "\u4e2d\u5b9a\u4e49\u4e86\u6a21\u578b\u7684\u53c2\u6570\u53d8\u91cf\u3002", "\u8bcd\u6027\u6807\u6ce8\u3001\u53e5\u6cd5\u5206\u6790\u3001\u8bed\u97f3\u8bc6\u522b", "1", "\u8bad\u7ec3\u6a21\u578b\u7684\u65b9\u5f0f\uff0c\u5373\uff1a\u628a\u6bcf\u4e2a\u6c49\u5b57\u5f53\u505a\u4e00\u4e2a\u8bcd\uff0c\u5e38\u7528\u6c49\u5b57\u4e5f\u5c31\u51e0\u5343\u4e2a\uff0c\u4f7f\u5f97\u5b57\u5178\u7684\u5927\u5c0f\u4e0d\u4f1a\u592a\u5927\u3001\u4e0d\u4f1a\u4e22\u5931\u592a\u591a\u4fe1\u606f\uff0c\u4f46\u6c49\u8bed\u4e2d\u540c\u4e00\u4e2a\u5b57\u5728\u4e0d\u540c\u8bcd\u4e2d\u8bed\u4e49\u76f8\u5dee\u5f88\u5927\uff0c\u6709\u65f6\u5bfc\u81f4\u6a21\u578b\u6548\u679c\u4e0d\u7406\u60f3\u3002\u5efa\u8bae\u591a\u8bd5\u8bd5\u3001\u6839\u636e\u5b9e\u9645\u60c5\u51b5\u9009\u62e9\u662f", "\u6309\u9700\u8c03\u6574", "Beam", "train_file"], "path": "/models/en/develop/generate_sequence_by_rnn_lm/README.html", "title": "\u4f7f\u7528\u5faa\u73af\u795e\u7ecf\u7f51\u8bed\u8a00\u6a21\u578b\u751f\u6210\u6587\u672c"}, {"content": ["\u8bf7\u66f4\u65b0", "and", "glove-path", "train.log", "image", "Reader", "year", "tee", "\u5168\u7403\u6807\u51c6\u5316\u9605\u8bfb\u5668", "cd", "infer.log", "\u9884\u5904\u7406\u5e76\u7279\u5f81\u5316\u6570\u636e\uff1a", "in", "featurize.py", "John", "\u5b89\u88c5", "data/glove.840B.300d.txt", "Jonathan", "data_dir", "Language"], "path": "/models/en/develop/globally_normalized_reader/README.cn.html", "title": "\u5168\u7403\u6807\u51c6\u5316\u9605\u8bfb\u5668"}, {"content": ["all", "code", "paper", "data/glove.840B.300d.txt", "title", "EMNLP", "train.py", "to", "2017", "then", "Training", "Installation", "get", "python", "Configurate", "modifying", "raiman2015gnr", "Inferring", "endverbatim", "necessary"], "path": "/models/en/develop/globally_normalized_reader/README.html", "title": "Globally Normalized Reader"}, {"content": ["02d", "\uff0c\u7136\u540e\u4f9d\u636e\u7a97\u53e3\u5927\u5c0f\uff08\u672c\u6587\u4e3a5\uff09\uff0c\u4ece\u5934\u5230\u5c3e\u6bcf\u6b21\u5411\u53f3\u6ed1\u52a8\u7a97\u53e3\u5e76\u751f\u6210\u4e00\u6761\u6570\u636e\u3002\u5982", "ngram_lm", "PTB\u6570\u636e", "\uff0c\u5206\u522b\u662f\u7b2c\u4e00\u4e2a\u5230\u7b2c\u56db\u4e2a\u8bcd\u5728\u5b57\u5178\u4e2d\u7684id\uff0cPaddlePaddle\u4f1a\u8fdb\u4e00\u6b65\u5c06\u8be5list\u8f6c\u5316\u6210", "predict_lbl", "infer.py", "results", "Param", "\u4f5c\u4e3a\u771f\u5b9e\u8bcd\uff0c\u65b9\u4fbf\u8bc4\u4f30\u3002\u89e3\u7801\u51fd\u6570\u7684\u8f93\u5165\u662f\u4e00\u4e2abatch\u6837\u672c\u7684\u9884\u6d4b\u6982\u7387\u4ee5\u53ca\u8bcd\u8868\u7684\u5927\u5c0f\uff0c\u91cc\u9762\u7684\u5faa\u73af\u662f\u5bf9\u6bcf\u6761\u6837\u672c\u7684\u8f93\u51fa\u6982\u7387\u8fdb\u884c\u89e3\u7801\uff0c\u89e3\u7801\u65b9\u5f0f\u5c31\u662f\u6309\u7167\u5de60\u53f31\u7684\u51c6\u5219\uff0c\u4e0d\u65ad\u904d\u5386\u8def\u5f84\uff0c\u76f4\u81f3\u5230\u8fbe\u53f6\u5b50\u8282\u70b9\u3002", "layer\u7684\u8f93\u5165\uff0c\u4e00\u4e2a\u5c01\u88c5\u6837\u4f8b\u5982\u4e0b\uff1a", "hidden_layer", "Bengio", "initial_std", "\u53c2\u6570\u542b\u4e49\u5982\u4e0b\uff1a", "label", "go", "\uff0c\u8be6\u7ec6\u7406\u8bba\u7ec6\u8282\u53ef\u53c2\u7167\u8bba\u6587", "\u53c2\u8003\u6587\u732e", "\u5b89\u88c5\u6587\u6863"], "path": "/models/en/develop/hsigmoid/README.html", "title": "Hsigmoid\u52a0\u901f\u8bcd\u5411\u91cf\u8bad\u7ec3"}, {"content": ["load", "'Paddle_ResNet50.tar.gz", "is", "Inception-ResNet-v2", "\u901a\u8fc7", "dropout_rate", "128000", "cross_entropy_cost", "Test", "init_from_tar", "\u6307\u5b9a\u8f93\u5165\u5c42", "\uff1a", "L2Regularization", "Label", "1", "EndIteration", "224", "test_reader", "parameters", "\u4f7f\u7528ResNet\u6a21\u578b"], "path": "/models/en/develop/image_classification/README.html", "title": "\u56fe\u50cf\u5206\u7c7b"}, {"content": ["fc_layer", "\u6982\u7387", "\u7684\u503c\u4f20\u5165\uff0c\u8fd9\u6837\u5728\u547d\u540d\u4fdd\u5b58layer\u4e2d\u7684\u53c2\u6570\u65f6\u5c06\u4f7f\u7528\u76f8\u5e94\u7684layer", "name_map", "\u4e2d\u7684\u503c\u548c\u8be5layer\u7c7b\u578b\u7684\u8c03\u7528\u8ba1\u6570\u6784\u9020layer", "\u5373\u53ef\uff08\u4e00\u65b9\u9762\uff0c\u5bf9\u4e8ePooling\u7b49\u65e0\u9700\u8bad\u7ec3\u7684layer\u4e0d\u9700\u8981\u4fdd\u5b58\uff0c\u6545\u8fd9\u91cc\u6ca1\u6709\u63d0\u4f9b\u8f6c\u6362\u63a5\u53e3\uff1b\u53e6\u4e00\u65b9\u9762\uff0c\u5bf9\u4e8eCaffe\u4e2d\u7684Scale\u7c7b\u522b\u7684layer\uff0c\u7531\u4e8eCaffe\u548cPaddlePaddle\u5728\u5b9e\u73b0\u4e0a\u7684\u4e00\u4e9b\u5dee\u522b\uff0cPaddlePaddle\u4e2d\u7684batch_norm\u5c42\u662fBatchNorm\u548cScale\u5c42\u7684\u590d\u5408\uff0c\u6545\u8fd9\u91cc\u5bf9Scale\u8fdb\u884c\u4e86\u7279\u6b8a\u5904\u7406\uff09\u3002", "img", "\u521d\u59cb\u5316\uff0c\u4ece\u6307\u5b9a\u6587\u4ef6\u52a0\u8f7d\u6a21\u578b", "\u5982\u679c\u4e0d\u5e0c\u671b\u4f7f\u7528\u9ed8\u8ba4\u7684\u547d\u540d\uff0c\u5e76\u4e14\u5728PaddlePaddle\u7f51\u7edc\u914d\u7f6e\u4e2d\u6307\u5b9a\u4e86layer", "\u4e2d\u5c01\u88c5\u4e86\u4f7f\u7528Caffe", "API\u52a0\u8f7d\u4f7f\u7528\u7684\u6a21\u578b\u6587\u4ef6\u3002", "\u3002", "\u4e2d\u5df2\u63d0\u4f9b\u4ee5\u4e0a\u6b65\u9aa4\uff0c\u4fee\u6539\u5176\u4e2d\u6587\u4ef6\u76f8\u5173\u53d8\u91cf\u7684\u503c\u540e\u6267\u884c", "wrap_name_default", "\u9700\u8981\u6ce8\u610f\uff0c\u5728\u6a21\u578b\u8f6c\u6362\u65f6\u4f1a\u5bf9layer\u7684\u53c2\u6570\u8fdb\u884c\u547d\u540d\uff0c\u8fd9\u91cc\u9ed8\u8ba4\u4f7f\u7528PaddlePaddle\u4e2d\u9ed8\u8ba4\u7684layer\u548c\u53c2\u6570\u547d\u540d\u89c4\u5219\uff1a\u4ee5", "\u4e3a\u6b64\uff0c\u5728\u9a8c\u8bc1\u548c\u4f7f\u7528\u8f6c\u6362\u5f97\u5230\u7684\u6a21\u578b\u65f6\uff0c\u7f16\u5199PaddlePaddle\u7f51\u7edc\u914d\u7f6e\u65e0\u9700\u6307\u5b9alayer", "imagenet/ilsvrc_2012_mean.npy", "\u8fdb\u884c\u6a21\u578b\u8f6c\u6362", "ModelConverter", "caffe_pretrained_file"], "path": "/models/en/develop/image_classification/caffe2paddle/README.html", "title": "\u4f7f\u7528\u8bf4\u660e"}, {"content": ["\u76f8\u5e94\u4f4d\u7f6e\u7684\u53c2\u6570\u3002", "\u4f20\u5165", "\uff0c\u4ee5\u6b64\u83b7\u53d6\u5bf9\u5e94", "\u7684\u987a\u5e8f\u4e3a\uff1a\u5148\u53ef\u5b66\u4e60", "saver", "restore", "paddle.layer.XX", "\u6a21\u578b\u8f6c\u6362\u7684\u57fa\u672c\u6d41\u7a0b\u662f\uff1a", "1", "Full", "\u7684Operator\u8bad\u7ec3\u51fa\u7684\u53c2\u6570\u5411PaddlePaddle\u6a21\u578b\u53c2\u6570\u8f6c\u6362\u3002", "\u7684\u540d\u5b57\u6620\u5c04\u4e3a\u6240\u5bf9\u5e94\u7684", "dict", "tf", "\u4e0b\u9762\u662f\u4e00\u4e2a\u7b80\u5355\u7684\u8c03\u7528\u793a\u4f8b\uff1a", "main", "return", "\u76ee\u524d\u4ec5\u652f\u6301\u5c06TensorFlow\u4e2d", "python", "Batch"], "path": "/models/en/develop/image_classification/tf2paddle/README.html", "title": "\u4f7f\u7528\u8bf4\u660e"}, {"content": ["partial", "relevance_score", "\u7684\u7279\u5f81\u7684\u7ef4\u5ea6\u3002\u6bcf\u6761\u8f93\u5165\u6837\u672c\u4e3alabel\uff0c", "K=Z", "1+e", "\u4f8b\u5982\u8c03\u7528\u63a5\u53e3", "\u63a8\u8350\u7cfb\u7edf", "paddle.layer.fc", "\uff0c\u662f\u7279\u5f81\u5411\u91cf\u7684\u5e8f\u5217\u8f93\u5165\uff0c", "left_doc", "gm", "\u603b\u7ed3", "feeding", "\uff0c\u5bf9\u4e8e\u7ed9\u5b9a\u7684\u4e00\u4e2a\u6587\u6863\u5bf9", "0", "1+S", "\u4e8b\u4ef6\u3002", "\uff0c\u5219", "\u6307\u4e0a\u4e00\u5c42\u4e2d\u7684\u6bcf\u4e2a\u8282\u70b9\u90fd\u8fde\u63a5\u5230\u4e0b\u5c42\u7f51\u7edc\u3002\u672c\u4f8b\u5b50\u4e2d\u540c\u6837\u4f7f\u7528", "\u7684\u65b9\u6cd5\u4f18\u5316\u5ea6\u91cf\u6807\u51c6NDCG"], "path": "/models/en/develop/ltr/README.html", "title": "\u6392\u5e8f\u5b66\u4e60(Learning To Rank)"}, {"content": ["\u9ed8\u8ba4\u63d0\u4f9b\u4e86\u4e00\u4e2a\u7ecf\u8fc7\u9884\u5904\u7406\u7684\u3001\u8f83\u5c0f\u89c4\u6a21\u7684", "limited", "data_utils.py", "dict_size", "LayerOutput", "V", "\u7eb8\u5e26", "RNNs\uff09", "Wayne", "WM^B", "\u57fa\u4e8e\u4f4d\u7f6e\u7684\u5bfb\u5740", "\uff0c\u4ee5\u4fdd\u6301\u548c", "LSTM", "\u901a\u8fc7\u57fa\u4e8e\u5185\u5bb9\u7684\u5bfb\u5740\uff0c\u8ba1\u7b97\u5f97\u5230\u8bfb\u5199\u64cd\u4f5c\u7684\u5bfb\u5740\u5f3a\u5ea6\u3002", "Memory\uff0c\u5373\u8bb0\u5fc6\u69fd\u6570\u91cf\u53ef\u53d8", "decoder_target", "Convolutional", "\u4e8e\u662f\uff0c\u6211\u4eec\u6df7\u5408\u4e0a\u8ff0\u7684\u4e09\u79cd\u52a8\u6001\u8bb0\u5fc6\u673a\u5236\uff0c\u5373RNN", "\u5176\u4ed6\u8ba8\u8bba", "\u6307\u7684\u662f\u5411\u91cf\u7ec4\u7684\u5411\u91cf\u4e2a\u6570\u975e\u56fa\u5b9a\uff0c\u800c\u662f\u968f\u7740\u6e90\u8bed\u53e5\u7684\u5b57\u7b26\u6570\u7684\u53d8\u5316\u800c\u53d8\u5316\uff0c\u6570\u91cf\u4e0d\u53d7\u9650\u3002\u5728\u6e90\u8bed\u53e5\u7684\u7f16\u7801\u5b8c\u6210\u65f6\uff0c\u8be5\u5916\u90e8\u5b58\u50a8\u5373\u88ab\u521d\u59cb\u5316\u4e3a\u5404\u5b57\u7b26\u7684\u72b6\u6001\u5411\u91cf\uff0c\u800c\u5728\u5176\u540e\u7684\u6574\u4e2a\u89e3\u7801\u8fc7\u7a0b\u4e2d\u88ab\u8bfb\u53d6\u4f7f\u7528\u3002"], "path": "/models/en/develop/mt_with_external_memory/README.html", "title": "\u5e26\u5916\u90e8\u8bb0\u5fc6\u673a\u5236\u7684\u795e\u7ecf\u673a\u5668\u7ffb\u8bd1"}, {"content": ["dict_size", "ptb", "omega_t|w", "\uff1a", "P_\\theta", "\u51fd\u6570\u6765\u523b\u753b\u5f0f", "use_gpu", "\u8bed\u8a00\u6a21\u578b\u662f\u8bb8\u591a\u81ea\u7136\u8bed\u8a00\u5904\u7406\u4efb\u52a1\u7684\u57fa\u7840\uff0c\u4e5f\u662f\u83b7\u5f97\u8bcd\u5411\u91cf\u8868\u793a\u7684\u4e00\u79cd\u6709\u6548\u65b9\u6cd5\u3002\u795e\u7ecf\u6982\u7387\u8bed\u8a00\u6a21\u578b\uff08Neural", "to", "\u5168\u8fde\u63a5\u9690\u5c42", "\uff09\u7684\u6982\u7387\u3002\u5047\u8bbe\u6765\u81ea\u566a\u58f0\u5206\u5e03\u7684\u8d1f\u7c7b\u6837\u672c\u7684\u6570\u91cf", "Training", "activation", "P", "\uff0c\u518d\u901a\u8fc7", "endverbatim", "unigram", "Probabilistic", "\u4e2a\u8d1f\u7c7b\u6837\u672c\u5b83\u4eec\u7684\u7c7b\u522b\u5206\u522b\u8bb0\u4f5c\uff1a", "\u9700\u8981\u6ce8\u610f\u7684\u662f\uff1a"], "path": "/models/en/develop/nce_cost/README.html", "title": "\u4f7f\u7528\u566a\u58f0\u5bf9\u6bd4\u4f30\u8ba1\u52a0\u901f\u8bed\u8a00\u6a21\u578b\u8bad\u7ec3"}, {"content": ["\u5e8f\u5217\u662f\u8bb8\u591a\u673a\u5668\u5b66\u4e60\u548c\u6570\u636e\u6316\u6398\u4efb\u52a1\u9762\u5bf9\u7684\u4e00\u79cd\u8f93\u5165\u6570\u636e\u7c7b\u578b\uff0c\u4ee5\u81ea\u7136\u8bed\u8a00\u5904\u7406\u4efb\u52a1\u4e3a\u4f8b\uff1a\u53e5\u5b50\u7531\u8bcd\u8bed\u6784\u6210\uff0c\u800c\u591a\u4e2a\u53e5\u5b50\u8fdb\u4e00\u6b65\u6784\u6210\u4e86\u6bb5\u843d\u3002\u56e0\u6b64\uff0c\u6bb5\u843d\u53ef\u4ee5\u770b\u4f5c\u662f\u4e00\u4e2a\u5d4c\u5957\u7684\u5e8f\u5217\uff08\u6216\u8005\u53eb\u4f5c\uff1a\u53cc\u5c42\u5e8f\u5217\uff09\uff0c\u8fd9\u4e2a\u5e8f\u5217\u7684\u6bcf\u4e2a\u5143\u7d20\u53c8\u662f\u4e00\u4e2a\u5e8f\u5217\u3002", "endverbatim", "verbatim", "\u7b80\u4ecb", "\u53cc\u5c42\u5e8f\u5217\u662f", "PaddlePaddle", "\u57fa\u4e8e\u53cc\u5c42\u5e8f\u5217\u7684\u6587\u672c\u5206\u7c7b", "\u80fd\u591f\u5e2e\u52a9\u6211\u4eec\u66f4\u597d\u5730\u63cf\u8ff0\u6bb5\u843d\u3001\u591a\u8f6e\u5bf9\u8bdd\u7b49\u66f4\u4e3a\u590d\u6742\u7684\u6570\u636e\u3002\u4ee5\u53cc\u5c42\u5e8f\u5217\u4f5c\u4e3a\u8f93\u5165\uff0c\u6211\u4eec\u53ef\u4ee5\u8bbe\u8ba1\u4e00\u4e2a\u5c42\u6b21\u5316\u7684\u7f51\u7edc\uff0c\u4ece\u800c\u66f4\u597d\u5730\u5b8c\u6210\u4e00\u4e9b\u590d\u6742\u7684\u4efb\u52a1\u3002", "\u652f\u6301\u7684\u4e00\u79cd\u975e\u5e38\u7075\u6d3b\u7684\u6570\u636e\u7ec4\u7ec7\u65b9\u5f0f\uff0c", "\u4e2d\u4f7f\u7528\u53cc\u5c42\u5e8f\u5217\u3002", "\u672c\u5355\u5143\u5c06\u4ecb\u7ecd\u5982\u4f55\u5728"], "path": "/models/en/develop/nested_sequence/README.html", "title": "\u7b80\u4ecb"}, {"content": ["default", "What", "cnn_cov_group", "0", "LayerOutput", "TEXT", "\u901a\u8fc7", "'_cov_value_bias", "label", "word_dict", "\u6307\u5b9a\u8bad\u7ec3\u914d\u7f6e\u53c2\u6570", "avg_pool", "doc_ids", "\uff1a", "\u7684\u903b\u8f91\u3002\u4ee5\u4e0b\u4ee3\u7801\u7247\u6bb5\u5b9e\u73b0\u4e86\u8bfb\u53d6\u539f\u59cb\u6570\u636e\u8fd4\u56de\u7c7b\u578b\u4e3a\uff1a", "networks", "help", "\u5c06\u53e5\u5b50\u62c6\u89e3\u4e3a\u8bcd\u8bed\uff0c", "group", "Linear"], "path": "/models/en/develop/nested_sequence/text_classification/README.html", "title": "\u57fa\u4e8e\u53cc\u5c42\u5e8f\u5217\u7684\u6587\u672c\u5206\u7c7b"}, {"content": ["WebQA", "code", "evidence", "Wei", "supported", "results", "Jie", "li:2016", "using", "CPU", "equation-0", "Zhou", "title", "should", "2016", "correct_chunks=2335", "arXiv:1607.06275", "Training", "Installation", "python"], "path": "/models/en/develop/neural_qa/README.html", "title": "Neural Recurrent Sequence Labeling Model for Open-Domain Factoid Question Answering"}, {"content": ["Van", "\u7b26\u53f7", "\u4e3a\u6bcf\u4e2a\u8f93\u5165\u6837\u672c\u751f\u6210", "skip", "\u7b2c\u4e8c", "PaddleBook\u4e2d", "last_seq", "LSTM", "if-else", "\u5bf9\u5e94\u7684\u76ee\u6807\u8bed\u8a00\u82f1\u6587\u7ffb\u8bd1\u7ed3\u679c\u4e3a\uff1a", "\u672c\u4f8b\u6240\u7528\u5230\u7684\u6570\u636e\u6765\u81ea", "0", "Step", "\u5728\u9884\u5904\u7406\u9636\u6bb5\uff0c\u51c6\u5907\u6e90\u8bed\u8a00\u4e0e\u76ee\u6807\u8bed\u8a00\u4e92\u8bd1\u7684\u5e73\u884c\u8bed\u6599\u6570\u636e\uff0c\u5e76\u5206\u522b\u6784\u5efa\u6e90\u8bed\u8a00\u548c\u76ee\u6807\u8bed\u8a00\u7684\u8bcd\u5178\uff1b\u5728\u8bad\u7ec3\u9636\u6bb5\uff0c\u7528\u8fd9\u6837\u6210\u5bf9\u7684\u5e73\u884c\u8bed\u6599\u8bad\u7ec3\u6a21\u578b\uff1b\u5728\u6a21\u578b\u6d4b\u8bd5\u9636\u6bb5\uff0c\u8f93\u5165\u4e2d\u6587\u53e5\u5b50\uff0c\u6a21\u578b\u81ea\u52a8\u751f\u6210\u5bf9\u5e94\u7684\u82f1\u8bed\u7ffb\u8bd1\uff0c\u7136\u540e\u5c06\u751f\u6210\u7ed3\u679c\u4e0e\u6807\u51c6\u7ffb\u8bd1\u5bf9\u6bd4\u8fdb\u884c\u8bc4\u4f30\u3002\u5728\u673a\u5668\u7ffb\u8bd1\u9886\u57df\uff0cBLEU", "end_pos", "decoder_group_name", "target_language_next_word", "\u884c\u662f\u67f1\u641c\u7d22\u751f\u6210\u7684", "activation", "nmt_without_att"], "path": "/models/en/develop/nmt_without_attention/README.cn.html", "title": "\u795e\u7ecf\u7f51\u7edc\u673a\u5668\u7ffb\u8bd1\u6a21\u578b"}, {"content": ["all", "code", "semantic", "illustrates", "global", "sub-sequence", "results", "single-step", "Take", "consists", "Van", "LSTM", "saved", "if-else", "row", "calculate", "graph", "Encodes", "subcomponents", "to"], "path": "/models/en/develop/nmt_without_attention/README.html", "title": "Neural Machine Translation Model"}, {"content": ["41G", "help", "TEXT", "infer.py", "Competitions", "not", "train_file_list", "filters", "PROPER", "'173,46", "1", "Challenge2_Training_Task3_Images_GT.zip", "group", "use_gpu", "requirements.txt", "train.py", "\u811a\u672c\u4e2d\u5305\u542b\u4e86\u6a21\u578b\u914d\u7f6e\u548c\u8bad\u7ec3\u76f8\u5173\u7684\u53c2\u6570\u4ee5\u53ca\u5bf9\u5e94\u7684\u8be6\u7ec6\u89e3\u91ca\uff0c\u4ee3\u7801\u7247\u6bb5\u5982\u4e0b\uff1a", "to", "model_save_dir", "8"], "path": "/models/en/develop/scene_text_recognition/README.html", "title": "\u573a\u666f\u6587\u5b57\u8bc6\u522b (STR, Scene Text Recognition)"}, {"content": ["iterator", "LayerOutput", "decay_a", "decay_b", "Scheduled", "\u548c\u76ee\u6807\u5e8f\u5217\u4e0b\u4e00\u4e2a\u5143\u7d20", "epsilon_i=max", "0", "init", "\uff0c\u5219", "\u4f5c\u4e3a\u89e3\u7801\u5668\u8f93\u5165\u3002", "\u4f5c\u4e3a\u751f\u6210\u7684\u8bcd\u8bed\u3002", "activation", "endverbatim", "Jaitly", "False", "Sampling\u9700\u8981\u5728\u5e8f\u5217\u5230\u5e8f\u5217\u6a21\u578b\u7684\u57fa\u7840\u4e0a\u589e\u52a0\u4e00\u4e2a\u8f93\u5165", "decoder_inputs", "vector", "\u5c42\u4f1a\u5728\u771f\u5b9e\u5143\u7d20"], "path": "/models/en/develop/scheduled_sampling/README.html", "title": "Scheduled Sampling"}, {"content": ["\u4e3a\u8fd0\u884c\u672c\u4f8b\uff0c\u8bf7\u9996\u5148\u5728", "\u2502", "from", "over", "infer.py", "Research", "four", "\u6211\u4eec\u5728", "london", "I-PER", "\u4e3a\u8bcd\u5178\u4e2d\u7684\u8bcd\u8bed\u63d0\u4f9b\u9884\u8bad\u7ec3\u597d\u7684\u8bcd\u5411\u91cf", "\uff0c\u672c\u4f8b\u53ea\u8003\u8651Segment", "Test", "\u201c", "Bottou", "1", "\u6587\u4ef6\u4e2d\u653e\u7f6e\u5c11\u6570\u6837\u672c\u7528\u4ee5\u793a\u4f8b\u8f93\u5165\u6570\u636e\u683c\u5f0f\u3002", "\u8fd9\u4e24\u4e2a\u6587\u4ef6\u3002", "parameters", "with"], "path": "/models/en/develop/sequence_tagging_for_ner/README.html", "title": "\u547d\u540d\u5b9e\u4f53\u8bc6\u522b"}, {"content": ["osp", "findall", "\u4f4d\u7f6e\u3002\u8bc4\u4f30\u7ed3\u675f\u4f1a\u8f93\u51fa", "\u8fd9\u91cc\u4e00\u6761\u6807\u6ce8\u6570\u636e\u5305\u62ec\uff1alabel\u3001xmin\u3001ymin\u3001xmax\u3001ymax\u548cis_difficult\uff0cis_difficult\u8868\u793a\u8be5object\u662f\u5426\u4e3a\u96be\u4f8b\uff0c\u5b9e\u9645\u4e2d\u5982\u679c\u4e0d\u9700\u8981\uff0c\u53ea\u9700\u628a\u8be5\u5b57\u6bb5\u7f6e\u96f6\u5373\u53ef\u3002\u81ea\u6709\u6570\u636e\u4e5f\u9700\u8981\u63d0\u4f9b\u5bf9\u5e94\u7684\u89e3\u6790\u903b\u8f91\uff0c\u5047\u8bbe\u6807\u6ce8\u6570\u636e\uff08\u6bd4\u5982image00000_annotation_file_path\uff09\u5b58\u50a8\u683c\u5f0f\u5982\u4e0b\uff1a", "\u4f5c\u4e3a\u57fa\u7840\u7f51\u7edc\u8fdb\u884c\u56fe\u50cf\u7279\u5f81\u63d0\u53d6\u3002\u4f46SSD\u5bf9\u539f\u59cbVGG16\u7f51\u7edc\u505a\u4e86\u4e00\u4e9b\u6539\u53d8\uff1a", "\u7c7b\u522b\u5217\u8868", "use_gpu", "Object", "0", "init", "size\u4e3a2\uff09\u66f4\u6539\u4e3a3x3-s1-p1\uff08kernel\u5927\u5c0f\u4e3a3x3\uff0cstride", "image00001_file_path", "\u5728\u81ea\u6709\u6570\u636e\u4e0a\u8bad\u7ec3PaddlePaddle", "\u53c2\u89c1\u8bad\u7ec3\u9636\u6bb5\u7684\u914d\u7f6e\uff1b\u8c03\u7528", "Classes", "\u9884\u8bad\u7ec3\u6a21\u578b\u51c6\u5907", "endverbatim", "walk_dir", "item", "data_provider.py"], "path": "/models/en/develop/ssd/README.cn.html", "title": "SSD\u76ee\u6807\u68c0\u6d4b"}, {"content": ["represent", "code", "osp", "results", "four", "rectangles", "CPU", "row", "layers", "use_gpu", "Object", "bbox", "environment", "0", "finally", "Now", "312.913876176", "image00001_file_path", "Training", "very"], "path": "/models/en/develop/ssd/README.html", "title": "Single Shot MultiBox Detector (SSD) Object Detection"}, {"content": ["all", "\u6a21\u578b\u5b66\u4e60\u6548\u679c\u8bc4\u4f30\u3002", "\u901a\u8fc7", "LSTM", "\uff1a", "N+1-h", "issues", "poorly", "\u5316\uff0c\u5373\u5c06\u8bcd\u8bed\u8f6c\u5316\u6210\u5728\u5b57\u5178\u4e2d\u7684\u5e8f\u53f7\u3002", "tv", "\u5bf9\u5377\u79ef\u5f97\u5230\u7684\u5404\u4e2a\u7279\u5f81\u56fe\u5206\u522b\u8fdb\u884c\u6700\u5927\u6c60\u5316\u64cd\u4f5c\u3002\u7531\u4e8e\u7279\u5f81\u56fe\u672c\u8eab\u5df2\u7ecf\u662f\u5411\u91cf\uff0c\u56e0\u6b64\u8fd9\u91cc\u7684\u6700\u5927\u6c60\u5316\u5b9e\u9645\u4e0a\u5c31\u662f\u7b80\u5355\u5730\u9009\u51fa\u5404\u4e2a\u5411\u91cf\u4e2d\u7684\u6700\u5927\u5143\u7d20\u3002\u5404\u4e2a\u6700\u5927\u5143\u7d20\u53c8\u88ab\u62fc\u63a5\u5728\u4e00\u8d77\uff0c\u7ec4\u6210\u65b0\u7684\u5411\u91cf\uff0c\u663e\u7136\uff0c\u8be5\u5411\u91cf\u7684\u7ef4\u5ea6\u7b49\u4e8e\u7279\u5f81\u56fe\u7684\u6570\u91cf\uff0c\u4e5f\u5c31\u662f\u5377\u79ef\u6838\u7684\u6570\u91cf\u3002\u4e3e\u4f8b\u6765\u8bf4\uff0c\u5047\u8bbe\u6211\u4eec\u4f7f\u7528\u4e86\u56db\u4e2a\u4e0d\u540c\u7684\u5377\u79ef\u6838\uff0c\u5377\u79ef\u4ea7\u751f\u7684\u7279\u5f81\u56fe\u5206\u522b\u4e3a\uff1a", "\u751f\u6210\u5668\u5b9e\u73b0", "seriously", "to", "\u8f93\u51fa\u65e5\u5fd7\u6bcf\u4e00\u884c\u662f\u5bf9\u4e00\u6761\u6837\u672c\u9884\u6d4b\u7684\u7ed3\u679c\uff0c\u4ee5", "train_reader", "paddle.batch", "\u5168\u8fde\u63a5\u9690\u5c42", "\u6a21\u578b\u8be6\u89e3", "very"], "path": "/models/en/develop/text_classification/README.html", "title": "\u6587\u672c\u5206\u7c7b"}, {"content": ["'_proj_history_clicked_tags", "\u662f\u6a21\u957f\u5f52\u4e00\u5316\u540e\u7684", "\u6574\u4e2a\u63a8\u8350\u7cfb\u7edf\u6709\u4e24\u90e8\u5206\u7ec4\u6210", "\u7684\u7c7b\u522b\uff0c", "feeding", "\u56e0\u4e3a\u5b9e\u65f6\u53ec\u56de\u9700\u8981\u5927\u91cf\u673a\u5668\u8d44\u6e90\uff0c\u8fd9\u8fb9\u4e5f\u53ef\u4ee5\u79bb\u7ebf\u6316\u6398\u4ea7\u51fa\u6570\u636e\uff0c\u7ebf\u4e0a\u53ec\u56de\u4f7f\u7528\u6316\u6398\u597d\u7684\u6570\u636e\u3002\u53ef\u4ee5\u4ea7\u51fa\u6700\u70ed\uff0c\u7528\u6237\u4e2a\u6027\u5316\uff0c\u89c6\u9891\u76f8\u5173\u7b49\u6570\u636e\u3002\u4e0b\u9762\u7684\u793a\u4f8b\u4ea7\u51fa\u4e86\u7528\u6237\u4e2a\u6027\u5316\u6570\u636e\u3002", "\u5386\u53f2\u70b9\u51fb\u89c6\u98911\uff0c\u5386\u53f2\u70b9\u51fb\u89c6\u98912", "Relu2", "Relu1", "window_size=2", "to", "\u672c\u6587\u5bf9", "h\uff0c", "lstm_last", "norm", "5Cfrac", "USER_ID_15", "activation", "reader", "\u7ebf\u4e0a\u4f7f\u7528\u65f6\uff0c\u4e3a\u4fdd\u7559\u7cbe\u5ea6\uff0c\u53ef\u4ee5\u4e0d\u9664\u4ee5"], "path": "/models/en/develop/youtube_recall/README.cn.html", "title": "Youtube DNN\u63a8\u8350\u6a21\u578b"}, {"content": ["represent", "all", "code", "similarity", "shuffle", "dynamic", "existing", "illustrated", "code.google.com/archive/p/word2vec", "initial_std", "consists", "query", "LSTM", "saved", "poorly", "layers", "feeding", "calculate", "prodct", "Relu2"], "path": "/models/en/develop/youtube_recall/README.html", "title": "Deep Neural Networks for YouTube Recommendations"}, {"path": "/api/en/develop/evaluator.html#evaluator", "prefix": "evaluator", "title": "evaluator"}, {"path": "/api/en/develop/evaluator.html#evaluator", "prefix": "evaluator", "title": "evaluator"}, {"path": "/api/en/develop/evaluator.html#chunkevaluator", "prefix": "evaluator", "title": "ChunkEvaluator"}, {"path": "/api/en/develop/evaluator.html#editdistance", "prefix": "evaluator", "title": "EditDistance"}, {"path": "/api/en/develop/evaluator.html#detectionmap", "prefix": "evaluator", "title": "DetectionMAP"}, {"path": "/api/en/develop/executor.html#executor", "prefix": "executor", "title": "executor"}, {"path": "/api/en/develop/executor.html#executor", "prefix": "executor", "title": "executor"}, {"path": "/api/en/develop/executor.html#id1", "prefix": "executor", "title": "Executor"}, {"path": "/api/en/develop/executor.html#global-scope", "prefix": "executor", "title": "global_scope"}, {"path": "/api/en/develop/executor.html#scope-guard", "prefix": "executor", "title": "scope_guard"}, {"path": "/api/en/develop/executor.html#switch-scope", "prefix": "executor", "title": "switch_scope"}, {"path": "/api/en/develop/initializer.html#initializer", "prefix": "initializer", "title": "initializer"}, {"path": "/api/en/develop/initializer.html#initializer", "prefix": "initializer", "title": "initializer"}, {"path": "/api/en/develop/initializer.html#constant", "prefix": "initializer", "title": "Constant"}, {"path": "/api/en/develop/initializer.html#uniform", "prefix": "initializer", "title": "Uniform"}, {"path": "/api/en/develop/initializer.html#normal", "prefix": "initializer", "title": "Normal"}, {"path": "/api/en/develop/initializer.html#xavier", "prefix": "initializer", "title": "Xavier"}, {"path": "/api/en/develop/initializer.html#msra", "prefix": "initializer", "title": "MSRA"}, {"path": "/api/en/develop/initializer.html#constantinitializer", "prefix": "initializer", "title": "ConstantInitializer"}, {"path": "/api/en/develop/initializer.html#uniforminitializer", "prefix": "initializer", "title": "UniformInitializer"}, {"path": "/api/en/develop/initializer.html#normalinitializer", "prefix": "initializer", "title": "NormalInitializer"}, {"path": "/api/en/develop/initializer.html#xavierinitializer", "prefix": "initializer", "title": "XavierInitializer"}, {"path": "/api/en/develop/initializer.html#msrainitializer", "prefix": "initializer", "title": "MSRAInitializer"}, {"path": "/api/en/develop/io.html#io", "prefix": "io", "title": "io"}, {"path": "/api/en/develop/io.html#io", "prefix": "io", "title": "io"}, {"path": "/api/en/develop/io.html#save-vars", "prefix": "io", "title": "save_vars"}, {"path": "/api/en/develop/io.html#save-params", "prefix": "io", "title": "save_params"}, {"path": "/api/en/develop/io.html#save-persistables", "prefix": "io", "title": "save_persistables"}, {"path": "/api/en/develop/io.html#load-vars", "prefix": "io", "title": "load_vars"}, {"path": "/api/en/develop/io.html#load-params", "prefix": "io", "title": "load_params"}, {"path": "/api/en/develop/io.html#load-persistables", "prefix": "io", "title": "load_persistables"}, {"path": "/api/en/develop/io.html#save-inference-model", "prefix": "io", "title": "save_inference_model"}, {"path": "/api/en/develop/io.html#load-inference-model", "prefix": "io", "title": "load_inference_model"}, {"path": "/api/en/develop/io.html#get-inference-program", "prefix": "io", "title": "get_inference_program"}, {"path": "/api/en/develop/layers.html#layers", "prefix": "layers", "title": "layers"}, {"path": "/api/en/develop/layers.html#layers", "prefix": "layers", "title": "layers"}, {"path": "/api/en/develop/layers.html#control-flow", "prefix": "layers", "title": "control_flow"}, {"path": "/api/en/develop/layers.html#split-lod-tensor", "prefix": "layers", "title": "split_lod_tensor"}, {"path": "/api/en/develop/layers.html#merge-lod-tensor", "prefix": "layers", "title": "merge_lod_tensor"}, {"path": "/api/en/develop/layers.html#blockguard", "prefix": "layers", "title": "BlockGuard"}, {"path": "/api/en/develop/layers.html#blockguardwithcompletion", "prefix": "layers", "title": "BlockGuardWithCompletion"}, {"path": "/api/en/develop/layers.html#staticrnnmemorylink", "prefix": "layers", "title": "StaticRNNMemoryLink"}, {"path": "/api/en/develop/layers.html#whileguard", "prefix": "layers", "title": "WhileGuard"}, {"path": "/api/en/develop/layers.html#while", "prefix": "layers", "title": "While"}, {"path": "/api/en/develop/layers.html#lod-rank-table", "prefix": "layers", "title": "lod_rank_table"}, {"path": "/api/en/develop/layers.html#max-sequence-len", "prefix": "layers", "title": "max_sequence_len"}, {"path": "/api/en/develop/layers.html#topk", "prefix": "layers", "title": "topk"}, {"path": "/api/en/develop/layers.html#lod-tensor-to-array", "prefix": "layers", "title": "lod_tensor_to_array"}, {"path": "/api/en/develop/layers.html#array-to-lod-tensor", "prefix": "layers", "title": "array_to_lod_tensor"}, {"path": "/api/en/develop/layers.html#increment", "prefix": "layers", "title": "increment"}, {"path": "/api/en/develop/layers.html#array-write", "prefix": "layers", "title": "array_write"}, {"path": "/api/en/develop/layers.html#create-array", "prefix": "layers", "title": "create_array"}, {"path": "/api/en/develop/layers.html#less-than", "prefix": "layers", "title": "less_than"}, {"path": "/api/en/develop/layers.html#array-read", "prefix": "layers", "title": "array_read"}, {"path": "/api/en/develop/layers.html#shrink-memory", "prefix": "layers", "title": "shrink_memory"}, {"path": "/api/en/develop/layers.html#array-length", "prefix": "layers", "title": "array_length"}, {"path": "/api/en/develop/layers.html#ifelse", "prefix": "layers", "title": "IfElse"}, {"path": "/api/en/develop/layers.html#dynamicrnn", "prefix": "layers", "title": "DynamicRNN"}, {"path": "/api/en/develop/layers.html#conditionalblock", "prefix": "layers", "title": "ConditionalBlock"}, {"path": "/api/en/develop/layers.html#staticrnn", "prefix": "layers", "title": "StaticRNN"}, {"path": "/api/en/develop/layers.html#reorder-lod-tensor-by-rank", "prefix": "layers", "title": "reorder_lod_tensor_by_rank"}, {"path": "/api/en/develop/layers.html#paralleldo", "prefix": "layers", "title": "ParallelDo"}, {"path": "/api/en/develop/layers.html#print", "prefix": "layers", "title": "Print"}, {"path": "/api/en/develop/layers.html#device", "prefix": "layers", "title": "device"}, {"path": "/api/en/develop/layers.html#get-places", "prefix": "layers", "title": "get_places"}, {"path": "/api/en/develop/layers.html#io", "prefix": "layers", "title": "io"}, {"path": "/api/en/develop/layers.html#data", "prefix": "layers", "title": "data"}, {"path": "/api/en/develop/layers.html#blockguardserv", "prefix": "layers", "title": "BlockGuardServ"}, {"path": "/api/en/develop/layers.html#listenandserv", "prefix": "layers", "title": "ListenAndServ"}, {"path": "/api/en/develop/layers.html#send", "prefix": "layers", "title": "Send"}, {"path": "/api/en/develop/layers.html#nn", "prefix": "layers", "title": "nn"}, {"path": "/api/en/develop/layers.html#fc", "prefix": "layers", "title": "fc"}, {"path": "/api/en/develop/layers.html#embedding", "prefix": "layers", "title": "embedding"}, {"path": "/api/en/develop/layers.html#dynamic-lstm", "prefix": "layers", "title": "dynamic_lstm"}, {"path": "/api/en/develop/layers.html#dynamic-lstmp", "prefix": "layers", "title": "dynamic_lstmp"}, {"path": "/api/en/develop/layers.html#dynamic-gru", "prefix": "layers", "title": "dynamic_gru"}, {"path": "/api/en/develop/layers.html#gru-unit", "prefix": "layers", "title": "gru_unit"}, {"path": "/api/en/develop/layers.html#linear-chain-crf", "prefix": "layers", "title": "linear_chain_crf"}, {"path": "/api/en/develop/layers.html#crf-decoding", "prefix": "layers", "title": "crf_decoding"}, {"path": "/api/en/develop/layers.html#cos-sim", "prefix": "layers", "title": "cos_sim"}, {"path": "/api/en/develop/layers.html#cross-entropy", "prefix": "layers", "title": "cross_entropy"}, {"path": "/api/en/develop/layers.html#square-error-cost", "prefix": "layers", "title": "square_error_cost"}, {"path": "/api/en/develop/layers.html#accuracy", "prefix": "layers", "title": "accuracy"}, {"path": "/api/en/develop/layers.html#chunk-eval", "prefix": "layers", "title": "chunk_eval"}, {"path": "/api/en/develop/layers.html#sequence-conv", "prefix": "layers", "title": "sequence_conv"}, {"path": "/api/en/develop/layers.html#conv2d", "prefix": "layers", "title": "conv2d"}, {"path": "/api/en/develop/layers.html#sequence-pool", "prefix": "layers", "title": "sequence_pool"}, {"path": "/api/en/develop/layers.html#pool2d", "prefix": "layers", "title": "pool2d"}, {"path": "/api/en/develop/layers.html#batch-norm", "prefix": "layers", "title": "batch_norm"}, {"path": "/api/en/develop/layers.html#layer-norm", "prefix": "layers", "title": "layer_norm"}, {"path": "/api/en/develop/layers.html#beam-search-decode", "prefix": "layers", "title": "beam_search_decode"}, {"path": "/api/en/develop/layers.html#conv2d-transpose", "prefix": "layers", "title": "conv2d_transpose"}, {"path": "/api/en/develop/layers.html#sequence-expand", "prefix": "layers", "title": "sequence_expand"}, {"path": "/api/en/develop/layers.html#lstm-unit", "prefix": "layers", "title": "lstm_unit"}, {"path": "/api/en/develop/layers.html#reduce-sum", "prefix": "layers", "title": "reduce_sum"}, {"path": "/api/en/develop/layers.html#reduce-mean", "prefix": "layers", "title": "reduce_mean"}, {"path": "/api/en/develop/layers.html#reduce-max", "prefix": "layers", "title": "reduce_max"}, {"path": "/api/en/develop/layers.html#reduce-min", "prefix": "layers", "title": "reduce_min"}, {"path": "/api/en/develop/layers.html#sequence-first-step", "prefix": "layers", "title": "sequence_first_step"}, {"path": "/api/en/develop/layers.html#sequence-last-step", "prefix": "layers", "title": "sequence_last_step"}, {"path": "/api/en/develop/layers.html#dropout", "prefix": "layers", "title": "dropout"}, {"path": "/api/en/develop/layers.html#split", "prefix": "layers", "title": "split"}, {"path": "/api/en/develop/layers.html#ctc-greedy-decoder", "prefix": "layers", "title": "ctc_greedy_decoder"}, {"path": "/api/en/develop/layers.html#edit-distance", "prefix": "layers", "title": "edit_distance"}, {"path": "/api/en/develop/layers.html#l2-normalize", "prefix": "layers", "title": "l2_normalize"}, {"path": "/api/en/develop/layers.html#matmul", "prefix": "layers", "title": "matmul"}, {"path": "/api/en/develop/layers.html#warpctc", "prefix": "layers", "title": "warpctc"}, {"path": "/api/en/develop/layers.html#sequence-reshape", "prefix": "layers", "title": "sequence_reshape"}, {"path": "/api/en/develop/layers.html#transpose", "prefix": "layers", "title": "transpose"}, {"path": "/api/en/develop/layers.html#im2sequence", "prefix": "layers", "title": "im2sequence"}, {"path": "/api/en/develop/layers.html#nce", "prefix": "layers", "title": "nce"}, {"path": "/api/en/develop/layers.html#beam-search", "prefix": "layers", "title": "beam_search"}, {"path": "/api/en/develop/layers.html#row-conv", "prefix": "layers", "title": "row_conv"}, {"path": "/api/en/develop/layers.html#multiplex", "prefix": "layers", "title": "multiplex"}, {"path": "/api/en/develop/layers.html#label-smooth", "prefix": "layers", "title": "label_smooth"}, {"path": "/api/en/develop/layers.html#roi-pool", "prefix": "layers", "title": "roi_pool"}, {"path": "/api/en/develop/layers.html#ops", "prefix": "layers", "title": "ops"}, {"path": "/api/en/develop/layers.html#mean", "prefix": "layers", "title": "mean"}, {"path": "/api/en/develop/layers.html#mul", "prefix": "layers", "title": "mul"}, {"path": "/api/en/develop/layers.html#reshape", "prefix": "layers", "title": "reshape"}, {"path": "/api/en/develop/layers.html#pad", "prefix": "layers", "title": "pad"}, {"path": "/api/en/develop/layers.html#scale", "prefix": "layers", "title": "scale"}, {"path": "/api/en/develop/layers.html#sigmoid-cross-entropy-with-logits", "prefix": "layers", "title": "sigmoid_cross_entropy_with_logits"}, {"path": "/api/en/develop/layers.html#elementwise-add", "prefix": "layers", "title": "elementwise_add"}, {"path": "/api/en/develop/layers.html#elementwise-div", "prefix": "layers", "title": "elementwise_div"}, {"path": "/api/en/develop/layers.html#elementwise-sub", "prefix": "layers", "title": "elementwise_sub"}, {"path": "/api/en/develop/layers.html#elementwise-mul", "prefix": "layers", "title": "elementwise_mul"}, {"path": "/api/en/develop/layers.html#elementwise-max", "prefix": "layers", "title": "elementwise_max"}, {"path": "/api/en/develop/layers.html#elementwise-min", "prefix": "layers", "title": "elementwise_min"}, {"path": "/api/en/develop/layers.html#elementwise-pow", "prefix": "layers", "title": "elementwise_pow"}, {"path": "/api/en/develop/layers.html#clip", "prefix": "layers", "title": "clip"}, {"path": "/api/en/develop/layers.html#clip-by-norm", "prefix": "layers", "title": "clip_by_norm"}, {"path": "/api/en/develop/layers.html#sequence-softmax", "prefix": "layers", "title": "sequence_softmax"}, {"path": "/api/en/develop/layers.html#sigmoid", "prefix": "layers", "title": "sigmoid"}, {"path": "/api/en/develop/layers.html#logsigmoid", "prefix": "layers", "title": "logsigmoid"}, {"path": "/api/en/develop/layers.html#exp", "prefix": "layers", "title": "exp"}, {"path": "/api/en/develop/layers.html#relu", "prefix": "layers", "title": "relu"}, {"path": "/api/en/develop/layers.html#tanh", "prefix": "layers", "title": "tanh"}, {"path": "/api/en/develop/layers.html#tanh-shrink", "prefix": "layers", "title": "tanh_shrink"}, {"path": "/api/en/develop/layers.html#softshrink", "prefix": "layers", "title": "softshrink"}, {"path": "/api/en/develop/layers.html#sqrt", "prefix": "layers", "title": "sqrt"}, {"path": "/api/en/develop/layers.html#abs", "prefix": "layers", "title": "abs"}, {"path": "/api/en/develop/layers.html#ceil", "prefix": "layers", "title": "ceil"}, {"path": "/api/en/develop/layers.html#floor", "prefix": "layers", "title": "floor"}, {"path": "/api/en/develop/layers.html#round", "prefix": "layers", "title": "round"}, {"path": "/api/en/develop/layers.html#reciprocal", "prefix": "layers", "title": "reciprocal"}, {"path": "/api/en/develop/layers.html#log", "prefix": "layers", "title": "log"}, {"path": "/api/en/develop/layers.html#square", "prefix": "layers", "title": "square"}, {"path": "/api/en/develop/layers.html#softplus", "prefix": "layers", "title": "softplus"}, {"path": "/api/en/develop/layers.html#softsign", "prefix": "layers", "title": "softsign"}, {"path": "/api/en/develop/layers.html#brelu", "prefix": "layers", "title": "brelu"}, {"path": "/api/en/develop/layers.html#leaky-relu", "prefix": "layers", "title": "leaky_relu"}, {"path": "/api/en/develop/layers.html#soft-relu", "prefix": "layers", "title": "soft_relu"}, {"path": "/api/en/develop/layers.html#elu", "prefix": "layers", "title": "elu"}, {"path": "/api/en/develop/layers.html#relu6", "prefix": "layers", "title": "relu6"}, {"path": "/api/en/develop/layers.html#pow", "prefix": "layers", "title": "pow"}, {"path": "/api/en/develop/layers.html#stanh", "prefix": "layers", "title": "stanh"}, {"path": "/api/en/develop/layers.html#hard-shrink", "prefix": "layers", "title": "hard_shrink"}, {"path": "/api/en/develop/layers.html#thresholded-relu", "prefix": "layers", "title": "thresholded_relu"}, {"path": "/api/en/develop/layers.html#hard-sigmoid", "prefix": "layers", "title": "hard_sigmoid"}, {"path": "/api/en/develop/layers.html#swish", "prefix": "layers", "title": "swish"}, {"path": "/api/en/develop/layers.html#tensor", "prefix": "layers", "title": "tensor"}, {"path": "/api/en/develop/layers.html#create-tensor", "prefix": "layers", "title": "create_tensor"}, {"path": "/api/en/develop/layers.html#create-parameter", "prefix": "layers", "title": "create_parameter"}, {"path": "/api/en/develop/layers.html#create-global-var", "prefix": "layers", "title": "create_global_var"}, {"path": "/api/en/develop/layers.html#cast", "prefix": "layers", "title": "cast"}, {"path": "/api/en/develop/layers.html#concat", "prefix": "layers", "title": "concat"}, {"path": "/api/en/develop/layers.html#sums", "prefix": "layers", "title": "sums"}, {"path": "/api/en/develop/layers.html#assign", "prefix": "layers", "title": "assign"}, {"path": "/api/en/develop/layers.html#fill-constant-batch-size-like", "prefix": "layers", "title": "fill_constant_batch_size_like"}, {"path": "/api/en/develop/layers.html#fill-constant", "prefix": "layers", "title": "fill_constant"}, {"path": "/api/en/develop/layers.html#ones", "prefix": "layers", "title": "ones"}, {"path": "/api/en/develop/layers.html#zeros", "prefix": "layers", "title": "zeros"}, {"path": "/api/en/develop/layers.html#id1", "prefix": "layers", "title": "topk"}, {"path": "/api/en/develop/nets.html#nets", "prefix": "nets", "title": "nets"}, {"path": "/api/en/develop/nets.html#nets", "prefix": "nets", "title": "nets"}, {"path": "/api/en/develop/nets.html#simple-img-conv-pool", "prefix": "nets", "title": "simple_img_conv_pool"}, {"path": "/api/en/develop/nets.html#sequence-conv-pool", "prefix": "nets", "title": "sequence_conv_pool"}, {"path": "/api/en/develop/nets.html#glu", "prefix": "nets", "title": "glu"}, {"path": "/api/en/develop/nets.html#scaled-dot-product-attention", "prefix": "nets", "title": "scaled_dot_product_attention"}, {"path": "/api/en/develop/optimizer.html#optimizer", "prefix": "optimizer", "title": "optimizer"}, {"path": "/api/en/develop/optimizer.html#optimizer", "prefix": "optimizer", "title": "optimizer"}, {"path": "/api/en/develop/optimizer.html#sgd", "prefix": "optimizer", "title": "SGD"}, {"path": "/api/en/develop/optimizer.html#momentum", "prefix": "optimizer", "title": "Momentum"}, {"path": "/api/en/develop/optimizer.html#adagrad", "prefix": "optimizer", "title": "Adagrad"}, {"path": "/api/en/develop/optimizer.html#adam", "prefix": "optimizer", "title": "Adam"}, {"path": "/api/en/develop/optimizer.html#adamax", "prefix": "optimizer", "title": "Adamax"}, {"path": "/api/en/develop/optimizer.html#decayedadagrad", "prefix": "optimizer", "title": "DecayedAdagrad"}, {"path": "/api/en/develop/optimizer.html#sgdoptimizer", "prefix": "optimizer", "title": "SGDOptimizer"}, {"path": "/api/en/develop/optimizer.html#momentumoptimizer", "prefix": "optimizer", "title": "MomentumOptimizer"}, {"path": "/api/en/develop/optimizer.html#adagradoptimizer", "prefix": "optimizer", "title": "AdagradOptimizer"}, {"path": "/api/en/develop/optimizer.html#adamoptimizer", "prefix": "optimizer", "title": "AdamOptimizer"}, {"path": "/api/en/develop/optimizer.html#adamaxoptimizer", "prefix": "optimizer", "title": "AdamaxOptimizer"}, {"path": "/api/en/develop/optimizer.html#decayedadagradoptimizer", "prefix": "optimizer", "title": "DecayedAdagradOptimizer"}, {"path": "/api/en/develop/optimizer.html#adadelta", "prefix": "optimizer", "title": "Adadelta"}]
+var index = [{"content": ["\u4e0b\u9762\u7684\u4e00\u4e2a\u5b50\u76ee\u5f55\u91cc\u3002\u5f53\u8fd9\u4e00\u7ae0\u7a33\u5b9a\u4e0b\u6765\uff0c\u6211\u4eec\u4e00\u8d77\u628a\u60a8\u7684\u76ee\u5f55\u632a\u5230\u6839\u76ee\u5f55\u3002", "\u201c", "\u6211\u4eec\u628aJupyter\u3001PaddlePaddle\u3001\u4ee5\u53ca\u5404\u79cd\u88ab\u4f9d\u8d56\u7684\u8f6f\u4ef6\u90fd\u6253\u5305\u8fdb\u4e00\u4e2aDocker", "We", "\u89c6\u9891\u8bfe\u5802", "use_gpu", "\u5e76\u53ef\u4ee5\u7528", "image\u3002\u9605\u8bfb\u548c\u5728\u7ebf\u7f16\u8f91\u672c\u4e66\u8bf7\u5728\u6d4f\u89c8\u5668\u91cc\u8bbf\u95ee", "init", "pending", "\u6765\u751f\u6210\u65b0\u7684Docker", "image\u4e86\u3002\u6240\u4ee5\u60a8\u4e0d\u9700\u8981\u81ea\u5df1\u6765\u5b89\u88c5\u5404\u79cd\u8f6f\u4ef6\uff0c\u53ea\u9700\u8981\u5b89\u88c5Docker\u5373\u53ef\u3002\u5bf9\u4e8e\u5404\u79cdLinux\u53d1\u884c\u7248\uff0c\u8bf7\u53c2\u8003", "\u4f7f\u7528GPU\u8bad\u7ec3", "\u60a8\u73b0\u5728\u5728\u770b\u7684\u8fd9\u672c\u4e66\u662f\u4e00\u672c", "endverbatim", "False", "d", "\u4e2a\u6027\u5316\u63a8\u8350", "p", "latest-gpu"], "path": "/book/en/develop/index.cn.html", "title": "\u6df1\u5ea6\u5b66\u4e60\u5165\u95e8"}, {"content": ["Web", "all", "code", "Fit", "contributed", "move", "chapters", "GPU", "follow", "CPU", "Running", "write", "We", "slow", "somewhere", "use_gpu", "Pull", "except", "to", "init"], "path": "/book/en/develop/index.html", "title": "Deep Learning with PaddlePaddle"}, {"content": ["8:2", "Hyperparameter", "AGE", "global", "\u7684\u6d4b\u8bd5\u96c6\uff0c", "layers", "feeding", "use_gpu", "TAX", "\u8bad\u7ec3\u8fc7\u7a0b", "to", "init", "\u3002\u8fd9\u91cc\u6211\u4eec\u4f7f\u7528\u4e00\u79cd\u5f88\u5e38\u89c1\u7684\u64cd\u4f5c\u65b9\u6cd5\uff1a\u51cf\u6389\u5747\u503c\uff0c\u7136\u540e\u9664\u4ee5\u539f\u53d6\u503c\u8303\u56f4\u3002", "\u3002\u56e0\u6b64\uff0c\u6211\u4eec\u7684\u6a21\u578b\u5c31\u53ef\u4ee5\u8868\u793a\u6210\uff1a", "\uff09\u6a21\u578b\u5f00\u59cb\u8fd9\u4efd\u6559\u7a0b\u3002\u5728\u8fd9\u4e00\u7ae0\u91cc\uff0c\u4f60\u5c06\u4f7f\u7528\u771f\u5b9e\u7684\u6570\u636e\u96c6\u5efa\u7acb\u8d77\u4e00\u4e2a\u623f\u4ef7\u9884\u6d4b\u6a21\u578b\uff0c\u5e76\u4e14\u4e86\u89e3\u5230\u673a\u5668\u5b66\u4e60\u4e2d\u7684\u82e5\u5e72\u91cd\u8981\u6982\u5ff5\u3002", "activation", "print", "every", "fit_a_line/trainer.py", "endverbatim"], "path": "/book/en/develop/01.fit_a_line/index.cn.html", "title": "\u7ebf\u6027\u56de\u5f52"}, {"content": ["represent", "all", "code", "consider", "8:2", "AGE", "global", "results", "highways", "concept", "comments", "computation", "go", "relationships", "feeding", "use_gpu", "1940", "technique", "TAX", "0"], "path": "/book/en/develop/01.fit_a_line/index.html", "title": "Linear Regression"}, {"content": ["AISTATS", "global", "results", "1+e", "0.361920", "computation", "\u4ee3\u8868\u56fe\u7247\u5206\u7c7b\u4e3a\u7b2c", "'pixel", "Convolutional", "layer.data", "\u662f\u8f93\u51fa\uff1a\u5206\u7c7b\u5668\u7684\u8f93\u51fa\u662f10\u7c7b\u6570\u5b57\uff080-9\uff09\uff0c\u5373", "\u7ecf\u8fc7\u7b2c\u4e8c\u4e2a\u9690\u85cf\u5c42\uff0c\u53ef\u4ee5\u5f97\u5230", "use_gpu", "Dan", "\u5206\u522b\u505a\u8bad\u7ec3\u548c\u6d4b\u8bd5\u6570\u636e\u96c6\u3002\u8fd9\u4e24\u4e2a\u51fd\u6570\u5404\u81ea\u8fd4\u56de\u4e00\u4e2areader\u2014\u2014PaddlePaddle\u4e2d\u7684reader\u662f\u4e00\u4e2aPython\u51fd\u6570\uff0c\u6bcf\u6b21\u8c03\u7528\u7684\u65f6\u5019\u8fd4\u56de\u4e00\u4e2aPython", "Platt", "0", "\u591a\u5c42\u611f\u77e5\u5668\u7f51\u7edc\u7ed3\u6784\u56fe", "W_2H_1", "\u5377\u79ef\u5c42\u56fe\u7247"], "path": "/book/en/develop/02.recognize_digits/index.cn.html", "title": "\u8bc6\u522b\u6570\u5b57"}, {"content": ["limited", "all", "code", "consider", "illustrates", "During", "2278-2324", "global", "dynamic", "results", "1+e", "0.361920", "Bureau", "Fig", "computation", "go", "'pixel", "rectangles", "magnifying", "two-dimensional"], "path": "/book/en/develop/02.recognize_digits/index.html", "title": "Recognize Digits"}, {"content": ["\u76f4\u8fde", "nStages", "Van", "Olshausen", "Jia", "\u3002\u964d\u91c7\u6837\u4e5f\u662f\u56fe\u50cf\u5904\u7406\u4e2d\u5e38\u89c1\u7684\u4e00\u79cd\u64cd\u4f5c\uff0c\u53ef\u4ee5\u8fc7\u6ee4\u6389\u4e00\u4e9b\u4e0d\u91cd\u8981\u7684\u9ad8\u9891\u4fe1\u606f\u3002", "\u901a\u8fc7", "results", "R.R", "ICLR", "Krizhevsky", "\u603b\u7ed3", "feeding", "\u5206\u7c7b\u5668\u8bbe\u8ba1", "use_gpu", "Object", "Residual", "GoogleNet", "0", "init"], "path": "/book/en/develop/03.image_classification/index.cn.html", "title": "\u56fe\u50cf\u5206\u7c7b"}, {"content": ["consists", "Van", "feeding", "patches", "Residual", "fine-grained", "swap", "under", "preprocess", "activation", "connects", "updates", "0.6875", "Classification", "vector", "enhance", "direct", "second", "max-pooling", "contributed"], "path": "/book/en/develop/03.image_classification/index.html", "title": "Image Classification"}, {"content": ["dict_size", "similarity", "0.0678709", "1,1", "\u4e3a\u6240\u6709\u8bcd\u8bed\u7684\u8bcd\u5411\u91cf\u8fde\u63a5\u6210\u7684\u5927\u5411\u91cf\uff0c\u8868\u793a\u6587\u672c\u5386\u53f2\u7279\u5f81\uff1b", "\u7136\u540e\uff0c\u6307\u5b9a\u8bad\u7ec3\u76f8\u5173\u7684\u53c2\u6570\uff1a", "Testing", "\uff1a", "\u7531\u4e8e\u5f88\u591a\u8bcd\u6ca1\u6709\u51fa\u73b0\uff0c\u5bfc\u81f4\u77e9\u9635\u6781\u5176\u7a00\u758f\uff0c\u56e0\u6b64\u9700\u8981\u5bf9\u8bcd\u9891\u505a\u989d\u5916\u5904\u7406\u6765\u8fbe\u5230\u597d\u7684\u77e9\u9635\u5206\u89e3\u6548\u679c\uff1b", "\u603b\u7ed3", "0.999591588973999", "\uff0c\u7531\u4e8e\u5176\u5411\u91cf\u6b63\u4ea4\uff0c\u90fd\u8ba4\u4e3a\u8fd9\u4e24\u4e2a\u8bcd\u6beb\u65e0\u76f8\u5173\u6027\u3002", "imikolov", "use_gpu", "\u4e2a\u8bcd\u7684\u6982\u7387\u3002", "0", "init", "\u4e2d\u4ecb\u7ecd\u5982\u4f55\u5b66\u4e60\u4e00\u4e2a\u795e\u7ecf\u5143\u7f51\u7edc\u8868\u793a\u7684\u8bcd\u5411\u91cf\u6a21\u578b\u3002\u6587\u4e2d\u7684\u795e\u7ecf\u6982\u7387\u8bed\u8a00\u6a21\u578b\uff08Neural", "paddle.batch", "0.2069038"], "path": "/book/en/develop/04.word2vec/index.cn.html", "title": "\u8bcd\u5411\u91cf"}, {"content": ["limited", "all", "code", "semantic", "two-dimensional", "iterator", "similarity", "0.0678709", "caused", "results", "1,1", "illustrated", "concept", "Take", "computation", "initial_std", "go", "query", "saved", "dict_size"], "path": "/book/en/develop/04.word2vec/index.html", "title": "Word2Vec"}, {"content": ["copy", "global", "\u81f3\u6b64\uff0c\u6211\u4eec\u7684\u4f18\u5316\u76ee\u6807\u5c31\u662f\u8fd9\u4e2a\u7f51\u7edc\u914d\u7f6e\u4e2d\u7684", "'gender_id", "\u3001\u57fa\u4e8e\u7269\u54c1\uff08Item-Based\uff09\u7684\u63a8\u8350", "\u603b\u7ed3", "\u57fa\u4e8e\u5185\u5bb9\u8fc7\u6ee4\u63a8\u8350", "use_gpu", "Collaborative", "0", "init", "\u6700\u8fd1\u7684", "\u6559\u7a0b\uff09\u5904\u7406\u3002\u6700\u540e\uff0c\u8bad\u7ec3\u65f6\u5c06MLP\u7684\u8f93\u51fa\u7ed9softmax\u505a\u5206\u7c7b\uff0c\u9884\u6d4b\u65f6\u8ba1\u7b97\u7528\u6237\u7684\u7efc\u5408\u7279\u5f81\uff08MLP\u7684\u8f93\u51fa\uff09\u4e0e\u6240\u6709\u89c6\u9891\u7684\u76f8\u4f3c\u5ea6\uff0c\u53d6\u5f97\u5206\u6700\u9ad8\u7684", "Feed", "\u5176\u6b21\uff0c\u8fdb\u884c\u5377\u79ef\u64cd\u4f5c\uff1a\u628a\u5377\u79ef\u6838", "movie_id", "activation", "u'___embedding_layer_1__.w0", "P", "every"], "path": "/book/en/develop/05.recommender_system/index.cn.html", "title": "\u4e2a\u6027\u5316\u63a8\u8350"}, {"content": ["represent", "all", "code", "semantic", "two-dimensional", "Web", "shuffle", "results", "four", "'gender_id", "during", "go", "Personalized", "consider", "Convolutional", "dimensionality", "calculate", "use_gpu", "fixed-length", "to"], "path": "/book/en/develop/05.recommender_system/index.html", "title": "Personalized Recommendation"}, {"content": ["\u3002\u81ea\u7136\u8bed\u8a00\u662f\u4e00\u79cd\u5178\u578b\u7684\u5e8f\u5217\u6570\u636e\uff08\u8bcd\u5e8f\u5217\uff09\uff0c\u8fd1\u5e74\u6765\uff0c\u5faa\u73af\u795e\u7ecf\u7f51\u7edc\u53ca\u5176\u53d8\u4f53\uff08\u5982long", "Van", "forget", "0.294321", "global", "\u63a8\u8350\u7cfb\u7edf", "computation", "\u7f51\u7edc\u53c2\u6570", "\u5f88\u4e0d\u597d\u770b\uff0c\u597d\u50cf\u4e00\u4e2a\u5730\u65b9\u53f0\u7684\u7535\u89c6\u5267", "semantic", "LSTM", "\uff1a", "layers", "\u4e00\u8282\u4ecb\u7ecd\u8fc7\u7684\u6587\u672c\u5377\u79ef\u795e\u7ecf\u7f51\u7edc\uff0c\u4ee5\u53ca", "\u53ca\u5176\u6269\u5c55\u3002\u4e0b\u9762\u4f9d\u6b21\u4ecb\u7ecd\u8fd9\u51e0\u4e2a\u6a21\u578b\u3002", "use_gpu", "\uff08\u5411\u91cf\u8868\u793a\uff09\u53ca\u524d\u4e00\u65f6\u523b\u9690\u5c42\u7684\u72b6\u6001\u503c", "to", "init", "train_reader"], "path": "/book/en/develop/06.understand_sentiment/index.cn.html", "title": "\u60c5\u611f\u5206\u6790"}, {"content": ["represent", "all", "code", "semantic", "forget", "illustrates", "leads", "computation", "consists", "Van", "LSTM", "\uff1a", "issues", "Convolutional", "Turing-complete", "feeding", "use_gpu", "similarity", "to", "init"], "path": "/book/en/develop/06.understand_sentiment/index.html", "title": "Sentiment Analysis"}, {"content": ["semantic", "\u7136\u800c\uff0c\u5b8c\u5168\u53e5\u6cd5\u5206\u6790\u9700\u8981\u786e\u5b9a\u53e5\u5b50\u6240\u5305\u542b\u7684\u5168\u90e8\u53e5\u6cd5\u4fe1\u606f\uff0c\u5e76\u786e\u5b9a\u53e5\u5b50\u5404\u6210\u5206\u4e4b\u95f4\u7684\u5173\u7cfb\uff0c\u662f\u4e00\u4e2a\u975e\u5e38\u56f0\u96be\u7684\u4efb\u52a1\uff0c\u76ee\u524d\u6280\u672f\u4e0b\u7684\u53e5\u6cd5\u5206\u6790\u51c6\u786e\u7387\u5e76\u4e0d\u9ad8\uff0c\u53e5\u6cd5\u5206\u6790\u7684\u7ec6\u5fae\u9519\u8bef\u90fd\u4f1a\u5bfc\u81f4SRL\u7684\u9519\u8bef\u3002\u4e3a\u4e86\u964d\u4f4e\u95ee\u9898\u7684\u590d\u6742\u5ea6\uff0c\u540c\u65f6\u83b7\u5f97\u4e00\u5b9a\u7684\u53e5\u6cd5\u7ed3\u6784\u4fe1\u606f\uff0c", "\u901a\u8fc7", "feature_out", "proposition", "Van", "\uff1a", "\u603b\u7ed3", "feeding", "use_gpu", "test.wsj", "book/label_semantic_roles", "0", "init", "\u56fe3\u662f\u6700\u7ec8\u5f97\u5230\u7684\u6808\u5f0f\u5faa\u73af\u795e\u7ecf\u7f51\u7edc\u7ed3\u6784\u793a\u610f\u56fe\u3002", "CRF\u4ee5\u7b2c3\u6b65\u4e2dLSTM\u5b66\u4e60\u5230\u7684\u7279\u5f81\u4e3a\u8f93\u5165\uff0c\u4ee5\u6807\u8bb0\u5e8f\u5217\u4e3a\u76d1\u7763\u4fe1\u53f7\uff0c\u5b8c\u6210\u5e8f\u5217\u6807\u6ce8\uff1b", "\u8f93\u5165\u6587\u672c\u5e8f\u5217", "predicate", "\u5927\u5bb6\u53ef\u4ee5\u5c1d\u8bd5\u4e0a\u9762\u8fd9\u79cd\u65b9\u6cd5\u3002\u8fd9\u91cc\uff0c\u6211\u4eec\u63d0\u51fa\u4e00\u4e9b\u6539\u8fdb\uff0c\u5f15\u5165\u4e24\u4e2a\u7b80\u5355\u4f46\u5bf9\u63d0\u9ad8\u7cfb\u7edf\u6027\u80fd\u975e\u5e38\u6709\u6548\u7684\u7279\u5f81\uff1a", "role"], "path": "/book/en/develop/07.label_semantic_roles/index.cn.html", "title": "\u8bed\u4e49\u89d2\u8272\u6807\u6ce8"}, {"content": ["proposition", "all", "code", "semantic", "consider", "chain", "illustrates", "skip", "2012", "hidden-to-hidden", "concern", "results", "test_creator", "illustrated", "paddle.layer.fc", "\u4e86", "Fig", "computation", "initial_std", "consists"], "path": "/book/en/develop/07.label_semantic_roles/index.html", "title": "Semantic Role Labeling"}, {"content": ["Van", "dict_size", "\u901a\u8fc7", "name=decoder_group_name", "\u76ee\u6807\u8bed\u8a00\u5b57\u5178\u7ef4\u5ea6", "\u7684\u8ba1\u7b97\u516c\u5f0f\u5982\u4e0b\uff1a", "ICLR", "\uff0c\u5373\u9488\u5bf9\u6bcf\u4e00\u4e2a\u76ee\u6807\u8bed\u8a00\u4e2d\u7684\u8bcd", "\u603b\u7ed3", "\u673a\u5668\u7ffb\u8bd1\u4efb\u52a1\u7684\u751f\u6210\u8fc7\u7a0b\uff0c\u901a\u4fd7\u6765\u8bb2\u5c31\u662f\u6839\u636e\u9884\u5148\u8bad\u7ec3\u7684\u6a21\u578b\u6765\u7ffb\u8bd1\u6e90\u8bed\u8a00\u53e5\u5b50\u3002\u751f\u6210\u8fc7\u7a0b\u4e2d\u7684\u89e3\u7801\u9636\u6bb5\u548c\u4e0a\u8ff0\u8bad\u7ec3\u8fc7\u7a0b\u7684\u6709\u6240\u5dee\u5f02\uff0c\u5177\u4f53\u4ecb\u7ecd\u8bf7\u89c1", "backward_first", "use_gpu", "\u76ee\u6807\u5b57\u5178", "148.444983", "\u884c\u8fde\u63a5\uff0c\u7528't'\u5206\u9694\u3002", "0", "init", "vector\u8868\u793a\u5b58\u5728\u4e24\u4e2a\u95ee\u9898\uff0c1\uff09\u751f\u6210\u7684\u5411\u91cf\u7ef4\u5ea6\u5f80\u5f80\u5f88\u5927\uff0c\u5bb9\u6613\u9020\u6210\u7ef4\u6570\u707e\u96be\uff1b2\uff09\u96be\u4ee5\u523b\u753b\u8bcd\u4e0e\u8bcd\u4e4b\u95f4\u7684\u5173\u7cfb\uff08\u5982\u8bed\u4e49\u76f8\u4f3c\u6027\uff0c\u4e5f\u5c31\u662f\u65e0\u6cd5\u5f88\u597d\u5730\u8868\u8fbe\u8bed\u4e49\uff09\u3002\u56e0\u6b64\uff0c\u9700\u518done-hot", "z_i", "x=\\left"], "path": "/book/en/develop/08.machine_translation/index.cn.html", "title": "\u673a\u5668\u7ffb\u8bd1"}, {"content": ["limited", "all", "code", "semantic", "consider", "breadth", "global", "focus", "computation", "embeding", "query", "LSTM", "research", "CPU", "forget", "layers", "dimensionality", "calculate", "backward_first", "Attention-based"], "path": "/book/en/develop/08.machine_translation/index.html", "title": "Machine Translation"}, {"content": ["code", "just", "github.com/sugyan/tensorflow-mnist", "front", "Usage", "its", "We", "parameters", "Javascript", "0", "Front", "build", "them", "localhost:5000", "familiar", "stop", "topylogy", "s3.us-east-2.amazonaws.com/models.paddlepaddle/end-to-end-mnist/param.tar", "not", "name"], "path": "/book/en/develop/mnist-client/index.html", "title": "MNIST classification by PaddlePaddle"}, {"content": ["load", "all", "code", "stores", "0", "just", "show", "numbers", "text", "supported", "obtain", "save_parameter_to_tar", "already", "0.10060056298971176", "integer", "mentioned", "139", "cases", "layers", "content"], "path": "/book/en/develop/serve/index.html", "title": "Inference Server Example"}, {"content": ["Index"], "path": "/documentation/en/develop/genindex.html", "title": "Index"}, {"content": ["all", "protocol", "square_error_cost", "avg_cost", "shape", "Lookup", "computation", "Evaluates", "Scope", "CPU", "layers", "Fluid", "paddle/fluid/framework/tensor.h", "Base", "Initialize", "input", "has", "variables", "handle", "fluid"], "path": "/documentation/en/develop/read_source.html", "title": "PaddlePaddle Fluid Source Code Overview"}, {"content": ["all", "code", "consider", "BLAS", "saves", "Does", "go", "follow", "disk", "saved", "CPU", "build/python/dist/*.whl", "environment", "to", "Step", "under", "Not", "Frequently", "include", "paddle/paddle/scripts/docker/build.sh"], "path": "/documentation/en/develop/build_and_install/build_from_source_en.html", "title": "Build from Sources"}, {"content": ["all", "code", "proc/cpuinfo", "DockerHub", "edit", "BLAS", "move", "python", "libs", "serving", "work/train.py", "using", "go", "CPU", "OpenBLAS", "before", "one", "We", "ll", "train.py"], "path": "/documentation/en/develop/build_and_install/docker_install_en.html", "title": "Run in Docker Containers"}, {"content": ["all", "code", "Bottom", "Does", "during", "still", "We", "\u521b\u5efaissue", "environment", "to", "choose", "has", "python", "not", "using", "instructions", "approach", "Warning", "Learning", "require"], "path": "/documentation/en/develop/build_and_install/index_en.html", "title": "install and Compile"}, {"content": ["MacOS", "code", "just", "supported", "Access", "using", "still", "\u201c", "cuda8.0_cudnn7_avx_mkl", "suffix", "libraries", "We", "trying", "ll", "2.7.x", "seems", "libc.so", "environment", "to", "only"], "path": "/documentation/en/develop/build_and_install/pip_install_en.html", "title": "Install using pip"}, {"content": ["and", "Management", "Network", "Gradient", "Multi-Device", "Distributed", "Complex", "Memory", "Type", "Core", "Multi-Language", "Update", "Concurrent", "Important", "Goals", "Interface", "Execution", "Data", "Support", "Structure"], "path": "/documentation/en/develop/design/index_en.html", "title": "Design"}, {"content": ["Algorithm", "Gradient", "Averaging", "PaddlePaddle", "Update", "\u00c2\u00b6", "in", "Parameter"], "path": "/documentation/en/develop/design/algorithm/index_en.html", "title": "Gradient Update Algorithm"}, {"content": ["operations", "represent", "all", "code", "Based", "proposed", "Juditsky", "over", "desirable", "obtain", "re-writing", "through", "computation", "using", "mentioned", "speed", "its", "before", "layers", "1"], "path": "/documentation/en/develop/design/algorithm/parameter_average.html", "title": "Averaging Parameter in PaddlePaddle"}, {"content": ["default", "all", "global", "unique", "scalar", "attrs", "computation", "Var", "if-else", "layers", "const", "graph", "description", "languages", "0", "init", "program", "ifelse", "string", "void"], "path": "/documentation/en/develop/design/concepts/block.html", "title": "Design Doc: Block and Scope"}, {"content": ["all", "Feeding", "EOF", "MultiPassReader", "leads", "prefix", "code", "feeding", "const", "ReadNext", "graph", "Fluid", "environment", "to", "program", "friendly", "And", "very", "void", "HasNextOp"], "path": "/documentation/en/develop/design/concepts/cpp_data_feeding.html", "title": "C++ Data Feeding"}, {"content": ["operations", "code", "Also", "layer.fc", "using", "Var", "fine", "O2", "O1", "layers", "We", "compose", "implementation", "with", "explain", "should", "to", "add", "going", "illustrative"], "path": "/documentation/en/develop/design/concepts/functions_operators_layers.html", "title": "Design Doc: Functions, Operators, and Layers"}, {"content": ["and", "Layers", "Functions", "Feeding", "Tensor", "Executor", "PaddlePaddle", "Paddle", "ParallelExecutor", "in", "Scope", "Type", "Function", "Core", "for", "Operators", "Level-of-Detail", "Variable", "Data", "TensorArray"], "path": "/documentation/en/develop/design/concepts/index_en.html", "title": "Core Concepts"}, {"content": ["represent", "What", "just", "show", "numbers", "indeed", "Images", "all", "ending", "underlying", "through", "consists", "revisit", "image", "adopt", "convolution", "Level-of-Detail", "We", "2-level", "640x480"], "path": "/documentation/en/develop/design/concepts/lod_tensor.html", "title": "Design Doc: LoD (Level-of-Detail) Tensor"}, {"content": ["represent", "all", "concept", "results", "computation", "CPU", "duplicated", "graph", "Fluid", "to", "program", "helps", "nodes", "Parallel", "string", "synchronized", "very", "MultiGPUs", "multi-GPUs", "endverbatim"], "path": "/documentation/en/develop/design/concepts/parallel_executor.html", "title": "ParallelExecutor"}, {"content": ["concept", "vars", "global", "true_block", "IfElseOp", "shape", "stored", "attrs", "consists", "20", "parameters", "with", "should", "0", "program", "has", "local", "into", "int32", "return"], "path": "/documentation/en/develop/design/concepts/program.html", "title": "Design Doc: PaddlePaddle Programs"}, {"content": ["all", "code", "global", "cudnn", "go", "lgtest", "depend", "graph", "cc_test", "to", "under", "nodes", "Now", "very", "ops.h", "choice", "cc_binary", "GO", "difference", "nv_library"], "path": "/documentation/en/develop/design/concepts/README.html", "title": "Required CMake Function"}, {"content": ["default", "all", "concept", "just", "but", "global", "Also", "entity", "stored", "relation", "same", "computation", "Var", "Scope", "pointer", "its", "some", "unique", "const", "vars"], "path": "/documentation/en/develop/design/concepts/scope.html", "title": "Design of Scope in Paddle"}, {"content": ["represent", "all", "code", "consider", "iterator", "caused", "global", "dynamic", "four", "manages", "consists", "CPU", "row", "const", "pointing", "to", "tail", "include", "friendly", "resources"], "path": "/documentation/en/develop/design/concepts/tensor.html", "title": "Tensor: An Unified Data Type in PaddlePaddle"}, {"content": ["all", "concept", "dynamic", "splitting", "existing", "scalar", "code", "Var", "Supports", "layers", "whose", "batches", "const", "segments", "presents", "to", "init", "th", "preprocess", "ta"], "path": "/documentation/en/develop/design/concepts/tensor_array.html", "title": "Design for TensorArray"}, {"content": ["operations", "represent", "executed", "from", "TensorDesc", "supported", "concept", "looks", "computation", "consists", "saved", "workers", "distributed", "should", "to", "only", "program", "4", "5", "interprets"], "path": "/documentation/en/develop/design/concepts/var_desc.html", "title": "Design Doc: Var_desc"}, {"content": ["code", "Requirements", "Similarly", "able", "saves", "anyway", "syntax", "held", "pointer", "its", "utilizes", "Lazy", "implementation", "graph", "technique", "should", "Allocation", "to", "only", "blob"], "path": "/documentation/en/develop/design/concepts/variable.html", "title": "Design Doc: Variable"}, {"content": ["limited", "they", "Closes", "being", "supported", "synchronous", "shape", "ch", "through", "to/from", "sent/received", "still", "before", "note", "We", "on", "Channels", "send", "should", "0"], "path": "/documentation/en/develop/design/concurrent/channel.html", "title": "Channel Design"}, {"content": ["represent", "all", "executed", "managed", "0", "just", "parallel_for", "process/node", "saves", "control-flow", "splitting", "syntax", "listen_and_do", "attrs", "questions", "integer", "its", "Executor.Run", "Tensor", "layers"], "path": "/documentation/en/develop/design/concurrent/concurrent_programming.html", "title": "Design Doc: Concurrent Programming with Fluid"}, {"content": ["all", "concept", "illustrates", "control-flow", "focus", "four", "code", "computation", "go", "if-else", "layers", "kqueue", "compose", "unblocked", "graph", "Fluid", "send", "to", "program", "Now"], "path": "/documentation/en/develop/design/concurrent/csp.html", "title": "Design Doc: CSP in PaddlePaddle Fluid"}, {"content": ["all", "executed", "creation", "vars", "less", "being", "captures", "global", "executing", "shape", "Take", "X", "go", "only", "before", "whose", "We", "Backward", "3", "should"], "path": "/documentation/en/develop/design/concurrent/go_op.html", "title": "go_op Design"}, {"content": ["Parallel_Do", "Doc", "Programming", "Fluid", "Concurrent", "\u00c2\u00b6", "Design", "PaddlePaddle", "in", "with"], "path": "/documentation/en/develop/design/concurrent/index_en.html", "title": "Concurrent Programming"}, {"content": ["represent", "all", "executed", "vars", "Imporvement", "over", "AddInput", "kPlaces", "looks", "CPU", "duplicated", "layers", "Inputs", "We", "copy", "20", "decay", "implementation", "use_gpu", "param"], "path": "/documentation/en/develop/design/concurrent/parallel_do.html", "title": "Design Doc: Parallel_Do in PaddlePaddle"}, {"content": ["operations", "fluid.Select.case", "all", "executed", "set", "true", "random", "while_cond", "elementwise_add", "shape", "simplier", "executes", "attrs", "X", "Scope", "channel_102", "its", "channel_101", "fill_constant_109.tmp_0", "whose"], "path": "/documentation/en/develop/design/concurrent/select_op.html", "title": "select_op Design"}, {"content": ["all", "code", "global", "However", "executing", "computation", "consists", "CPU", "layers", "whose", "feeding", "const", "Fluid", "send", "intensive", "to", "finally", "program", "norm", "numpy.uint16"], "path": "/documentation/en/develop/design/data_type/float16.html", "title": "Design Doc: float16"}, {"content": ["float16", "Data", "\u00c2\u00b6", "Design", "Doc", "Type"], "path": "/documentation/en/develop/design/data_type/index_en.html", "title": "Data Type"}, {"content": ["all", "global", "synchronous", "competitive", "Threadpool", "immediately", "find", "before", "whose", "We", "Optimize", "parameters", "Schedule", "distributed", "Update", "send", "than", "better", "to", "finally"], "path": "/documentation/en/develop/design/dist_train/async_update.html", "title": "Design Doc: Asynchronous Update With Distributed Training"}, {"content": ["all", "code", "Take", "computation", "'pixel", "if-else", "layers", "mem_per_pserver", "Kubernetes", "Executors", "graph", "Fluid", "send", "environment", "to", "fluid.layer.fc", "program", "train_reader", "nodes", "Executor.run"], "path": "/documentation/en/develop/design/dist_train/distributed_architecture.html", "title": "Design Doc: Fluid Distributed Training Architecture"}, {"content": ["Service", "illustrates", "when", "both", "symbols", "performance", "Does", "Lookup", "solutions", "Forward", "its", "non-zero", "We", "one", "distributed", "except", "former", "better", "to", "Distributed"], "path": "/documentation/en/develop/design/dist_train/distributed_lookup_table_design.html", "title": "Design Doc: Distributed Lookup Table Operator"}, {"content": ["all", "Norouzi", "results", "computation", "LSTM", "issues", "Schuster", "layers", "depend", "technique", "to", "stable", "worth", "resources", "case", "Training", "worse", "advantage", "very", "N+M"], "path": "/documentation/en/develop/design/dist_train/distributed_traing_review.html", "title": "Parallelism, Asynchronous,  Synchronous, Codistillation"}, {"content": ["Training", "Doc", "Distributed", "Fluid", "Server", "\u00c2\u00b6", "Design", "Architecture", "Table", "Operator", "Lookup", "Parameter"], "path": "/documentation/en/develop/design/dist_train/index_en.html", "title": "Distributed Training"}, {"content": ["operations", "all", "code", "distribute", "find", "mpirun", "bring", "identify", "using", "compiling.Add", "CPU", "low-level", "before", "one", "write", "We", "copy", "group", "system", "integer"], "path": "/documentation/en/develop/design/dist_train/mpi_enabled_design.html", "title": "MPI-enabled PaddlePaddle Design doc"}, {"content": ["all", "executed", "thread_pool", "atomic", "CPU", "parameters", "graph", "0", "Initialize", "input", "Wait", "then", "Execute", "propose", "endverbatim", "TODO", "Op", "execute", "user-defined", "Multi-CPU"], "path": "/documentation/en/develop/design/dist_train/multi_cpu.html", "title": "Design Doc: Execute the Program with Multi CPU"}, {"content": ["represent", "now", "code", "scheduler", "both", "Heterogeneous", "bandwidth", "Systems", "computation", "Async", "mentioned", "converting", "its", "previous", "layers", "heuristic", "We", "slow", "graph", "workers"], "path": "/documentation/en/develop/design/dist_train/parameter_server.html", "title": "Design Doc: Parameter Server"}, {"content": ["BATCH_BARRIER_MESSAGE", "all", "consider", "just", "synchronous", "go", "speed", "its", "parameters", "distributed", "Fluid", "send", "should", "0", "Using", "Now", "local", "then", "them", "handle"], "path": "/documentation/en/develop/design/dist_train/README.html", "title": "Distributed training overview doc"}, {"content": ["RNNOp", "RNN", "\u53d8\u957f\u8f93\u5165\u8bbe\u8ba1", "Dynamic", "design", "\u00b6"], "path": "/documentation/en/develop/design/dynamic_rnn/index_en.html", "title": "Dynamic RNN"}, {"content": ["default", "all", "step-scopes", "illustrates", "text", "128-dim", "dynamic", "128", "shape", "including", "using", "mentioned", "its", "previous", "whose", "We", "compose", "20", "parameters", "also"], "path": "/documentation/en/develop/design/dynamic_rnn/rnn.html", "title": "RNNOp design"}, {"content": ["\uff0c\u5e76\u4e14\u8c03\u7528", "\u6765\u4fdd\u8bc1\u4f20\u9012\u53ea\u8fdb\u884c\u4e00\u6b21", "producer", "\u5373\u5c06\u4e00\u4e2amini-batch\u5185\u4e0d\u540c\u957f\u5ea6\u7684\u5e8f\u5217\u88650\u5230\u56fa\u5b9a\u957f\u5ea6\u53c2\u4e0e\u8ba1\u7b97\u3002", "Bucketing", "not", "\u800c\u6240\u6709\u4f9d\u8d56\u5176", "\u7684\u4f20\u9012", "ShareConstLODFrom", "\u7684\u5b9e\u73b0\u4e2d\uff0c\u64cd\u4f5c\u66f4\u65b0\u81ea\u5df1\u7684", "size_t", "\u9700\u8981\u4fee\u6539", "sould", "boot_memories", "\u5fc5\u987b\u7528zero-padding\u7684\u65b9\u5f0f\u5c06\u53d8\u957f\u5e8f\u5217\u8865\u5168\u4e3a\u56fa\u5b9ashape\u7684tensor\u3002", "Tensor", "\u7b49\u5747\u4f7f\u7528\u4e86padding\u7684\u65b9\u5f0f\uff0c", "member", "typedef", "\u7684\u4fe1\u606f\u7684\u4f20\u9012\u5bf9\u975eOLD\u7684Op\u7684\u5b9e\u73b0\u662f\u5b8c\u5168\u900f\u660e\u7684\u3002"], "path": "/documentation/en/develop/design/dynamic_rnn/rnn_design.html", "title": "RNN \u53d8\u957f\u8f93\u5165\u8bbe\u8ba1"}, {"content": ["all", "chain", "existing", "sould", "const", "to", "Argument.subSequenceStartPositions", "mainstream", "void", "replication", "endverbatim", "level", "vector", "prevent", "consistent", "PaddlePaddle", "cost", "design", "pass", "implemented"], "path": "/documentation/en/develop/design/dynamic_rnn/rnn_design_en.html", "title": "Varient Length supported RNN Design"}, {"content": ["true_block", "shape", "identify", "differs", "20", "true", "either", "to", "add", "program", "usage", "bool", "mini-batch", "15", "Example", "condition", "endverbatim", "d", "follows", "scalar"], "path": "/documentation/en/develop/design/execution/if_else_op.html", "title": "The"}, {"content": ["Process", "Doc", "Switch", "\u00c2\u00b6", "Design", "Operator", "IfElse", "The", "Execution"], "path": "/documentation/en/develop/design/execution/index_en.html", "title": "Execution Process"}, {"content": ["control-flow", "scalar", "differs", "if-else", "We", "implementation", "Fluid", "should", "languages", "0", "add", "program", "usage", "matched", "easier", "Semantics", "which", "matches", "Many", "fluid"], "path": "/documentation/en/develop/design/execution/switch.html", "title": "Design Doc: Switch"}, {"content": ["Multi-Language", "Interface", "\u00c2\u00b6", "TBD"], "path": "/documentation/en/develop/design/interface/index_en.html", "title": "Multi-Language Interface"}, {"content": ["\u00c2\u00b6", "Management", "Optimization", "Memory"], "path": "/documentation/en/develop/design/memory/index_en.html", "title": "Memory Management"}, {"content": ["limited", "all", "go", "follow", "expressions", "graph", "technique", "to", "init", "program", "Not", "nodes", "very", "activation", "every", "formula", "endverbatim", "Reference", "large", "solved"], "path": "/documentation/en/develop/design/memory/memory_optimization.html", "title": "Memory Optimization"}, {"content": ["initial_pool_size", "cout", "max_pool_size", "both", "gpu_id", "Region-based", "not", "looks", "Usage", "CPU", "to", "fallback", "hold", "buddy", "parameters", "meta-data", "Free", "0", "only", "extends"], "path": "/documentation/en/develop/design/memory/README.html", "title": "Region-based Heterogeneous Memory Management"}, {"content": ["all", "code", "chain", "scalar", "block=s_block", "nested", "whose", "Backward", "calculate", "graph", "Args", "father", "Fluid", "to", "program", "marks", "Cases", "Now", "uninitialized", "updated"], "path": "/documentation/en/develop/design/modules/backward.html", "title": "Backward Building"}, {"content": ["Kernels", "What", "inferencing", "0", "Once", "suggests", "all", "var_cache", "assigned", "cudnn", "batch_var", "through", "contains", "computation", "Attributes", "go", "estimated_mean", "its", "previous", "As"], "path": "/documentation/en/develop/design/modules/batch_norm_op.html", "title": "Batch Normalization"}, {"content": ["and", "Regularization", "Network", "Gradient", "PaddlePaddle", "in", "Optimizer", "Python", "Operators", "Important", "Structure", "Building", "Prune", "InferVarType", "Doc", "Modules", "Backward", "Code", "\u00c2\u00b6", "API"], "path": "/documentation/en/develop/design/modules/index_en.html", "title": "Code Structure and Important Modules"}, {"content": ["just", "follow", "its", "We", "const", "struct", "implementation", "ease", "should", "to", "input", "them", "return", "void", "Register", "not", "using", "some_op", "replace", "endverbatim"], "path": "/documentation/en/develop/design/modules/infer_var_type.html", "title": "Design Doc: InferVarType"}, {"content": ["default", "all", "code", "Network.Run", "just", "neural", "hierarchy", "AddOptimization", "RecurrentNet", "stored", "attrs", "OpIndex", "managed", "label", "networks", "BaseNet", "member", "typedef", "unique", "const"], "path": "/documentation/en/develop/design/modules/net_op_design.html", "title": "Network Design"}, {"content": ["code", "0", "being", "cost_np", "op_desc.is_traget", "session", "attrs", "Var", "its", "count", "We", "const", "depended", "also", "situations", "foward", "add", "4", "bool", "into"], "path": "/documentation/en/develop/design/modules/prune.html", "title": "Prune"}, {"content": ["all", "code", "uniform_random", "global", "dynamic", "attrs", "follow", "layers", "param.stop_gradient=True", "to", "init", "program", "manipulates", "string", "activation", "endverbatim", "InferShape", "Global", "list", "fewer"], "path": "/documentation/en/develop/design/modules/python_api.html", "title": "Design Doc: Python API"}, {"content": ["now", "code", "just", "existing", "stored", "virtual", "computation", "consists", "minus_grad", "Developers", "its", "whose", "We", "rapid", "const", "struct", "field", "CreateGradientOperator", "should", "to"], "path": "/documentation/en/develop/design/modules/register_grad_op.html", "title": "Design Doc: Gradient Operators Registration"}, {"content": ["operations", "represent", "code", "background", "just", "spaces", "computation", "using", "implement", "CPU", "before", "layers", "adding", "We", "based", "parameters", "also", "implementation", "graph", "should"], "path": "/documentation/en/develop/design/modules/regularization.html", "title": "Regularization in PaddlePaddle"}, {"content": ["represent", "SelectedRow", "Selected", "shape", "So", "satisfies", "saved", "its", "row", "non-zero", "3,4", "written", "should", "to", "only", "4", "5", "has", "them", "string"], "path": "/documentation/en/develop/design/modules/selected_rows.html", "title": "Design Doc: Selected Rows"}, {"content": ["all", "code", "semantic", "global", "existing", "computation", "ParameterSet", "layers", "compose", "Kubernetes", "environment", "to", "program", "Not", "include", "sent", "Training", "updater", "decide", "read_from_realistic_images"], "path": "/documentation/en/develop/design/motivation/api.html", "title": "PaddlePaddle Design Doc"}, {"content": ["Existed", "all", "concept", "Based", "four", "layers", "Turing-complete", "graph", "Fluid", "removal", "languages", "to", "program", "very", "difference", "condition", "look", "endverbatim", "False", "PaddlePaddle"], "path": "/documentation/en/develop/design/motivation/fluid.html", "title": "Design Doc: PaddlePaddle Fluid"}, {"content": ["code", "vars", "global", "rest", "looks", "consists", "interpreter", "its", "before", "We", "const", "Generator", "generator", "instance", "distributed", "Fluid", "should", "0", "program", "input"], "path": "/documentation/en/develop/design/motivation/fluid_compiler.html", "title": "PaddlePaddle Fluid: Towards a Compiled Programming Language"}, {"content": ["Motivations", "and", "Towards", "Language", "Doc", "PaddlePaddle", "Fluid", "Refactorization", "\u00c2\u00b6", "Design", "Goals", "a", "Overview", "Programming", "Compiled"], "path": "/documentation/en/develop/design/motivation/index_en.html", "title": "Design Motivations and Goals"}, {"content": ["represent", "code", "migrated", "particular", "Macros", "Take", "computation", "consists", "CPU", "layers", "whose", "OpKernel", "OpWithKernel/Kernel", "Kubernetes", "graph", "send", "languages", "to", "program", "Registry"], "path": "/documentation/en/develop/design/motivation/refactorization.html", "title": "Design Doc: Refactorization Overview"}, {"content": ["Kernel", "Selection", "Hint", "Keys", "Doc", "Multi-Device", "\u00c2\u00b6", "Design", "of", "Operator", "The", "Type", "Support"], "path": "/documentation/en/develop/design/multi_devices/index_en.html", "title": "Multi-Device Support"}, {"content": ["all", "code", "global", "kernel_hint", "same", "kForceCPU", "CPU", "layer_helper", "layout", "struct", "fit", "also", "personal", "send", "should", "to", "add", "choose", "input", "GetExpectedKernelType"], "path": "/documentation/en/develop/design/multi_devices/kernel_hint_design.html", "title": "Kernel Hint Design"}, {"content": ["global", "find", "rest", "Take", "Var", "follow", "its", "library_type", "Inputs", "We", "layout", "struct", "situations", "new_scope", "Kernel", "CPU", "Fluid", "should", "better", "to"], "path": "/documentation/en/develop/design/multi_devices/kernel_selection.html", "title": "Kernel Selection"}, {"content": ["represent", "all", "supported", "mainly", "four", "course", "shape", "identify", "computation", "ifdef", "follow", "stride", "pointer", "CPU", "based", "member", "We", "now", "layout", "struct"], "path": "/documentation/en/develop/design/multi_devices/operator_kernel_type.html", "title": "Design Doc: The Keys of Operator Kernel Type"}, {"content": ["all", "WER/CER", "Improvement", "prefix", "computation", "Department", "row", "devoted", "batches", "fixed-length", "Standard", "to", "CTC-LM", "Not", "supported", "fully-connected", "DNNs", "very", "CTC-loss", "word"], "path": "/documentation/en/develop/design/network/deep_speech_2.html", "title": "DeepSpeech2 on PaddlePaddle: Design Doc"}, {"content": ["Generating", "Network", "Sequence", "LoDTensors", "\u00c2\u00b6", "Complex", "Design", "Decoder"], "path": "/documentation/en/develop/design/network/index_en.html", "title": "Complex Network Design"}, {"content": ["limited", "all", "semantic", "dict_size", "represent", "results", "prefix", "last_seq", "lookup", "Introducing", "Encoder", "prefixes", "to", "init", "Changing", "updated", "very", "activation", "translations", "source_dict_dim"], "path": "/documentation/en/develop/design/network/sequence_decoder.html", "title": "Design: Sequence Decoder Generating LoDTensors"}, {"content": ["all", "code", "fluid_model", "global", "dynamic", "focus", "four", "whose", "inverse", "graph", "Fluid", "to", "finally", "program", "nodes", "Now", "case", "mainstream", "very", "two-way"], "path": "/documentation/en/develop/design/onnx/onnx_convertor.html", "title": "Background"}, {"content": ["restore", "absolute", "accurate", "supported", "backward_op", "stability", "shape", "computation", "GPU", "CPP", "check_grad", "suffer", "CPU", "Calculate", "whose", "Theory", "instance", "cpu", "shoule", "should"], "path": "/documentation/en/develop/design/others/auto_gradient_check.html", "title": "Auto Gradient Check Design"}, {"content": ["all", "0", "github.com/rajathkmp/DCGAN", "if-else", "Convolutional", "graph", "to", "init", "tf", "001", "decide", "summarized", "G_h1_bn", "Declare", "endverbatim", "paddle.v2", "list", "load_mnist", "small", "the"], "path": "/documentation/en/develop/design/others/gan_api.html", "title": "Design for GAN"}, {"content": ["all", "code", "chain", "results", "1", "four", "looks", "computation", "Forward", "followed", "saved", "its", "fifth", "before", "layers", "layer.data", "ConstructOptimizationGraph", "vars", "parameters", "also"], "path": "/documentation/en/develop/design/others/graph.html", "title": "Design Doc: Computations as a Graph"}, {"content": ["represent", "all", "concept", "help", "produces", "name='fc1", "cg.print_graphviz", "global", "focus", "flatten0", "produced", "128", "session", "including", "same", "Expression", "follow", "label", "pointer", "find"], "path": "/documentation/en/develop/design/others/graph_survey.html", "title": "Survey on Graph"}, {"content": ["concept", "stores", "just", "from", "ParameterUpdater", "Multi-GPU", "multi-GPU/CPU", "code", "CPP", "its", "MultiGradientMachine", "before", "We", "const", "parameters", "implementation", "should", "to", "only", "Make"], "path": "/documentation/en/develop/design/others/parameters_in_cpp.html", "title": "Design Doc: The C++ Class"}, {"content": ["What", "code", "sigmoid", "all", "existing", "opdesc", "not", "elif", "AddInput", "OpProtoMaker", "attrs", "implement", "desc", "CosineOpProto", "\u201c", "low-level", "register", "layers", "1", "We"], "path": "/documentation/en/develop/design/others/simple_op_design.html", "title": "Interaction between C++ and Python"}, {"content": ["coresponding", "N-1", "less", "layer_norm", "results", "ParamAttr|list", "shape", "num_flatten_dims=1", "dynamic_lstm", "Take", "Under", "LSTM", "act=None", "its", "Tensor", "whose", "parameters", "omit", "Args", "should"], "path": "/documentation/en/develop/dev/api_doc_std_en.html", "title": "API Doc Standard"}, {"content": ["Git", "reviewers", "Fix", "code", "follow", "contributions", "Skipped", "Workflow", "Standard", "environment", "0", "helps", "7", "Fixes", "every", "upstream", "forked", "endverbatim", "level", "button"], "path": "/documentation/en/develop/dev/contribute_to_paddle_en.html", "title": "Contribute Code"}, {"content": ["Kernels", "\u2019", "Device/Library", "Convention", "PaddlePaddle", "Paddle", "How", "Add", "in", "operator", "New", "Parameter", "Development", "Supporting", "use", "for", "Format", "Operators", "Releasing", "Standard"], "path": "/documentation/en/develop/dev/index_en.html", "title": "Development"}, {"content": ["unified", "all", "show", "Comments", "AddInput", "Attributes", "follow", "Tensor", "We", "op_checker", "explain", "OpProtoAndCheckerMaker", "Here", "should", "examples", "to", "only", "listing", "usage", "input"], "path": "/documentation/en/develop/dev/name_convention.html", "title": "Operator's Parameter Name Convention"}, {"content": ["code", "device_context", "Describes", "attrs", "computation", "consists", "CPU", "OpKernel", "const", "constructor\u00ef\u00bc\u0161", "InferShapeContext", "0", "MulOp", "include", "string", "void", "VariableNameMap", "Where", "entire", "endverbatim"], "path": "/documentation/en/develop/dev/new_op_en.html", "title": "How to write a new operator"}, {"content": ["Kernels", "code", "GemmConvKernel", "your", "virtual", "using", "CPU", "before", "library_type", "resource", "We", "kCUDNN", "const", "struct", "implementation", "Kernel", "Fluid", "should", "to", "add"], "path": "/documentation/en/develop/dev/new_op_kernel.html", "title": "Add Kernels for a New Device"}, {"content": ["all", "words", "proposed", "rest", "relevant", "including", "mentioned", "follow", "paper/article", "Inputs", "should", "to", "other", "array_to_lod_tensor", "covered", "has", "them", "above", "get", "overall"], "path": "/documentation/en/develop/dev/op_markdown_format.html", "title": "Standard Markdown Format for Operators"}, {"content": ["all", "dist", "Artifacts", "major", "chapters", "feature", "not", "manages", "including", "merge", "Test", "follow", "Advise", "each", "its", "previous", "We", "Personalized", "Linear", "V1"], "path": "/documentation/en/develop/dev/releasing_process_en.html", "title": "PaddlePaddle Releasing Process"}, {"content": ["represent", "all", "consider", "Crop", "REGISTER_OP_CUDA_KERNEL", "demand", "mainly", "obtain", "However", "performance", "shape", "CropGradKernel", "cudaStream_t", "file", "using", "Integrate", "pointer", "ptr", "low-level", "CPUDeviceContext"], "path": "/documentation/en/develop/dev/support_new_device.html", "title": "Design Doc: Supporting new Device/Library"}, {"content": ["all", "code", "Using", "show", "being", "developed", "bring", "shape", "stored", "computation", "using", "Usage", "still", "pointer", "its", "note", "OpKernel", "input0/input1/output", "const", "Computation"], "path": "/documentation/en/develop/dev/use_eigen_en.html", "title": "How to use Eigen in Paddle"}, {"content": ["WITH_DOC=ON", "documentation\uff0cPlease", "code", "set", "produces", "find", "results", "processors", "installed", "indicates", "Release", "stored", "through", "PRs", "affect", "\u2018", "facilitates", "once", "its", "developer"], "path": "/documentation/en/develop/dev/write_docs_en.html", "title": "Contribute Documentation"}, {"content": ["\u00c2\u00b6", "FAQ"], "path": "/documentation/en/develop/faq/index_en.html", "title": "FAQ"}, {"content": ["and", "is", "PaddlePaddle", "an", "Start", "including", "want", "regression", "linear", "use", "GET", "STARTED", "format", "please", "Here", "quickly", "how", "Running", "program", "build"], "path": "/documentation/en/develop/getstarted/index_en.html", "title": "GET STARTED"}, {"content": ["operations", "limited", "all", "code", "just", "increasingly", "over", "move", "syntax", "through", "research", "networks", "before", "layers", "Think", "how", "implementation", "graph", "Fluid", "should"], "path": "/documentation/en/develop/getstarted/overview_en.html", "title": "Overview"}, {"content": ["feed", "MacOS", "code", "fluid_model", "installed", "housing.py", "paddle.dataset.uci_housing", "Scope", "should", "0", "voila", "details", "format", "python", "paddle.fluid", "fluid", "new", "Quick", "GPU", "CentOS"], "path": "/documentation/en/develop/getstarted/quickstart_en.html", "title": "Quick Start"}, {"content": ["all", "executed", "subsections", "syntax", "code", "through", "same", "concurrency", "Scope", "yes", "\u201c", "feedforward", "Aspects", "interesting", "epoch", "actually", "to", "program", "under", "input"], "path": "/documentation/en/develop/getstarted/running_a_program_en.html", "title": "Running a program"}, {"content": ["\u00c2\u00b6", "TBD", "Concepts"], "path": "/documentation/en/develop/getstarted/concepts/index_en.html", "title": "Concepts"}, {"content": ["code", "skip", "existing", "compose", "Returning", "to", "paddle.batch", "supported", "true_label", "advantage", "very", "true_image", "entries", "endverbatim", "False", "list", "try", "item", "PaddlePaddle", "second"], "path": "/documentation/en/develop/getstarted/concepts/reader/README.html", "title": "Python Data Reader Design Doc"}, {"content": ["all", "code", "stores", "0", "text", "computation", "consists", "TensorDesc.dims", "disk", "saved", "its", "bunch", "We", "parameters", "to", "version", "has", "Optional", "By", "string"], "path": "/documentation/en/develop/getstarted/concepts/save_model/model_format.html", "title": "Design Doc: Model Format"}, {"content": ["Performance", "HOW", "\u00c2\u00b6", "Optimization", "TO"], "path": "/documentation/en/develop/howto/index_en.html", "title": "HOW TO"}, {"content": ["all", "code", "DWITH_SWIG_PY", "layers", "Introducing", "Fluid", "0", "program", "train_reader", "exit", "nodes", "Now", "Training", "fully-connected", "paddle.fluid", "endverbatim", "list", "try", "PaddlePaddle", "192"], "path": "/documentation/en/develop/howto/cluster/fluid_cluster_train_en.html", "title": "Fluid Distributed Training"}, {"content": ["0", "InfiniBand", "ping", "supported", "installed", "libmlx4.so", "including", "7.4", "GPU", "on", "PADDLE_WORKERS", "find", "before", "Files", "1", "We", "Multiple", "www.mellanox.com/page/products_dyn", "everything", "somewhere"], "path": "/documentation/en/develop/howto/cluster/nccl2_rdma_training.html", "title": "Distributed Training with NCCL2 and RDMA"}, {"content": ["code", "focus", "go", "follow", "CPU", "ncalls", "home/yuyang/perf_test/.env/lib/python2.7/site-packages/paddle/fluid/framework.py:219", "to", "init", "program", "include", "Identifying", "very", "results", "libgoogle-perftools-dev", "Where", "endverbatim", "Directing", "p", "Go"], "path": "/documentation/en/develop/howto/optimization/cpu_profiling_en.html", "title": "Profiling the Python Code"}, {"content": ["profile", "do", "use", "Profiling", "how", "Python", "timeline", "tool", "Benchmark", "to", "\u00c2\u00b6", "Optimization", "Performance", "the", "Code"], "path": "/documentation/en/develop/howto/optimization/index_en.html", "title": "Performance Optimization"}, {"content": ["load", "code", "batches", "should", "to", "train_reader", "main", "details", "do", "python", "fluid", "not", "endverbatim", "like", "button", "cmd", "Warning", "'All", "timeline.py", "Run"], "path": "/documentation/en/develop/howto/optimization/timeline.html", "title": "how to use timeline tool to do profile"}, {"content": ["Training", "for", "vgg16", "Distributed", "Benchmark", "Cluster", "\u00c2\u00b6", "Performance"], "path": "/documentation/en/develop/howto/optimization/benchmark/index_en.html", "title": "Benchmark"}, {"content": ["5GB", "over", "trainer/pserver", "v1.6.2", "per", "CPU", "TN", "We", "20", "Kubernetes", "Benchmark", "Fluid", "Efficiency", "Cases", "By", "Training", "Setup", "mini-batch", "Batch", "Variable"], "path": "/documentation/en/develop/howto/optimization/benchmark/README.html", "title": "Cluster Training Benchmark"}, {"content": ["77.47", "59.20", "Fuild", "190.20", "detail", "Test", "CPU", "247.40", "GLOG_v=3", "20", "fix", "cache", "distributed", "Fluid", "should", "environment", "to", "add", "samples", "under"], "path": "/documentation/en/develop/howto/optimization/benchmark/vgg16/README.html", "title": "Performance for Distributed vgg16"}, {"content": ["represent", "all", "less", "being", "assigned", "has_key", "target", "employ", "Usage", "isinstance", "its", "before", "whose", "Notice", "myErrorClip", "instance", "Here", "should", "to", "only"], "path": "/documentation/en/develop/howto/performance/error_clip.html", "title": "Error Clip"}, {"content": ["xrange", "all", "executed", "object", "text", "accurate", "global", "synchronous", "speedup", "RecordEvent", "ending", "disabled", "timing", "code", "including", "during", "mentioned", "yep", "kCPU", "CPU"], "path": "/documentation/en/develop/howto/performance/profiler.html", "title": "Introduction"}, {"content": ["fwd_primitive", "all", "is_cpu_place", "prim", "help", "just", "dst", "device_context", "satisfied", "best", "four", "createMemory", "executed", "forward_training", "paddle", "GetLibrary", "Test", "follow", "Sulution", "fwd_PD"], "path": "/documentation/en/develop/howto/third_party/mkldnn_fluid.html", "title": "Design Doc: Add MKLDNN Kernel in Fluid Operator"}, {"content": ["operations", "all", "executed", "aggregate", "aggregating", "Transpiler", "over", "Multi-GPU", "synchronous", "rank0", "utilize", "assigned", "Cards", "bandwidth", "neighberhood", "during", "plug-in", "mentioned", "all-gather", "its"], "path": "/documentation/en/develop/howto/third_party/paddle_nccl.html", "title": "Design Doc: NCCL support in Paddle Fluid"}, {"content": ["and", "program", "smaller", "inference", "methods", "Mobilenet", "demo", "into", "some", "ENet", "library", "one", "How", "as", "minimum", "deploying", "Build", "file", "in", "Android"], "path": "/mobile/en/develop/README.html", "title": "Mobile"}, {"content": ["single-threaded", "1.0", "Mobilenet", "calculation", "Paddle", "Resnet", "Note", "7.0", "learning", "Input", "Android", "implemented", "3x224x224", "Benchmark", "Xiaomi", "5", "122.607", "ENet", "verbatim", "means"], "path": "/mobile/en/develop/benchmark/README.html", "title": "Paddle Benchmark"}, {"content": ["code", "inference.cc", "mentioned", "follow", "based", "environment", "to", "program", "adb", "include", "150528", "clang", "endverbatim", "follows", "API", "steps", "output", "For", "PaddlePaddle", "C"], "path": "/mobile/en/develop/benchmark/tool/C/README.html", "title": "Inference demo"}, {"content": ["demo/train.py", "process\u00e3\u20ac\u201a", "Mobilenet", "less", "being", "gradual", "0.9716", "participate", "Usage", "flower102", "layers", "increase", "We", "cut", "ratio", "parameters", "chosen", "cutted", "12M", "97.16"], "path": "/mobile/en/develop/model_compression/pruning/README.html", "title": "Pruning"}, {"content": ["Resnet18", "endverbatim", "used", "Classification", "verbatim", "Models", "of", "Image", "Paddle", "commonly", "MobileNet", "models", "Here", "file", "the", "config", "are"], "path": "/mobile/en/develop/models/README.html", "title": "Models"}, {"content": ["and", "Correctness", "layer", "30", "demo/mobilenet_without_bn.py", "convolution", "conv", "give", "process", "demo", "is", "demo/models/mobilenet_flowers102.tar.gz", "When", "python", "see", "in", "do_merge.sh", "acceleration", "NOTE", "We"], "path": "/mobile/en/develop/tool/merge_batch_normalization/README.html", "title": "Merge Batch Normarlization to fc or conv layer based on PaddlePaddle"}, {"content": ["represent", "96.96", "Mobilenet", "Rounding", "inverse-quantized", "BN", "its", "We", "inverse", "ratio", "parameters", "12M", "97.16", "0", "transformed", "input", "then", "endverbatim", "reduces", "performing"], "path": "/mobile/en/develop/tool/rounding/README.html", "title": "Rounding"}, {"content": ["\u81ea\u52a8\u95ee\u7b54\uff08Question", "\u662f\u4fe1\u606f\u68c0\u7d22\u548c\u641c\u7d22\u5f15\u64ce\u7814\u7a76\u7684\u6838\u5fc3\u95ee\u9898\u4e4b\u4e00\uff0c\u901a\u8fc7\u673a\u5668\u5b66\u4e60\u65b9\u6cd5\u5b66\u4e60\u4e00\u4e2a\u5206\u503c\u51fd\u6570\u5bf9\u5f85\u6392\u5e8f\u7684\u5019\u9009\u8fdb\u884c\u6253\u5206\uff0c\u518d\u6839\u636e\u5206\u503c\u7684\u9ad8\u4f4e\u786e\u5b9a\u5e8f\u5173\u7cfb\u3002\u6df1\u5ea6\u795e\u7ecf\u7f51\u7edc\u53ef\u4ee5\u7528\u6765\u5efa\u6a21\u5206\u503c\u51fd\u6570\uff0c\u6784\u6210\u5404\u7c7b\u57fa\u4e8e\u6df1\u5ea6\u5b66\u4e60\u7684LTR\u6a21\u578b\u3002", "\u6392\u5e8f\u5b66\u4e60", "CRF\uff09\u5728\u7279\u5f81\u57fa\u7840\u4e0a\u5b8c\u6210\u5e8f\u5217\u6807\u6ce8\u4efb\u52a1\uff0c\u9010\u6e10\u6210\u4e3a\u89e3\u51b3\u5e8f\u5217\u6807\u6ce8\u95ee\u9898\u7684\u6807\u914d\u89e3\u51b3\u65b9\u6848\u3002", "\u7684\u4ec5\u6709\u4e00\u4e9b\u503c\u4e3a0", "\u95ee\u9898-\u8bc1\u636e\uff08Evidence\uff09-\u7b54\u6848", "\u5728\u5e8f\u5217\u5230\u5e8f\u5217\u5b66\u4e60\u4efb\u52a1\u4e2d\uff0c\u6211\u4eec\u9996\u5148\u4ee5\u673a\u5668\u7ffb\u8bd1\u4efb\u52a1\u4e3a\u4f8b\uff0c\u63d0\u4f9b\u4e86\u591a\u79cd\u6539\u8fdb\u6a21\u578b\u4f9b\u5927\u5bb6\u5b66\u4e60\u548c\u4f7f\u7528\u3002\u5305\u62ec\uff1a\u4e0d\u5e26\u6ce8\u610f\u529b\u673a\u5236\u7684\u5e8f\u5217\u5230\u5e8f\u5217\u6620\u5c04\u6a21\u578b\uff0c\u8fd9\u4e00\u6a21\u578b\u662f\u6240\u6709\u5e8f\u5217\u5230\u5e8f\u5217\u5b66\u4e60\u6a21\u578b\u7684\u57fa\u7840\uff1b\u4f7f\u7528Scheduled", "\u4ecd\u6709\u4e00\u5b9a\u8ddd\u79bb\u3002\u4e3a\u4e86\u8861\u91cf\u73b0\u6709NLP\u6280\u672f\u5230", "\u63a8\u8350\u7cfb\u7edf", "\u57fa\u4e8ePairwise\u548cListwise\u7684\u6392\u5e8f\u5b66\u4e60", "11.2", "6.1", "11.3", "\u201c", "\u4e00\u8bfe\uff09\u3002", "10.1", "Sampling\u6539\u5584RNN\u6a21\u578b\u5728\u751f\u6210\u4efb\u52a1\u4e2d\u7684\u9519\u8bef\u7d2f\u79ef\u95ee\u9898\uff1b\u5e26\u5916\u90e8\u8bb0\u5fc6\u673a\u5236\u7684\u795e\u7ecf\u673a\u5668\u7ffb\u8bd1\uff0c\u901a\u8fc7\u589e\u5f3a\u795e\u7ecf\u7f51\u7edc\u7684\u8bb0\u5fc6\u80fd\u529b\uff0c\u6765\u5b8c\u6210\u590d\u6742\u7684\u5e8f\u5217\u5230\u5e8f\u5217\u5b66\u4e60\u4efb\u52a1\u3002\u9664\u673a\u5668\u7ffb\u8bd1\u4efb\u52a1\u4e4b\u5916\uff0c\u6211\u4eec\u4e5f\u63d0\u4f9b\u4e86\u4e00\u4e2a\u57fa\u4e8e\u6df1\u5c42LSTM\u7f51\u7edc\u751f\u6210\u53e4\u8bd7\u8bcd\uff0c\u5b9e\u73b0\u540c\u8bed\u8a00\u751f\u6210\u7684\u6a21\u578b\u3002", "\u57fa\u4e8e\u6df1\u5ea6\u56e0\u5b50\u5206\u89e3\u673a\u7684\u70b9\u51fb\u7387\u9884\u4f30\u6a21\u578b", "Residual", "to"], "path": "/models/en/develop/README.cn.html", "title": "models \u7b80\u4ecb"}, {"content": ["all", "semantic", "similarity", "focus", "poetry", "Convolutional", "word-vector", "Residual", "to", "finally", "under", "Training", "advantage", "bottleneck", "tagging", "endverbatim", "applicable", "Entity", "Listwise", "discrete"], "path": "/models/en/develop/README.html", "title": "Introduction to models"}, {"content": ["code", "0", "sentence", "paper", "consists", "pos_size", "data/src_dict", "Convolutional", "20", "use_gpu", "Modify", "train.py", "should", "to", "does", "src_dict_path", "2017", "Grangier", "test_data_path", "then"], "path": "/models/en/develop/conv_seq2seq/README.html", "title": "Convolutional Sequence to Sequence Learning"}, {"content": ["github.com/PaddlePaddle/Paddle/blob/develop/doc/api/v1/data_provider/pydataprovider2_en.rst", "all", "5", "14091123", "Generator", "Embedding\u5c06\u6bcf\u79cd\u503c\u6620\u5c04\u4e3a\u8fde\u7eed\u503c\u7684\u5411\u91cf\u3002", "\u4e2d", "app_id", "\u90e8\u5206\uff0c\u53ef\u4ee5\u8f93\u5165\u5f88", "not", "X", "\u9700\u8981\u5148\u626b\u63cf\u6570\u636e\u96c6\uff0c\u5f97\u5230\u8be5\u7c7b\u522b\u5bf9\u5e94\u7684\u9879\u96c6\u5408\uff0c\u4e4b\u540e\u624d\u80fd\u5f00\u59cb\u751f\u6210\u7279\u5f81\u3002", "Understanding", "\u786e\u5b9a\u8868\u793a\u7684\u6700\u5927\u7ef4\u5ea6", "features", "24", "concat_sparse_vectors", "21", "22", "\u6211\u4eec\u7684\u5b9e\u9a8c\u6570\u636e\u96c6"], "path": "/models/en/develop/ctr/dataset.html", "title": "\u6570\u636e\u53ca\u5904\u7406"}, {"content": ["\u6a21\u578b\u7684\u4f18\u52bf\u662f\u5bf9\u5927\u89c4\u6a21\u7a00\u758f\u7279\u5f81\u7684\u5bb9\u7eb3\u80fd\u529b\uff0c\u5305\u62ec\u5185\u5b58\u548c\u8ba1\u7b97\u91cf\u7b49\u65b9\u9762\uff0c\u5de5\u4e1a\u754c\u90fd\u6709\u975e\u5e38\u6210\u719f\u7684\u4f18\u5316\u65b9\u6cd5\uff1b", "\u800c\u6a21\u578b\u4e0b\u8fb9\u7684", "gz", "prefix", "\u5bf9\u5e94\u7684\u6570\u503c\u3002", "query", "\uff1a", "\u4e2a\u8f93\u5165\uff0c\u5206\u522b\u662f", "feeding", "use_gpu", "\u8bad\u597d\u7684\u6a21\u578b\u53ef\u4ee5\u7528\u6765\u9884\u6d4b\u65b0\u7684\u6570\u636e\uff0c", "test_data_path", "\u6570\u636e\u548c\u4efb\u52a1\u62bd\u8c61", "\u6a21\u578b\u7ed3\u6784\u5bf9\u6bd4", "0", "init", "\u8fdb\u884c\u5904\u7406\uff0c\u5177\u4f53\u4f7f\u7528\u65b9\u6cd5\u53c2\u8003\u5982\u4e0b\u8bf4\u660e\uff1a", "activation", "OUTPUT_DIR", "\u70b9\u51fb\u7387\u9884\u4f30"], "path": "/models/en/develop/ctr/README.cn.html", "title": "\u70b9\u51fb\u7387\u9884\u4f30"}, {"content": ["code", "results", "prefix", "go", "query", "\uff1a", "layers", "feeding", "Business", "use_gpu", "test_data_path", "Trees", "\u6570\u636e\u548c\u4efb\u52a1\u62bd\u8c61", "0", "init", "Training", "activation", "gz", "OUTPUT_DIR", "\uff1aClassification"], "path": "/models/en/develop/ctr/README.html", "title": "Click-Through Rate Prediction"}, {"content": ["code", "0", "proposed", "Ye", "both", "not", "paper", "paddle", "during", "90", "Criteo", "networks", "row", "raw", "1", "26", "data/raw", "based", "Linear", "parameters"], "path": "/models/en/develop/deep_fm/README.html", "title": "Deep Factorization Machine for Click-Through Rate prediction"}, {"content": ["semantic", "similarity", "\u888b", "prefix", "last_seq", "layers", "build_classification_or_regression_model", "test_data_path", "to", "\u4e4b\u95f4\u7684\u5927\u5c0f\u5173\u7cfb", "\u5206\u522b\u6253\u5206\u2014\u2014", "\u56de\u5f52", "P", "\u6574\u4f53\u4e0a\u7b97\u6761\u4ef6\u6982\u7387\u7528\u5bf9\u6570\u4f3c\u7136\u51fd\u6570\u4f5c\u4e3a\u635f\u5931\uff0c\u8fd9\u4e5f\u5c31\u662f\u56fe", "dim", "endverbatim", "0\u3002\u5b9e\u73b0\u8bf7\u53c2\u8003", "list", "Rank", "vector"], "path": "/models/en/develop/dssm/README.cn.html", "title": "\u6df1\u5ea6\u7ed3\u6784\u5316\u8bed\u4e49\u6a21\u578b (Deep Structured Semantic Models, DSSM)"}, {"content": ["code", "semantic", "similarity", "prefix", "last_seq", "query", "layers", "whose", "calculate", "test_data_path", "0", "finally", "Training", "activation", "P", "are\uff1a", "dim", "school", "level", "list"], "path": "/models/en/develop/dssm/README.html", "title": "Deep Structured Semantic Models (DSSM)"}, {"content": ["Currently", "Paddle", "are", "in", "still", "use", "codes", "APIs", "Models", "Fluid", "under", "development", "verbatim", "that", "models", "collection", "active", "The", "a", "endverbatim"], "path": "/models/en/develop/fluid/README.html", "title": "Paddle Fluid Models"}, {"content": ["code", "mnist_model.py", "results", "Tutorials", "ICLR", "contributions", "Keras", "Fluid", "to", "asterisks", "under", "adversary.py", "Moosavi-Dezfooli", "endverbatim", "Adversary", "Testing", "form", "small", "Szegedy", "MXNet"], "path": "/models/en/develop/fluid/adversarial/README.html", "title": "Advbox"}, {"content": ["precision:0.00474094141551", "recall:0.00727476217124", "\u5f00\u59cb\u9884\u6d4b\u3002", "precision:0", "\u7684", "\u8bad\u7ec3", "precision:0.0228873239437", "recall:0.00762112139358", "10-B", "precision:0.0120967741935", "endverbatim", "precision:0.000796178343949", "f1:0.00287769784173", "pass_id:2", "pass_id:0", "pass_id:1", "Test", "7793", "recall:0.0059880239521", "\u9884\u6d4b"], "path": "/models/en/develop/fluid/chinese_ner/README.html", "title": "\u4f7f\u7528ParallelExecutor\u7684\u4e2d\u6587\u547d\u540d\u5b9e\u4f53\u8bc6\u522b\u793a\u4f8b"}, {"content": ["and", "engine", "right", "tqdm", "one-dimensional", "Training", "is", "pass", "positioned", "back", "up", "How", "single", "Therefore", "in", "CartPole-v0", "The", "as", "mountain", "use"], "path": "/models/en/develop/fluid/DeepQNetwork/README.html", "title": "Reproduce DQN model"}, {"content": ["all", "code", "semantic", "paper", "pretrained_model", "still", "num_layers=50", "layers", "We", "password", "val_list.txt", "Untar", "train.py", "train/n04596742/n04596742_3032.jpeg", "under", "include", "into", "2012", "Training", "unzip"], "path": "/models/en/develop/fluid/image_classification/README.html", "title": "SE-ResNeXt for image classification"}, {"content": ["and", "code", "inference", "file", "Prepare", "show", "demo", "is", "models.caffe/alexnet", "caffemodel", "installed", "models.caffe/resnet", "How", "as", "stored", "Caffe", "trained", "in", "need", "converting"], "path": "/models/en/develop/fluid/image_classification/caffe2fluid/examples/imagenet/README.html", "title": "How to use"}, {"content": ["and", "show", "lenet.caffemodel", "demo", "eg", "How", "'mnist", "download", "converting", "models.caffe/lenet", "use", "prepare", "lenet.prototxt", "contains", "environment", "to", "caffe", "which", "lenet", "run"], "path": "/models/en/develop/fluid/image_classification/caffe2fluid/examples/mnist/README.html", "title": "How to use"}, {"content": ["ppl:335.877", "ppl:130.092", "ppl:211.507", "from", "ppl:203.581", "ppl:324.272", "hid_size", "model/epoch_6", "base_lr", "step:500", "ppl:364.404", "GPU", "\u5f53\u524d\u652f\u6301\u7684\u53c2\u6570\u53ef\u53c2\u89c1", "saved", "step:1400", "\u81ea\u5b9a\u4e49\u7f51\u7edc\u7ed3\u6784", "ppl:156.251", "20", "test_reader", "train.py"], "path": "/models/en/develop/fluid/language_model/README.html", "title": "\u8bed\u8a00\u6a21\u578b"}, {"content": ["all", "code", "Kaiser", "Systems", "still", "Jakob", "6000", "title", "implementation", "Fluid", "the", "Illia", "under", "2017", "L", "TODO", "6010", "endverbatim", "Noam", "arxiv"], "path": "/models/en/develop/fluid/neural_machine_translation/transformer/README.html", "title": "Attention is All You Need: A Paddle Fluid implementation"}, {"content": ["github.com/cocodataset/cocoapi.git", "Visualization", "code", "test/val", "'annotations/instances_minival2017.json", "skip", "global", "'coco2014", "own", "file", "eval_coco_map", "Test", "eval_coco_map.py", "pretrained_model", "VOC07+12", "ap_version", "We", "equation-0", "based", "'integral"], "path": "/models/en/develop/fluid/object_detection/README.html", "title": "SSD Object Detection"}, {"content": ["\u53ef\u67e5\u770b\u53c2\u6570\u8be6\u7ec6\u8bf4\u660e\u3002", "models/model_00052_15000", "help", "\u5728\u8be5\u811a\u672c\u4e2d\u5b9a\u4e49\u4e86\u8bad\u7ec3\u7f51\u7edc\u3001\u9884\u6d4b\u7f51\u7edc\u548cevaluate\u7f51\u7edc\u3002", "Character", "test_images", "\u8868\u793a\u4e2d\u6587\u5b57\u7b26\u9017\u53f7\u7684\u7d22\u5f15\u3002", "\u5bf9\u5e94\u7684label\u5982\u4e0b\u6240\u793a\uff1a", "\u5728\u8bc4\u4f30\u65f6\uff0c\u8bc4\u4f30\u96c6\u7684\u8def\u5f84\u901a\u8fc7eval.py\u7684\u9009\u9879", "\u2018", "data/test_images/00003.jpg", "input_images_dir", "\u4e0e\u8bad\u7ec3\u96c6\u7c7b\u4f3c", "\u4ece\u6807\u51c6\u8f93\u5165\u8bfb\u53d6\u4e00\u5f20\u56fe\u7247\u7684\u8def\u5f84\uff0c\u5e76\u5bf9\u9f50\u8fdb\u884c\u9884\u6d4b\uff1a", "\u4e24", "use_gpu", "train.py", "data/test.list", "\uff0c\u5176\u683c\u5f0f\u5982\u4e0b\u6240\u793a\uff1a", "\u4e2d\u5b9e\u73b0\u3002"], "path": "/models/en/develop/fluid/ocr_recognition/README.html", "title": "Optical Character Recognition"}, {"content": ["approx", "a_2", "a_1", "\u5927\u591a\u6570\u60c5\u51b5\uff0c\u6211\u4eec\u65e0\u6cd5\u7a77\u4e3e\u51fa\u6240\u6709\u7684", "r_t|s_t", "\u4e3a\u6b63\uff0c\u90a3\u4e48\u6211\u4eec\u5e0c\u671b\u5f53\u524d\u64cd\u4f5c\u7684\u51fa\u73b0\u7684\u6982\u7387", "\u5185\u5bb9\u5206\u4e3a", "entropy", "\u7684\u68af\u5ea6\u65b9\u5411\u66f4\u65b0", "\u2018", "\u6211\u4eec\u7684\u76ee\u6807\u51fd\u6570\u662f", "\u6211\u4eec\u600e\u4e48\u8bc4\u4f30\u4e00\u4e2aplayer", "1", "\u5f62\u5f0f\u5316\u8868\u793a\u5982\u4e0b\uff1a", "\u8868\u793a\u505a\u8fc7\u52a8\u4f5c", "\u4e3a\u540e\u9000\u7684\u6982\u7387\uff0c", "model\uff09", "42", "0", "4"], "path": "/models/en/develop/fluid/policy_gradient/README.html", "title": "Policy Gradient RL by PaddlePaddle"}, {"content": ["pass_id:127", "4.0867705", "being", "9585", "F1_score0.4473684", "caddick", "F1_score0.4097561", "3.6826708", "0.37846154", "\u201c", "TrainSet", "before", "4.1130385", "0.42424244", "pass_f1_score", "\u5728PaddlePaddle", "use_gpu", "discard", "train.py", "9560"], "path": "/models/en/develop/fluid/sequence_tagging_for_ner/README.html", "title": "\u547d\u540d\u5b9e\u4f53\u8bc6\u522b"}, {"content": ["0.850160", "0.026639", "0.137117", "bow_model/epoch11", "avg_cost", "\u4e2d\u81ea\u884c\u6dfb\u52a0\uff0c\u5e76\u8bbe\u7f6e", "0.951280", "\u5b66\u4e60\u7387\u5927\u5c0f", "lstm", "\u8bad\u7ec3\u6570\u636e", "24", "25", "26", "27", "20", "\u5728PaddlePaddle", "22", "23", "bow_model/epoch16", "28"], "path": "/models/en/develop/fluid/text_classification/README.html", "title": "\u6587\u672c\u5206\u7c7b"}, {"content": ["\u4e00", "\u4e0b\u8f7d\u539f\u59cb\u6570\u636e", "\u2502", "help", "\u674e\u767d", "10.4802", "data/dict.txt", "LSTM", "\u591a", "\u4e2d\u6536\u96c6\u7684\u5168\u5510\u8bd7\u4f5c\u4e3a\u8bad\u7ec3\u6570\u636e\uff0c\u5171\u6709\u7ea65.4\u4e07\u9996\u5510\u8bd7\u3002", "\u4e2d\u56fd\u53e4\u8bd7\u751f\u6210", "batch\u5927\u5c0f\uff0c\u9ed8\u8ba4\u4e3a1", "\u8bad\u7ec3\u6570\u636e\u8def\u5f84", "bi-directional", "data/raw", "use_gpu", "train.py", "to", "trainer\u6570\u76ee\uff0c\u9ed8\u8ba4\u4e3a1", "\u89c0\u7375"], "path": "/models/en/develop/generate_chinese_poetry/README.html", "title": "\u4e2d\u56fd\u53e4\u8bd7\u751f\u6210"}, {"content": ["\u2502", "words", "data/word_vocab.txt", "loss\uff1a\u5b9a\u4e49\u591a\u7c7b\u4ea4\u53c9\u71b5\u4f5c\u4e3a\u6a21\u578b\u7684\u635f\u5931\u51fd\u6570\u3002", "\u7b2c\u4e8c", "\u5047\u5982\u6307\u5b9a\u4e86", "\uff1a\u6307\u5b9a\u8f93\u51fa\u6587\u4ef6\u8def\u5f84\uff0c\u751f\u6210\u7ed3\u679c\u5c06\u5199\u5165\u6b64\u6587\u4ef6\u3002", "\u4e2d\u7684\u914d\u7f6e\uff0c\u7ec6\u8282\u548c\u9002\u914d\u5de5\u4f5c\u8be6\u60c5\u5982\u4e0b\uff1a", "\uff1a\u6307\u5b9a\u6bcf\u4e00\u53e5\u751f\u6210\u7684\u8bdd\u6700\u957f\u957f\u5ea6\uff0c\u5982\u679c\u6a21\u578b\u65e0\u6cd5\u751f\u6210\u51fa", "lstm", "\uff1a", "\u201c", "\u6784\u5efa\u8bcd\u5178\u65f6\uff0c\u4f1a\u81ea\u52a8\u52a0\u5165\u4e24\u4e2a\u7279\u6b8a\u7b26\u53f7\uff1a", "\u4e2d\u5b9a\u4e49\u4e86\u6a21\u578b\u7684\u53c2\u6570\u53d8\u91cf\u3002", "\u8bcd\u6027\u6807\u6ce8\u3001\u53e5\u6cd5\u5206\u6790\u3001\u8bed\u97f3\u8bc6\u522b", "1", "\u8bad\u7ec3\u6a21\u578b\u7684\u65b9\u5f0f\uff0c\u5373\uff1a\u628a\u6bcf\u4e2a\u6c49\u5b57\u5f53\u505a\u4e00\u4e2a\u8bcd\uff0c\u5e38\u7528\u6c49\u5b57\u4e5f\u5c31\u51e0\u5343\u4e2a\uff0c\u4f7f\u5f97\u5b57\u5178\u7684\u5927\u5c0f\u4e0d\u4f1a\u592a\u5927\u3001\u4e0d\u4f1a\u4e22\u5931\u592a\u591a\u4fe1\u606f\uff0c\u4f46\u6c49\u8bed\u4e2d\u540c\u4e00\u4e2a\u5b57\u5728\u4e0d\u540c\u8bcd\u4e2d\u8bed\u4e49\u76f8\u5dee\u5f88\u5927\uff0c\u6709\u65f6\u5bfc\u81f4\u6a21\u578b\u6548\u679c\u4e0d\u7406\u60f3\u3002\u5efa\u8bae\u591a\u8bd5\u8bd5\u3001\u6839\u636e\u5b9e\u9645\u60c5\u51b5\u9009\u62e9\u662f", "\u6309\u9700\u8c03\u6574", "Beam", "train_file"], "path": "/models/en/develop/generate_sequence_by_rnn_lm/README.html", "title": "\u4f7f\u7528\u5faa\u73af\u795e\u7ecf\u7f51\u8bed\u8a00\u6a21\u578b\u751f\u6210\u6587\u672c"}, {"content": ["\u8bf7\u66f4\u65b0", "and", "glove-path", "train.log", "image", "Reader", "year", "tee", "\u5168\u7403\u6807\u51c6\u5316\u9605\u8bfb\u5668", "cd", "infer.log", "\u9884\u5904\u7406\u5e76\u7279\u5f81\u5316\u6570\u636e\uff1a", "in", "featurize.py", "John", "\u5b89\u88c5", "data/glove.840B.300d.txt", "Jonathan", "data_dir", "Language"], "path": "/models/en/develop/globally_normalized_reader/README.cn.html", "title": "\u5168\u7403\u6807\u51c6\u5316\u9605\u8bfb\u5668"}, {"content": ["all", "code", "paper", "data/glove.840B.300d.txt", "title", "EMNLP", "train.py", "to", "2017", "then", "Training", "Installation", "get", "python", "Configurate", "modifying", "raiman2015gnr", "Inferring", "endverbatim", "necessary"], "path": "/models/en/develop/globally_normalized_reader/README.html", "title": "Globally Normalized Reader"}, {"content": ["02d", "\uff0c\u7136\u540e\u4f9d\u636e\u7a97\u53e3\u5927\u5c0f\uff08\u672c\u6587\u4e3a5\uff09\uff0c\u4ece\u5934\u5230\u5c3e\u6bcf\u6b21\u5411\u53f3\u6ed1\u52a8\u7a97\u53e3\u5e76\u751f\u6210\u4e00\u6761\u6570\u636e\u3002\u5982", "ngram_lm", "PTB\u6570\u636e", "\uff0c\u5206\u522b\u662f\u7b2c\u4e00\u4e2a\u5230\u7b2c\u56db\u4e2a\u8bcd\u5728\u5b57\u5178\u4e2d\u7684id\uff0cPaddlePaddle\u4f1a\u8fdb\u4e00\u6b65\u5c06\u8be5list\u8f6c\u5316\u6210", "predict_lbl", "infer.py", "results", "Param", "\u4f5c\u4e3a\u771f\u5b9e\u8bcd\uff0c\u65b9\u4fbf\u8bc4\u4f30\u3002\u89e3\u7801\u51fd\u6570\u7684\u8f93\u5165\u662f\u4e00\u4e2abatch\u6837\u672c\u7684\u9884\u6d4b\u6982\u7387\u4ee5\u53ca\u8bcd\u8868\u7684\u5927\u5c0f\uff0c\u91cc\u9762\u7684\u5faa\u73af\u662f\u5bf9\u6bcf\u6761\u6837\u672c\u7684\u8f93\u51fa\u6982\u7387\u8fdb\u884c\u89e3\u7801\uff0c\u89e3\u7801\u65b9\u5f0f\u5c31\u662f\u6309\u7167\u5de60\u53f31\u7684\u51c6\u5219\uff0c\u4e0d\u65ad\u904d\u5386\u8def\u5f84\uff0c\u76f4\u81f3\u5230\u8fbe\u53f6\u5b50\u8282\u70b9\u3002", "layer\u7684\u8f93\u5165\uff0c\u4e00\u4e2a\u5c01\u88c5\u6837\u4f8b\u5982\u4e0b\uff1a", "hidden_layer", "Bengio", "initial_std", "\u53c2\u6570\u542b\u4e49\u5982\u4e0b\uff1a", "label", "go", "\uff0c\u8be6\u7ec6\u7406\u8bba\u7ec6\u8282\u53ef\u53c2\u7167\u8bba\u6587", "\u53c2\u8003\u6587\u732e", "\u5b89\u88c5\u6587\u6863"], "path": "/models/en/develop/hsigmoid/README.html", "title": "Hsigmoid\u52a0\u901f\u8bcd\u5411\u91cf\u8bad\u7ec3"}, {"content": ["load", "'Paddle_ResNet50.tar.gz", "is", "Inception-ResNet-v2", "\u901a\u8fc7", "dropout_rate", "128000", "cross_entropy_cost", "Test", "init_from_tar", "\u6307\u5b9a\u8f93\u5165\u5c42", "\uff1a", "L2Regularization", "Label", "1", "EndIteration", "224", "test_reader", "parameters", "\u4f7f\u7528ResNet\u6a21\u578b"], "path": "/models/en/develop/image_classification/README.html", "title": "\u56fe\u50cf\u5206\u7c7b"}, {"content": ["fc_layer", "\u6982\u7387", "\u7684\u503c\u4f20\u5165\uff0c\u8fd9\u6837\u5728\u547d\u540d\u4fdd\u5b58layer\u4e2d\u7684\u53c2\u6570\u65f6\u5c06\u4f7f\u7528\u76f8\u5e94\u7684layer", "name_map", "\u4e2d\u7684\u503c\u548c\u8be5layer\u7c7b\u578b\u7684\u8c03\u7528\u8ba1\u6570\u6784\u9020layer", "\u5373\u53ef\uff08\u4e00\u65b9\u9762\uff0c\u5bf9\u4e8ePooling\u7b49\u65e0\u9700\u8bad\u7ec3\u7684layer\u4e0d\u9700\u8981\u4fdd\u5b58\uff0c\u6545\u8fd9\u91cc\u6ca1\u6709\u63d0\u4f9b\u8f6c\u6362\u63a5\u53e3\uff1b\u53e6\u4e00\u65b9\u9762\uff0c\u5bf9\u4e8eCaffe\u4e2d\u7684Scale\u7c7b\u522b\u7684layer\uff0c\u7531\u4e8eCaffe\u548cPaddlePaddle\u5728\u5b9e\u73b0\u4e0a\u7684\u4e00\u4e9b\u5dee\u522b\uff0cPaddlePaddle\u4e2d\u7684batch_norm\u5c42\u662fBatchNorm\u548cScale\u5c42\u7684\u590d\u5408\uff0c\u6545\u8fd9\u91cc\u5bf9Scale\u8fdb\u884c\u4e86\u7279\u6b8a\u5904\u7406\uff09\u3002", "img", "\u521d\u59cb\u5316\uff0c\u4ece\u6307\u5b9a\u6587\u4ef6\u52a0\u8f7d\u6a21\u578b", "\u5982\u679c\u4e0d\u5e0c\u671b\u4f7f\u7528\u9ed8\u8ba4\u7684\u547d\u540d\uff0c\u5e76\u4e14\u5728PaddlePaddle\u7f51\u7edc\u914d\u7f6e\u4e2d\u6307\u5b9a\u4e86layer", "\u4e2d\u5c01\u88c5\u4e86\u4f7f\u7528Caffe", "API\u52a0\u8f7d\u4f7f\u7528\u7684\u6a21\u578b\u6587\u4ef6\u3002", "\u3002", "\u4e2d\u5df2\u63d0\u4f9b\u4ee5\u4e0a\u6b65\u9aa4\uff0c\u4fee\u6539\u5176\u4e2d\u6587\u4ef6\u76f8\u5173\u53d8\u91cf\u7684\u503c\u540e\u6267\u884c", "wrap_name_default", "\u9700\u8981\u6ce8\u610f\uff0c\u5728\u6a21\u578b\u8f6c\u6362\u65f6\u4f1a\u5bf9layer\u7684\u53c2\u6570\u8fdb\u884c\u547d\u540d\uff0c\u8fd9\u91cc\u9ed8\u8ba4\u4f7f\u7528PaddlePaddle\u4e2d\u9ed8\u8ba4\u7684layer\u548c\u53c2\u6570\u547d\u540d\u89c4\u5219\uff1a\u4ee5", "\u4e3a\u6b64\uff0c\u5728\u9a8c\u8bc1\u548c\u4f7f\u7528\u8f6c\u6362\u5f97\u5230\u7684\u6a21\u578b\u65f6\uff0c\u7f16\u5199PaddlePaddle\u7f51\u7edc\u914d\u7f6e\u65e0\u9700\u6307\u5b9alayer", "imagenet/ilsvrc_2012_mean.npy", "\u8fdb\u884c\u6a21\u578b\u8f6c\u6362", "ModelConverter", "caffe_pretrained_file"], "path": "/models/en/develop/image_classification/caffe2paddle/README.html", "title": "\u4f7f\u7528\u8bf4\u660e"}, {"content": ["\u76f8\u5e94\u4f4d\u7f6e\u7684\u53c2\u6570\u3002", "\u4f20\u5165", "\uff0c\u4ee5\u6b64\u83b7\u53d6\u5bf9\u5e94", "\u7684\u987a\u5e8f\u4e3a\uff1a\u5148\u53ef\u5b66\u4e60", "saver", "restore", "paddle.layer.XX", "\u6a21\u578b\u8f6c\u6362\u7684\u57fa\u672c\u6d41\u7a0b\u662f\uff1a", "1", "Full", "\u7684Operator\u8bad\u7ec3\u51fa\u7684\u53c2\u6570\u5411PaddlePaddle\u6a21\u578b\u53c2\u6570\u8f6c\u6362\u3002", "\u7684\u540d\u5b57\u6620\u5c04\u4e3a\u6240\u5bf9\u5e94\u7684", "dict", "tf", "\u4e0b\u9762\u662f\u4e00\u4e2a\u7b80\u5355\u7684\u8c03\u7528\u793a\u4f8b\uff1a", "main", "return", "\u76ee\u524d\u4ec5\u652f\u6301\u5c06TensorFlow\u4e2d", "python", "Batch"], "path": "/models/en/develop/image_classification/tf2paddle/README.html", "title": "\u4f7f\u7528\u8bf4\u660e"}, {"content": ["partial", "relevance_score", "\u7684\u7279\u5f81\u7684\u7ef4\u5ea6\u3002\u6bcf\u6761\u8f93\u5165\u6837\u672c\u4e3alabel\uff0c", "K=Z", "1+e", "\u4f8b\u5982\u8c03\u7528\u63a5\u53e3", "\u63a8\u8350\u7cfb\u7edf", "paddle.layer.fc", "\uff0c\u662f\u7279\u5f81\u5411\u91cf\u7684\u5e8f\u5217\u8f93\u5165\uff0c", "left_doc", "gm", "\u603b\u7ed3", "feeding", "\uff0c\u5bf9\u4e8e\u7ed9\u5b9a\u7684\u4e00\u4e2a\u6587\u6863\u5bf9", "0", "1+S", "\u4e8b\u4ef6\u3002", "\uff0c\u5219", "\u6307\u4e0a\u4e00\u5c42\u4e2d\u7684\u6bcf\u4e2a\u8282\u70b9\u90fd\u8fde\u63a5\u5230\u4e0b\u5c42\u7f51\u7edc\u3002\u672c\u4f8b\u5b50\u4e2d\u540c\u6837\u4f7f\u7528", "\u7684\u65b9\u6cd5\u4f18\u5316\u5ea6\u91cf\u6807\u51c6NDCG"], "path": "/models/en/develop/ltr/README.html", "title": "\u6392\u5e8f\u5b66\u4e60(Learning To Rank)"}, {"content": ["\u9ed8\u8ba4\u63d0\u4f9b\u4e86\u4e00\u4e2a\u7ecf\u8fc7\u9884\u5904\u7406\u7684\u3001\u8f83\u5c0f\u89c4\u6a21\u7684", "limited", "data_utils.py", "dict_size", "LayerOutput", "V", "\u7eb8\u5e26", "RNNs\uff09", "Wayne", "WM^B", "\u57fa\u4e8e\u4f4d\u7f6e\u7684\u5bfb\u5740", "\uff0c\u4ee5\u4fdd\u6301\u548c", "LSTM", "\u901a\u8fc7\u57fa\u4e8e\u5185\u5bb9\u7684\u5bfb\u5740\uff0c\u8ba1\u7b97\u5f97\u5230\u8bfb\u5199\u64cd\u4f5c\u7684\u5bfb\u5740\u5f3a\u5ea6\u3002", "Memory\uff0c\u5373\u8bb0\u5fc6\u69fd\u6570\u91cf\u53ef\u53d8", "decoder_target", "Convolutional", "\u4e8e\u662f\uff0c\u6211\u4eec\u6df7\u5408\u4e0a\u8ff0\u7684\u4e09\u79cd\u52a8\u6001\u8bb0\u5fc6\u673a\u5236\uff0c\u5373RNN", "\u5176\u4ed6\u8ba8\u8bba", "\u6307\u7684\u662f\u5411\u91cf\u7ec4\u7684\u5411\u91cf\u4e2a\u6570\u975e\u56fa\u5b9a\uff0c\u800c\u662f\u968f\u7740\u6e90\u8bed\u53e5\u7684\u5b57\u7b26\u6570\u7684\u53d8\u5316\u800c\u53d8\u5316\uff0c\u6570\u91cf\u4e0d\u53d7\u9650\u3002\u5728\u6e90\u8bed\u53e5\u7684\u7f16\u7801\u5b8c\u6210\u65f6\uff0c\u8be5\u5916\u90e8\u5b58\u50a8\u5373\u88ab\u521d\u59cb\u5316\u4e3a\u5404\u5b57\u7b26\u7684\u72b6\u6001\u5411\u91cf\uff0c\u800c\u5728\u5176\u540e\u7684\u6574\u4e2a\u89e3\u7801\u8fc7\u7a0b\u4e2d\u88ab\u8bfb\u53d6\u4f7f\u7528\u3002"], "path": "/models/en/develop/mt_with_external_memory/README.html", "title": "\u5e26\u5916\u90e8\u8bb0\u5fc6\u673a\u5236\u7684\u795e\u7ecf\u673a\u5668\u7ffb\u8bd1"}, {"content": ["dict_size", "ptb", "omega_t|w", "\uff1a", "P_\\theta", "\u51fd\u6570\u6765\u523b\u753b\u5f0f", "use_gpu", "\u8bed\u8a00\u6a21\u578b\u662f\u8bb8\u591a\u81ea\u7136\u8bed\u8a00\u5904\u7406\u4efb\u52a1\u7684\u57fa\u7840\uff0c\u4e5f\u662f\u83b7\u5f97\u8bcd\u5411\u91cf\u8868\u793a\u7684\u4e00\u79cd\u6709\u6548\u65b9\u6cd5\u3002\u795e\u7ecf\u6982\u7387\u8bed\u8a00\u6a21\u578b\uff08Neural", "to", "\u5168\u8fde\u63a5\u9690\u5c42", "\uff09\u7684\u6982\u7387\u3002\u5047\u8bbe\u6765\u81ea\u566a\u58f0\u5206\u5e03\u7684\u8d1f\u7c7b\u6837\u672c\u7684\u6570\u91cf", "Training", "activation", "P", "\uff0c\u518d\u901a\u8fc7", "endverbatim", "unigram", "Probabilistic", "\u4e2a\u8d1f\u7c7b\u6837\u672c\u5b83\u4eec\u7684\u7c7b\u522b\u5206\u522b\u8bb0\u4f5c\uff1a", "\u9700\u8981\u6ce8\u610f\u7684\u662f\uff1a"], "path": "/models/en/develop/nce_cost/README.html", "title": "\u4f7f\u7528\u566a\u58f0\u5bf9\u6bd4\u4f30\u8ba1\u52a0\u901f\u8bed\u8a00\u6a21\u578b\u8bad\u7ec3"}, {"content": ["\u5e8f\u5217\u662f\u8bb8\u591a\u673a\u5668\u5b66\u4e60\u548c\u6570\u636e\u6316\u6398\u4efb\u52a1\u9762\u5bf9\u7684\u4e00\u79cd\u8f93\u5165\u6570\u636e\u7c7b\u578b\uff0c\u4ee5\u81ea\u7136\u8bed\u8a00\u5904\u7406\u4efb\u52a1\u4e3a\u4f8b\uff1a\u53e5\u5b50\u7531\u8bcd\u8bed\u6784\u6210\uff0c\u800c\u591a\u4e2a\u53e5\u5b50\u8fdb\u4e00\u6b65\u6784\u6210\u4e86\u6bb5\u843d\u3002\u56e0\u6b64\uff0c\u6bb5\u843d\u53ef\u4ee5\u770b\u4f5c\u662f\u4e00\u4e2a\u5d4c\u5957\u7684\u5e8f\u5217\uff08\u6216\u8005\u53eb\u4f5c\uff1a\u53cc\u5c42\u5e8f\u5217\uff09\uff0c\u8fd9\u4e2a\u5e8f\u5217\u7684\u6bcf\u4e2a\u5143\u7d20\u53c8\u662f\u4e00\u4e2a\u5e8f\u5217\u3002", "endverbatim", "verbatim", "\u7b80\u4ecb", "\u53cc\u5c42\u5e8f\u5217\u662f", "PaddlePaddle", "\u57fa\u4e8e\u53cc\u5c42\u5e8f\u5217\u7684\u6587\u672c\u5206\u7c7b", "\u80fd\u591f\u5e2e\u52a9\u6211\u4eec\u66f4\u597d\u5730\u63cf\u8ff0\u6bb5\u843d\u3001\u591a\u8f6e\u5bf9\u8bdd\u7b49\u66f4\u4e3a\u590d\u6742\u7684\u6570\u636e\u3002\u4ee5\u53cc\u5c42\u5e8f\u5217\u4f5c\u4e3a\u8f93\u5165\uff0c\u6211\u4eec\u53ef\u4ee5\u8bbe\u8ba1\u4e00\u4e2a\u5c42\u6b21\u5316\u7684\u7f51\u7edc\uff0c\u4ece\u800c\u66f4\u597d\u5730\u5b8c\u6210\u4e00\u4e9b\u590d\u6742\u7684\u4efb\u52a1\u3002", "\u652f\u6301\u7684\u4e00\u79cd\u975e\u5e38\u7075\u6d3b\u7684\u6570\u636e\u7ec4\u7ec7\u65b9\u5f0f\uff0c", "\u4e2d\u4f7f\u7528\u53cc\u5c42\u5e8f\u5217\u3002", "\u672c\u5355\u5143\u5c06\u4ecb\u7ecd\u5982\u4f55\u5728"], "path": "/models/en/develop/nested_sequence/README.html", "title": "\u7b80\u4ecb"}, {"content": ["default", "What", "cnn_cov_group", "0", "LayerOutput", "TEXT", "\u901a\u8fc7", "'_cov_value_bias", "label", "word_dict", "\u6307\u5b9a\u8bad\u7ec3\u914d\u7f6e\u53c2\u6570", "avg_pool", "doc_ids", "\uff1a", "\u7684\u903b\u8f91\u3002\u4ee5\u4e0b\u4ee3\u7801\u7247\u6bb5\u5b9e\u73b0\u4e86\u8bfb\u53d6\u539f\u59cb\u6570\u636e\u8fd4\u56de\u7c7b\u578b\u4e3a\uff1a", "networks", "help", "\u5c06\u53e5\u5b50\u62c6\u89e3\u4e3a\u8bcd\u8bed\uff0c", "group", "Linear"], "path": "/models/en/develop/nested_sequence/text_classification/README.html", "title": "\u57fa\u4e8e\u53cc\u5c42\u5e8f\u5217\u7684\u6587\u672c\u5206\u7c7b"}, {"content": ["WebQA", "code", "evidence", "Wei", "supported", "results", "Jie", "li:2016", "using", "CPU", "equation-0", "Zhou", "title", "should", "2016", "correct_chunks=2335", "arXiv:1607.06275", "Training", "Installation", "python"], "path": "/models/en/develop/neural_qa/README.html", "title": "Neural Recurrent Sequence Labeling Model for Open-Domain Factoid Question Answering"}, {"content": ["Van", "\u7b26\u53f7", "\u4e3a\u6bcf\u4e2a\u8f93\u5165\u6837\u672c\u751f\u6210", "skip", "\u7b2c\u4e8c", "PaddleBook\u4e2d", "last_seq", "LSTM", "if-else", "\u5bf9\u5e94\u7684\u76ee\u6807\u8bed\u8a00\u82f1\u6587\u7ffb\u8bd1\u7ed3\u679c\u4e3a\uff1a", "\u672c\u4f8b\u6240\u7528\u5230\u7684\u6570\u636e\u6765\u81ea", "0", "Step", "\u5728\u9884\u5904\u7406\u9636\u6bb5\uff0c\u51c6\u5907\u6e90\u8bed\u8a00\u4e0e\u76ee\u6807\u8bed\u8a00\u4e92\u8bd1\u7684\u5e73\u884c\u8bed\u6599\u6570\u636e\uff0c\u5e76\u5206\u522b\u6784\u5efa\u6e90\u8bed\u8a00\u548c\u76ee\u6807\u8bed\u8a00\u7684\u8bcd\u5178\uff1b\u5728\u8bad\u7ec3\u9636\u6bb5\uff0c\u7528\u8fd9\u6837\u6210\u5bf9\u7684\u5e73\u884c\u8bed\u6599\u8bad\u7ec3\u6a21\u578b\uff1b\u5728\u6a21\u578b\u6d4b\u8bd5\u9636\u6bb5\uff0c\u8f93\u5165\u4e2d\u6587\u53e5\u5b50\uff0c\u6a21\u578b\u81ea\u52a8\u751f\u6210\u5bf9\u5e94\u7684\u82f1\u8bed\u7ffb\u8bd1\uff0c\u7136\u540e\u5c06\u751f\u6210\u7ed3\u679c\u4e0e\u6807\u51c6\u7ffb\u8bd1\u5bf9\u6bd4\u8fdb\u884c\u8bc4\u4f30\u3002\u5728\u673a\u5668\u7ffb\u8bd1\u9886\u57df\uff0cBLEU", "end_pos", "decoder_group_name", "target_language_next_word", "\u884c\u662f\u67f1\u641c\u7d22\u751f\u6210\u7684", "activation", "nmt_without_att"], "path": "/models/en/develop/nmt_without_attention/README.cn.html", "title": "\u795e\u7ecf\u7f51\u7edc\u673a\u5668\u7ffb\u8bd1\u6a21\u578b"}, {"content": ["all", "code", "semantic", "illustrates", "global", "sub-sequence", "results", "single-step", "Take", "consists", "Van", "LSTM", "saved", "if-else", "row", "calculate", "graph", "Encodes", "subcomponents", "to"], "path": "/models/en/develop/nmt_without_attention/README.html", "title": "Neural Machine Translation Model"}, {"content": ["41G", "help", "TEXT", "infer.py", "Competitions", "not", "train_file_list", "filters", "PROPER", "'173,46", "1", "Challenge2_Training_Task3_Images_GT.zip", "group", "use_gpu", "requirements.txt", "train.py", "\u811a\u672c\u4e2d\u5305\u542b\u4e86\u6a21\u578b\u914d\u7f6e\u548c\u8bad\u7ec3\u76f8\u5173\u7684\u53c2\u6570\u4ee5\u53ca\u5bf9\u5e94\u7684\u8be6\u7ec6\u89e3\u91ca\uff0c\u4ee3\u7801\u7247\u6bb5\u5982\u4e0b\uff1a", "to", "model_save_dir", "8"], "path": "/models/en/develop/scene_text_recognition/README.html", "title": "\u573a\u666f\u6587\u5b57\u8bc6\u522b (STR, Scene Text Recognition)"}, {"content": ["iterator", "LayerOutput", "decay_a", "decay_b", "Scheduled", "\u548c\u76ee\u6807\u5e8f\u5217\u4e0b\u4e00\u4e2a\u5143\u7d20", "epsilon_i=max", "0", "init", "\uff0c\u5219", "\u4f5c\u4e3a\u89e3\u7801\u5668\u8f93\u5165\u3002", "\u4f5c\u4e3a\u751f\u6210\u7684\u8bcd\u8bed\u3002", "activation", "endverbatim", "Jaitly", "False", "Sampling\u9700\u8981\u5728\u5e8f\u5217\u5230\u5e8f\u5217\u6a21\u578b\u7684\u57fa\u7840\u4e0a\u589e\u52a0\u4e00\u4e2a\u8f93\u5165", "decoder_inputs", "vector", "\u5c42\u4f1a\u5728\u771f\u5b9e\u5143\u7d20"], "path": "/models/en/develop/scheduled_sampling/README.html", "title": "Scheduled Sampling"}, {"content": ["\u4e3a\u8fd0\u884c\u672c\u4f8b\uff0c\u8bf7\u9996\u5148\u5728", "\u2502", "from", "over", "infer.py", "Research", "four", "\u6211\u4eec\u5728", "london", "I-PER", "\u4e3a\u8bcd\u5178\u4e2d\u7684\u8bcd\u8bed\u63d0\u4f9b\u9884\u8bad\u7ec3\u597d\u7684\u8bcd\u5411\u91cf", "\uff0c\u672c\u4f8b\u53ea\u8003\u8651Segment", "Test", "\u201c", "Bottou", "1", "\u6587\u4ef6\u4e2d\u653e\u7f6e\u5c11\u6570\u6837\u672c\u7528\u4ee5\u793a\u4f8b\u8f93\u5165\u6570\u636e\u683c\u5f0f\u3002", "\u8fd9\u4e24\u4e2a\u6587\u4ef6\u3002", "parameters", "with"], "path": "/models/en/develop/sequence_tagging_for_ner/README.html", "title": "\u547d\u540d\u5b9e\u4f53\u8bc6\u522b"}, {"content": ["osp", "findall", "\u4f4d\u7f6e\u3002\u8bc4\u4f30\u7ed3\u675f\u4f1a\u8f93\u51fa", "\u8fd9\u91cc\u4e00\u6761\u6807\u6ce8\u6570\u636e\u5305\u62ec\uff1alabel\u3001xmin\u3001ymin\u3001xmax\u3001ymax\u548cis_difficult\uff0cis_difficult\u8868\u793a\u8be5object\u662f\u5426\u4e3a\u96be\u4f8b\uff0c\u5b9e\u9645\u4e2d\u5982\u679c\u4e0d\u9700\u8981\uff0c\u53ea\u9700\u628a\u8be5\u5b57\u6bb5\u7f6e\u96f6\u5373\u53ef\u3002\u81ea\u6709\u6570\u636e\u4e5f\u9700\u8981\u63d0\u4f9b\u5bf9\u5e94\u7684\u89e3\u6790\u903b\u8f91\uff0c\u5047\u8bbe\u6807\u6ce8\u6570\u636e\uff08\u6bd4\u5982image00000_annotation_file_path\uff09\u5b58\u50a8\u683c\u5f0f\u5982\u4e0b\uff1a", "\u4f5c\u4e3a\u57fa\u7840\u7f51\u7edc\u8fdb\u884c\u56fe\u50cf\u7279\u5f81\u63d0\u53d6\u3002\u4f46SSD\u5bf9\u539f\u59cbVGG16\u7f51\u7edc\u505a\u4e86\u4e00\u4e9b\u6539\u53d8\uff1a", "\u7c7b\u522b\u5217\u8868", "use_gpu", "Object", "0", "init", "size\u4e3a2\uff09\u66f4\u6539\u4e3a3x3-s1-p1\uff08kernel\u5927\u5c0f\u4e3a3x3\uff0cstride", "image00001_file_path", "\u5728\u81ea\u6709\u6570\u636e\u4e0a\u8bad\u7ec3PaddlePaddle", "\u53c2\u89c1\u8bad\u7ec3\u9636\u6bb5\u7684\u914d\u7f6e\uff1b\u8c03\u7528", "Classes", "\u9884\u8bad\u7ec3\u6a21\u578b\u51c6\u5907", "endverbatim", "walk_dir", "item", "data_provider.py"], "path": "/models/en/develop/ssd/README.cn.html", "title": "SSD\u76ee\u6807\u68c0\u6d4b"}, {"content": ["represent", "code", "osp", "results", "four", "rectangles", "CPU", "row", "layers", "use_gpu", "Object", "bbox", "environment", "0", "finally", "Now", "312.913876176", "image00001_file_path", "Training", "very"], "path": "/models/en/develop/ssd/README.html", "title": "Single Shot MultiBox Detector (SSD) Object Detection"}, {"content": ["all", "\u6a21\u578b\u5b66\u4e60\u6548\u679c\u8bc4\u4f30\u3002", "\u901a\u8fc7", "LSTM", "\uff1a", "N+1-h", "issues", "poorly", "\u5316\uff0c\u5373\u5c06\u8bcd\u8bed\u8f6c\u5316\u6210\u5728\u5b57\u5178\u4e2d\u7684\u5e8f\u53f7\u3002", "tv", "\u5bf9\u5377\u79ef\u5f97\u5230\u7684\u5404\u4e2a\u7279\u5f81\u56fe\u5206\u522b\u8fdb\u884c\u6700\u5927\u6c60\u5316\u64cd\u4f5c\u3002\u7531\u4e8e\u7279\u5f81\u56fe\u672c\u8eab\u5df2\u7ecf\u662f\u5411\u91cf\uff0c\u56e0\u6b64\u8fd9\u91cc\u7684\u6700\u5927\u6c60\u5316\u5b9e\u9645\u4e0a\u5c31\u662f\u7b80\u5355\u5730\u9009\u51fa\u5404\u4e2a\u5411\u91cf\u4e2d\u7684\u6700\u5927\u5143\u7d20\u3002\u5404\u4e2a\u6700\u5927\u5143\u7d20\u53c8\u88ab\u62fc\u63a5\u5728\u4e00\u8d77\uff0c\u7ec4\u6210\u65b0\u7684\u5411\u91cf\uff0c\u663e\u7136\uff0c\u8be5\u5411\u91cf\u7684\u7ef4\u5ea6\u7b49\u4e8e\u7279\u5f81\u56fe\u7684\u6570\u91cf\uff0c\u4e5f\u5c31\u662f\u5377\u79ef\u6838\u7684\u6570\u91cf\u3002\u4e3e\u4f8b\u6765\u8bf4\uff0c\u5047\u8bbe\u6211\u4eec\u4f7f\u7528\u4e86\u56db\u4e2a\u4e0d\u540c\u7684\u5377\u79ef\u6838\uff0c\u5377\u79ef\u4ea7\u751f\u7684\u7279\u5f81\u56fe\u5206\u522b\u4e3a\uff1a", "\u751f\u6210\u5668\u5b9e\u73b0", "seriously", "to", "\u8f93\u51fa\u65e5\u5fd7\u6bcf\u4e00\u884c\u662f\u5bf9\u4e00\u6761\u6837\u672c\u9884\u6d4b\u7684\u7ed3\u679c\uff0c\u4ee5", "train_reader", "paddle.batch", "\u5168\u8fde\u63a5\u9690\u5c42", "\u6a21\u578b\u8be6\u89e3", "very"], "path": "/models/en/develop/text_classification/README.html", "title": "\u6587\u672c\u5206\u7c7b"}, {"content": ["'_proj_history_clicked_tags", "\u662f\u6a21\u957f\u5f52\u4e00\u5316\u540e\u7684", "\u6574\u4e2a\u63a8\u8350\u7cfb\u7edf\u6709\u4e24\u90e8\u5206\u7ec4\u6210", "\u7684\u7c7b\u522b\uff0c", "feeding", "\u56e0\u4e3a\u5b9e\u65f6\u53ec\u56de\u9700\u8981\u5927\u91cf\u673a\u5668\u8d44\u6e90\uff0c\u8fd9\u8fb9\u4e5f\u53ef\u4ee5\u79bb\u7ebf\u6316\u6398\u4ea7\u51fa\u6570\u636e\uff0c\u7ebf\u4e0a\u53ec\u56de\u4f7f\u7528\u6316\u6398\u597d\u7684\u6570\u636e\u3002\u53ef\u4ee5\u4ea7\u51fa\u6700\u70ed\uff0c\u7528\u6237\u4e2a\u6027\u5316\uff0c\u89c6\u9891\u76f8\u5173\u7b49\u6570\u636e\u3002\u4e0b\u9762\u7684\u793a\u4f8b\u4ea7\u51fa\u4e86\u7528\u6237\u4e2a\u6027\u5316\u6570\u636e\u3002", "\u5386\u53f2\u70b9\u51fb\u89c6\u98911\uff0c\u5386\u53f2\u70b9\u51fb\u89c6\u98912", "Relu2", "Relu1", "window_size=2", "to", "\u672c\u6587\u5bf9", "h\uff0c", "lstm_last", "norm", "5Cfrac", "USER_ID_15", "activation", "reader", "\u7ebf\u4e0a\u4f7f\u7528\u65f6\uff0c\u4e3a\u4fdd\u7559\u7cbe\u5ea6\uff0c\u53ef\u4ee5\u4e0d\u9664\u4ee5"], "path": "/models/en/develop/youtube_recall/README.cn.html", "title": "Youtube DNN\u63a8\u8350\u6a21\u578b"}, {"content": ["represent", "all", "code", "similarity", "shuffle", "dynamic", "existing", "illustrated", "code.google.com/archive/p/word2vec", "initial_std", "consists", "query", "LSTM", "saved", "poorly", "layers", "feeding", "calculate", "prodct", "Relu2"], "path": "/models/en/develop/youtube_recall/README.html", "title": "Deep Neural Networks for YouTube Recommendations"}, {"path": "/api/en/develop/evaluator.html#evaluator", "prefix": "evaluator", "title": "evaluator"}, {"path": "/api/en/develop/evaluator.html#evaluator", "prefix": "evaluator", "title": "evaluator"}, {"path": "/api/en/develop/evaluator.html#chunkevaluator", "prefix": "evaluator", "title": "ChunkEvaluator"}, {"path": "/api/en/develop/evaluator.html#editdistance", "prefix": "evaluator", "title": "EditDistance"}, {"path": "/api/en/develop/evaluator.html#detectionmap", "prefix": "evaluator", "title": "DetectionMAP"}, {"path": "/api/en/develop/executor.html#executor", "prefix": "executor", "title": "executor"}, {"path": "/api/en/develop/executor.html#executor", "prefix": "executor", "title": "executor"}, {"path": "/api/en/develop/executor.html#id1", "prefix": "executor", "title": "Executor"}, {"path": "/api/en/develop/executor.html#global-scope", "prefix": "executor", "title": "global_scope"}, {"path": "/api/en/develop/executor.html#scope-guard", "prefix": "executor", "title": "scope_guard"}, {"path": "/api/en/develop/executor.html#switch-scope", "prefix": "executor", "title": "switch_scope"}, {"path": "/api/en/develop/initializer.html#initializer", "prefix": "initializer", "title": "initializer"}, {"path": "/api/en/develop/initializer.html#initializer", "prefix": "initializer", "title": "initializer"}, {"path": "/api/en/develop/initializer.html#constant", "prefix": "initializer", "title": "Constant"}, {"path": "/api/en/develop/initializer.html#uniform", "prefix": "initializer", "title": "Uniform"}, {"path": "/api/en/develop/initializer.html#normal", "prefix": "initializer", "title": "Normal"}, {"path": "/api/en/develop/initializer.html#xavier", "prefix": "initializer", "title": "Xavier"}, {"path": "/api/en/develop/initializer.html#msra", "prefix": "initializer", "title": "MSRA"}, {"path": "/api/en/develop/initializer.html#constantinitializer", "prefix": "initializer", "title": "ConstantInitializer"}, {"path": "/api/en/develop/initializer.html#uniforminitializer", "prefix": "initializer", "title": "UniformInitializer"}, {"path": "/api/en/develop/initializer.html#normalinitializer", "prefix": "initializer", "title": "NormalInitializer"}, {"path": "/api/en/develop/initializer.html#xavierinitializer", "prefix": "initializer", "title": "XavierInitializer"}, {"path": "/api/en/develop/initializer.html#msrainitializer", "prefix": "initializer", "title": "MSRAInitializer"}, {"path": "/api/en/develop/io.html#io", "prefix": "io", "title": "io"}, {"path": "/api/en/develop/io.html#io", "prefix": "io", "title": "io"}, {"path": "/api/en/develop/io.html#save-vars", "prefix": "io", "title": "save_vars"}, {"path": "/api/en/develop/io.html#save-params", "prefix": "io", "title": "save_params"}, {"path": "/api/en/develop/io.html#save-persistables", "prefix": "io", "title": "save_persistables"}, {"path": "/api/en/develop/io.html#load-vars", "prefix": "io", "title": "load_vars"}, {"path": "/api/en/develop/io.html#load-params", "prefix": "io", "title": "load_params"}, {"path": "/api/en/develop/io.html#load-persistables", "prefix": "io", "title": "load_persistables"}, {"path": "/api/en/develop/io.html#save-inference-model", "prefix": "io", "title": "save_inference_model"}, {"path": "/api/en/develop/io.html#load-inference-model", "prefix": "io", "title": "load_inference_model"}, {"path": "/api/en/develop/io.html#get-inference-program", "prefix": "io", "title": "get_inference_program"}, {"path": "/api/en/develop/layers.html#layers", "prefix": "layers", "title": "layers"}, {"path": "/api/en/develop/layers.html#layers", "prefix": "layers", "title": "layers"}, {"path": "/api/en/develop/layers.html#control-flow", "prefix": "layers", "title": "control_flow"}, {"path": "/api/en/develop/layers.html#split-lod-tensor", "prefix": "layers", "title": "split_lod_tensor"}, {"path": "/api/en/develop/layers.html#merge-lod-tensor", "prefix": "layers", "title": "merge_lod_tensor"}, {"path": "/api/en/develop/layers.html#blockguard", "prefix": "layers", "title": "BlockGuard"}, {"path": "/api/en/develop/layers.html#blockguardwithcompletion", "prefix": "layers", "title": "BlockGuardWithCompletion"}, {"path": "/api/en/develop/layers.html#staticrnnmemorylink", "prefix": "layers", "title": "StaticRNNMemoryLink"}, {"path": "/api/en/develop/layers.html#whileguard", "prefix": "layers", "title": "WhileGuard"}, {"path": "/api/en/develop/layers.html#while", "prefix": "layers", "title": "While"}, {"path": "/api/en/develop/layers.html#lod-rank-table", "prefix": "layers", "title": "lod_rank_table"}, {"path": "/api/en/develop/layers.html#max-sequence-len", "prefix": "layers", "title": "max_sequence_len"}, {"path": "/api/en/develop/layers.html#topk", "prefix": "layers", "title": "topk"}, {"path": "/api/en/develop/layers.html#lod-tensor-to-array", "prefix": "layers", "title": "lod_tensor_to_array"}, {"path": "/api/en/develop/layers.html#array-to-lod-tensor", "prefix": "layers", "title": "array_to_lod_tensor"}, {"path": "/api/en/develop/layers.html#increment", "prefix": "layers", "title": "increment"}, {"path": "/api/en/develop/layers.html#array-write", "prefix": "layers", "title": "array_write"}, {"path": "/api/en/develop/layers.html#create-array", "prefix": "layers", "title": "create_array"}, {"path": "/api/en/develop/layers.html#less-than", "prefix": "layers", "title": "less_than"}, {"path": "/api/en/develop/layers.html#array-read", "prefix": "layers", "title": "array_read"}, {"path": "/api/en/develop/layers.html#shrink-memory", "prefix": "layers", "title": "shrink_memory"}, {"path": "/api/en/develop/layers.html#array-length", "prefix": "layers", "title": "array_length"}, {"path": "/api/en/develop/layers.html#ifelse", "prefix": "layers", "title": "IfElse"}, {"path": "/api/en/develop/layers.html#dynamicrnn", "prefix": "layers", "title": "DynamicRNN"}, {"path": "/api/en/develop/layers.html#conditionalblock", "prefix": "layers", "title": "ConditionalBlock"}, {"path": "/api/en/develop/layers.html#staticrnn", "prefix": "layers", "title": "StaticRNN"}, {"path": "/api/en/develop/layers.html#reorder-lod-tensor-by-rank", "prefix": "layers", "title": "reorder_lod_tensor_by_rank"}, {"path": "/api/en/develop/layers.html#paralleldo", "prefix": "layers", "title": "ParallelDo"}, {"path": "/api/en/develop/layers.html#print", "prefix": "layers", "title": "Print"}, {"path": "/api/en/develop/layers.html#device", "prefix": "layers", "title": "device"}, {"path": "/api/en/develop/layers.html#get-places", "prefix": "layers", "title": "get_places"}, {"path": "/api/en/develop/layers.html#io", "prefix": "layers", "title": "io"}, {"path": "/api/en/develop/layers.html#data", "prefix": "layers", "title": "data"}, {"path": "/api/en/develop/layers.html#blockguardserv", "prefix": "layers", "title": "BlockGuardServ"}, {"path": "/api/en/develop/layers.html#listenandserv", "prefix": "layers", "title": "ListenAndServ"}, {"path": "/api/en/develop/layers.html#send", "prefix": "layers", "title": "Send"}, {"path": "/api/en/develop/layers.html#nn", "prefix": "layers", "title": "nn"}, {"path": "/api/en/develop/layers.html#fc", "prefix": "layers", "title": "fc"}, {"path": "/api/en/develop/layers.html#embedding", "prefix": "layers", "title": "embedding"}, {"path": "/api/en/develop/layers.html#dynamic-lstm", "prefix": "layers", "title": "dynamic_lstm"}, {"path": "/api/en/develop/layers.html#dynamic-lstmp", "prefix": "layers", "title": "dynamic_lstmp"}, {"path": "/api/en/develop/layers.html#dynamic-gru", "prefix": "layers", "title": "dynamic_gru"}, {"path": "/api/en/develop/layers.html#gru-unit", "prefix": "layers", "title": "gru_unit"}, {"path": "/api/en/develop/layers.html#linear-chain-crf", "prefix": "layers", "title": "linear_chain_crf"}, {"path": "/api/en/develop/layers.html#crf-decoding", "prefix": "layers", "title": "crf_decoding"}, {"path": "/api/en/develop/layers.html#cos-sim", "prefix": "layers", "title": "cos_sim"}, {"path": "/api/en/develop/layers.html#cross-entropy", "prefix": "layers", "title": "cross_entropy"}, {"path": "/api/en/develop/layers.html#square-error-cost", "prefix": "layers", "title": "square_error_cost"}, {"path": "/api/en/develop/layers.html#accuracy", "prefix": "layers", "title": "accuracy"}, {"path": "/api/en/develop/layers.html#chunk-eval", "prefix": "layers", "title": "chunk_eval"}, {"path": "/api/en/develop/layers.html#sequence-conv", "prefix": "layers", "title": "sequence_conv"}, {"path": "/api/en/develop/layers.html#conv2d", "prefix": "layers", "title": "conv2d"}, {"path": "/api/en/develop/layers.html#sequence-pool", "prefix": "layers", "title": "sequence_pool"}, {"path": "/api/en/develop/layers.html#pool2d", "prefix": "layers", "title": "pool2d"}, {"path": "/api/en/develop/layers.html#batch-norm", "prefix": "layers", "title": "batch_norm"}, {"path": "/api/en/develop/layers.html#layer-norm", "prefix": "layers", "title": "layer_norm"}, {"path": "/api/en/develop/layers.html#beam-search-decode", "prefix": "layers", "title": "beam_search_decode"}, {"path": "/api/en/develop/layers.html#conv2d-transpose", "prefix": "layers", "title": "conv2d_transpose"}, {"path": "/api/en/develop/layers.html#sequence-expand", "prefix": "layers", "title": "sequence_expand"}, {"path": "/api/en/develop/layers.html#lstm-unit", "prefix": "layers", "title": "lstm_unit"}, {"path": "/api/en/develop/layers.html#reduce-sum", "prefix": "layers", "title": "reduce_sum"}, {"path": "/api/en/develop/layers.html#reduce-mean", "prefix": "layers", "title": "reduce_mean"}, {"path": "/api/en/develop/layers.html#reduce-max", "prefix": "layers", "title": "reduce_max"}, {"path": "/api/en/develop/layers.html#reduce-min", "prefix": "layers", "title": "reduce_min"}, {"path": "/api/en/develop/layers.html#sequence-first-step", "prefix": "layers", "title": "sequence_first_step"}, {"path": "/api/en/develop/layers.html#sequence-last-step", "prefix": "layers", "title": "sequence_last_step"}, {"path": "/api/en/develop/layers.html#dropout", "prefix": "layers", "title": "dropout"}, {"path": "/api/en/develop/layers.html#split", "prefix": "layers", "title": "split"}, {"path": "/api/en/develop/layers.html#ctc-greedy-decoder", "prefix": "layers", "title": "ctc_greedy_decoder"}, {"path": "/api/en/develop/layers.html#edit-distance", "prefix": "layers", "title": "edit_distance"}, {"path": "/api/en/develop/layers.html#l2-normalize", "prefix": "layers", "title": "l2_normalize"}, {"path": "/api/en/develop/layers.html#matmul", "prefix": "layers", "title": "matmul"}, {"path": "/api/en/develop/layers.html#warpctc", "prefix": "layers", "title": "warpctc"}, {"path": "/api/en/develop/layers.html#sequence-reshape", "prefix": "layers", "title": "sequence_reshape"}, {"path": "/api/en/develop/layers.html#transpose", "prefix": "layers", "title": "transpose"}, {"path": "/api/en/develop/layers.html#im2sequence", "prefix": "layers", "title": "im2sequence"}, {"path": "/api/en/develop/layers.html#nce", "prefix": "layers", "title": "nce"}, {"path": "/api/en/develop/layers.html#beam-search", "prefix": "layers", "title": "beam_search"}, {"path": "/api/en/develop/layers.html#row-conv", "prefix": "layers", "title": "row_conv"}, {"path": "/api/en/develop/layers.html#multiplex", "prefix": "layers", "title": "multiplex"}, {"path": "/api/en/develop/layers.html#label-smooth", "prefix": "layers", "title": "label_smooth"}, {"path": "/api/en/develop/layers.html#roi-pool", "prefix": "layers", "title": "roi_pool"}, {"path": "/api/en/develop/layers.html#ops", "prefix": "layers", "title": "ops"}, {"path": "/api/en/develop/layers.html#mean", "prefix": "layers", "title": "mean"}, {"path": "/api/en/develop/layers.html#mul", "prefix": "layers", "title": "mul"}, {"path": "/api/en/develop/layers.html#reshape", "prefix": "layers", "title": "reshape"}, {"path": "/api/en/develop/layers.html#pad", "prefix": "layers", "title": "pad"}, {"path": "/api/en/develop/layers.html#scale", "prefix": "layers", "title": "scale"}, {"path": "/api/en/develop/layers.html#sigmoid-cross-entropy-with-logits", "prefix": "layers", "title": "sigmoid_cross_entropy_with_logits"}, {"path": "/api/en/develop/layers.html#elementwise-add", "prefix": "layers", "title": "elementwise_add"}, {"path": "/api/en/develop/layers.html#elementwise-div", "prefix": "layers", "title": "elementwise_div"}, {"path": "/api/en/develop/layers.html#elementwise-sub", "prefix": "layers", "title": "elementwise_sub"}, {"path": "/api/en/develop/layers.html#elementwise-mul", "prefix": "layers", "title": "elementwise_mul"}, {"path": "/api/en/develop/layers.html#elementwise-max", "prefix": "layers", "title": "elementwise_max"}, {"path": "/api/en/develop/layers.html#elementwise-min", "prefix": "layers", "title": "elementwise_min"}, {"path": "/api/en/develop/layers.html#elementwise-pow", "prefix": "layers", "title": "elementwise_pow"}, {"path": "/api/en/develop/layers.html#clip", "prefix": "layers", "title": "clip"}, {"path": "/api/en/develop/layers.html#clip-by-norm", "prefix": "layers", "title": "clip_by_norm"}, {"path": "/api/en/develop/layers.html#sequence-softmax", "prefix": "layers", "title": "sequence_softmax"}, {"path": "/api/en/develop/layers.html#sigmoid", "prefix": "layers", "title": "sigmoid"}, {"path": "/api/en/develop/layers.html#logsigmoid", "prefix": "layers", "title": "logsigmoid"}, {"path": "/api/en/develop/layers.html#exp", "prefix": "layers", "title": "exp"}, {"path": "/api/en/develop/layers.html#relu", "prefix": "layers", "title": "relu"}, {"path": "/api/en/develop/layers.html#tanh", "prefix": "layers", "title": "tanh"}, {"path": "/api/en/develop/layers.html#tanh-shrink", "prefix": "layers", "title": "tanh_shrink"}, {"path": "/api/en/develop/layers.html#softshrink", "prefix": "layers", "title": "softshrink"}, {"path": "/api/en/develop/layers.html#sqrt", "prefix": "layers", "title": "sqrt"}, {"path": "/api/en/develop/layers.html#abs", "prefix": "layers", "title": "abs"}, {"path": "/api/en/develop/layers.html#ceil", "prefix": "layers", "title": "ceil"}, {"path": "/api/en/develop/layers.html#floor", "prefix": "layers", "title": "floor"}, {"path": "/api/en/develop/layers.html#round", "prefix": "layers", "title": "round"}, {"path": "/api/en/develop/layers.html#reciprocal", "prefix": "layers", "title": "reciprocal"}, {"path": "/api/en/develop/layers.html#log", "prefix": "layers", "title": "log"}, {"path": "/api/en/develop/layers.html#square", "prefix": "layers", "title": "square"}, {"path": "/api/en/develop/layers.html#softplus", "prefix": "layers", "title": "softplus"}, {"path": "/api/en/develop/layers.html#softsign", "prefix": "layers", "title": "softsign"}, {"path": "/api/en/develop/layers.html#brelu", "prefix": "layers", "title": "brelu"}, {"path": "/api/en/develop/layers.html#leaky-relu", "prefix": "layers", "title": "leaky_relu"}, {"path": "/api/en/develop/layers.html#soft-relu", "prefix": "layers", "title": "soft_relu"}, {"path": "/api/en/develop/layers.html#elu", "prefix": "layers", "title": "elu"}, {"path": "/api/en/develop/layers.html#relu6", "prefix": "layers", "title": "relu6"}, {"path": "/api/en/develop/layers.html#pow", "prefix": "layers", "title": "pow"}, {"path": "/api/en/develop/layers.html#stanh", "prefix": "layers", "title": "stanh"}, {"path": "/api/en/develop/layers.html#hard-shrink", "prefix": "layers", "title": "hard_shrink"}, {"path": "/api/en/develop/layers.html#thresholded-relu", "prefix": "layers", "title": "thresholded_relu"}, {"path": "/api/en/develop/layers.html#hard-sigmoid", "prefix": "layers", "title": "hard_sigmoid"}, {"path": "/api/en/develop/layers.html#swish", "prefix": "layers", "title": "swish"}, {"path": "/api/en/develop/layers.html#tensor", "prefix": "layers", "title": "tensor"}, {"path": "/api/en/develop/layers.html#create-tensor", "prefix": "layers", "title": "create_tensor"}, {"path": "/api/en/develop/layers.html#create-parameter", "prefix": "layers", "title": "create_parameter"}, {"path": "/api/en/develop/layers.html#create-global-var", "prefix": "layers", "title": "create_global_var"}, {"path": "/api/en/develop/layers.html#cast", "prefix": "layers", "title": "cast"}, {"path": "/api/en/develop/layers.html#concat", "prefix": "layers", "title": "concat"}, {"path": "/api/en/develop/layers.html#sums", "prefix": "layers", "title": "sums"}, {"path": "/api/en/develop/layers.html#assign", "prefix": "layers", "title": "assign"}, {"path": "/api/en/develop/layers.html#fill-constant-batch-size-like", "prefix": "layers", "title": "fill_constant_batch_size_like"}, {"path": "/api/en/develop/layers.html#fill-constant", "prefix": "layers", "title": "fill_constant"}, {"path": "/api/en/develop/layers.html#ones", "prefix": "layers", "title": "ones"}, {"path": "/api/en/develop/layers.html#zeros", "prefix": "layers", "title": "zeros"}, {"path": "/api/en/develop/layers.html#id1", "prefix": "layers", "title": "topk"}, {"path": "/api/en/develop/nets.html#nets", "prefix": "nets", "title": "nets"}, {"path": "/api/en/develop/nets.html#nets", "prefix": "nets", "title": "nets"}, {"path": "/api/en/develop/nets.html#simple-img-conv-pool", "prefix": "nets", "title": "simple_img_conv_pool"}, {"path": "/api/en/develop/nets.html#sequence-conv-pool", "prefix": "nets", "title": "sequence_conv_pool"}, {"path": "/api/en/develop/nets.html#glu", "prefix": "nets", "title": "glu"}, {"path": "/api/en/develop/nets.html#scaled-dot-product-attention", "prefix": "nets", "title": "scaled_dot_product_attention"}, {"path": "/api/en/develop/optimizer.html#optimizer", "prefix": "optimizer", "title": "optimizer"}, {"path": "/api/en/develop/optimizer.html#optimizer", "prefix": "optimizer", "title": "optimizer"}, {"path": "/api/en/develop/optimizer.html#sgd", "prefix": "optimizer", "title": "SGD"}, {"path": "/api/en/develop/optimizer.html#momentum", "prefix": "optimizer", "title": "Momentum"}, {"path": "/api/en/develop/optimizer.html#adagrad", "prefix": "optimizer", "title": "Adagrad"}, {"path": "/api/en/develop/optimizer.html#adam", "prefix": "optimizer", "title": "Adam"}, {"path": "/api/en/develop/optimizer.html#adamax", "prefix": "optimizer", "title": "Adamax"}, {"path": "/api/en/develop/optimizer.html#decayedadagrad", "prefix": "optimizer", "title": "DecayedAdagrad"}, {"path": "/api/en/develop/optimizer.html#sgdoptimizer", "prefix": "optimizer", "title": "SGDOptimizer"}, {"path": "/api/en/develop/optimizer.html#momentumoptimizer", "prefix": "optimizer", "title": "MomentumOptimizer"}, {"path": "/api/en/develop/optimizer.html#adagradoptimizer", "prefix": "optimizer", "title": "AdagradOptimizer"}, {"path": "/api/en/develop/optimizer.html#adamoptimizer", "prefix": "optimizer", "title": "AdamOptimizer"}, {"path": "/api/en/develop/optimizer.html#adamaxoptimizer", "prefix": "optimizer", "title": "AdamaxOptimizer"}, {"path": "/api/en/develop/optimizer.html#decayedadagradoptimizer", "prefix": "optimizer", "title": "DecayedAdagradOptimizer"}, {"path": "/api/en/develop/optimizer.html#adadelta", "prefix": "optimizer", "title": "Adadelta"}, {"path": "/api/en/develop/regularizer.html#regularizer", "prefix": "regularizer", "title": "regularizer"}, {"path": "/api/en/develop/regularizer.html#regularizer", "prefix": "regularizer", "title": "regularizer"}, {"path": "/api/en/develop/regularizer.html#append-regularization-ops", "prefix": "regularizer", "title": "append_regularization_ops"}, {"path": "/api/en/develop/regularizer.html#l1decay", "prefix": "regularizer", "title": "L1Decay"}, {"path": "/api/en/develop/regularizer.html#l2decay", "prefix": "regularizer", "title": "L2Decay"}, {"path": "/api/en/develop/regularizer.html#l1decayregularizer", "prefix": "regularizer", "title": "L1DecayRegularizer"}, {"path": "/api/en/develop/regularizer.html#l2decayregularizer", "prefix": "regularizer", "title": "L2DecayRegularizer"}]
