+++ portal/portal/static/sass/navigation.sass
@@ -447,8 +447,8 @@ body
 // From https://github.com/bassjobsen/typeahead.js-bootstrap4-css/blob/master/typeaheadjs.css.
 span.twitter-typeahead .tt-menu
     cursor: pointer
-    width: 98%    // Bad way to account for paddings.
-
+    // Bad way to account for paddings.
+    width: 98%
 
 .dropdown-menu, span.twitter-typeahead .tt-menu
     position: absolute
@@ -459,7 +459,6 @@ span.twitter-typeahead .tt-menu
     float: left
     min-width: 160px
     padding: 5px 0
-    //margin: 2px 0 0
     font-size: 1rem
     color: #373a3c
     text-align: left
@@ -470,9 +469,6 @@ span.twitter-typeahead .tt-menu
     border-radius: 0.25rem
 
 span.twitter-typeahead .tt-suggestion
-    //display: block
-    // width: 100%
-    // padding: 3px 20px
     padding: 4px 0px
     margin: 0px
     clear: both
