+++ portal/portal/templatetags/app_tags.py
@@ -22,56 +22,6 @@ from portal import url_helper
 
 register = template.Library()
 
-
-@register.assignment_tag(takes_context=False)
-def get_dict_item(dictionary, key):
-    return dictionary.get(key)
-
-
-@register.simple_tag(takes_context=True)
-def translation(context, leaf_node):
-    """
-    The leaf node of the sitemap.json could be a dictionary of a string
-    When encountering a dictionary leaf node, load the value associated with the current language code
-    """
-    result = None
-
-    if isinstance(leaf_node, basestring):
-        result = leaf_node
-    elif isinstance(leaf_node, dict):
-        current_lang_code = context.request.LANGUAGE_CODE
-
-        if current_lang_code in leaf_node:
-            result = leaf_node[current_lang_code]
-
-    return result
-
-
-@register.assignment_tag(takes_context=True)
-def translation_assignment(context, leaf_node):
-    return translation(context, leaf_node)
-
-
-@register.assignment_tag(takes_context=True)
-def first_book_url_assignment(context, book, content_id):
-    # Finds the first url in the book in the default category
-    if book and 'default-category' in book:
-        category = book['default-category']
-
-        if content_id == portal_helper.Content.DOCUMENTATION or \
-           content_id == portal_helper.Content.API:
-            # For Documentation or API, we also filter by category
-            api_category = context.get('CURRENT_API_VERSION', None)
-            if api_category:
-                category = api_category
-
-        leaf_node = book['categories'][category]
-        if ('link' in leaf_node):
-            return translation(context, leaf_node['link'])
-
-    return ''
-
-
 @register.simple_tag(takes_context=True)
 def apply_class_if_template(context, template_file_name, class_name):
     '''
@@ -99,7 +49,6 @@ def nav_bar(context):
         current_lang_code
     )
 
-
     # TODO[thuan]: This is kinda hacky, need to find better way of removing visualdl docs from PPO
     if 'visualdl' in root_navigation:
         root_navigation.pop('visualdl')
@@ -126,17 +75,15 @@ def content_links(context, content_id):
     current_lang_code = context.request.LANGUAGE_CODE
     docs_version = context.get('CURRENT_DOCS_VERSION', None)
 
-    #side_nav_content, category = sitemap_helper.get_content_navigation(
     side_nav_content = sitemap_helper.get_content_navigation(
         context.request,
         content_id,
+        current_lang_code,
         docs_version,
-        current_lang_code
     )
 
     return _common_context(context, {
         'side_nav_content': side_nav_content,
-        # 'category': category,
         'allow_search': context.get('allow_search', False),
         'allow_version': context.get('allow_version', False),
         'search_url': context.get('search_url', None)
@@ -145,45 +92,12 @@ def content_links(context, content_id):
 
 @register.inclusion_tag('_version_links.html', takes_context=True)
 def version_links(context):
-    # if content_id == portal_helper.Content.DOCUMENTATION or \
-    #                 content_id == portal_helper.Content.API:
-    #     # API section needs to be additonally filtered by API Version
-    #     versions = _get_api_version_to_paddle_versions(content_id)
-    # else:
-    #     versions = [
-    #         {
-    #             "versions": sitemap_helper.get_available_versions(content_id)
-    #         }
-    #     ]
-
-    # Make a copy because otherwise settings persist in the process.
     versions = list(settings.VERSIONS)
     versions.reverse()
 
     return _common_context(context, {
         'versions':versions,
     })
-
-
-def _get_api_version_to_paddle_versions(content_id):
-    versions = sitemap_helper.get_available_versions(content_id)
-
-    fluid_min_version = '0.12.0'
-    v1v2_min_version = '0.9.0'  # TODO: Implement upper bounds for v1v2 once its deprecated
-
-    api_version_to_paddle_version = [
-        {
-            'key': 'fluid',
-            'title': 'Fluid',
-            'versions': [v for v in versions if sitemap_helper.is_version_greater_eq(v, fluid_min_version)]
-        },
-        {
-            'key': 'v2/v1',
-            'title': 'V2/V1',
-            'versions': [v for v in versions if sitemap_helper.is_version_greater_eq(v, v1v2_min_version)]
-        }
-    ]
-    return api_version_to_paddle_version
 
 
 def _common_context(context, additional_context):
@@ -201,3 +115,27 @@ def _common_context(context, additional_context):
     })
 
     return additional_context
+
+
+@register.assignment_tag(takes_context=True)
+def translation_assignment(context, leaf_node):
+    return translation(context, leaf_node)
+
+
+@register.simple_tag(takes_context=True)
+def translation(context, leaf_node):
+    """
+    The leaf node of the sitemap.json could be a dictionary of a string
+    When encountering a dictionary leaf node, load the value associated with the current language code
+    """
+    result = None
+
+    if isinstance(leaf_node, basestring):
+        result = leaf_node
+    elif isinstance(leaf_node, dict):
+        current_lang_code = context.request.LANGUAGE_CODE
+
+        if current_lang_code in leaf_node:
+            result = leaf_node[current_lang_code]
+
+    return result
