+++ portal/portal/urls.py
@@ -42,7 +42,7 @@ urlpatterns = [
     # HOME PAGE URLS
     # ---------------
     url(r'^$', views.home_root, name='home'),
-    url(r'^zh', views.cn_home_root, name='cn_home'),
+    url(r'^zh', views.zh_home_root, name='zh_home'),
 
     # ---------------
     # BLOG URLS
@@ -50,19 +50,6 @@ urlpatterns = [
     # NOTE: Temporary remove the links to Blog. Wait until we have more contents.
     # url(r'^%s$' % url_helper.BLOG_ROOT, views.blog_root, name=url_helper.URL_NAME_BLOG_ROOT),
     # url(r'^%s(?P<path>.+html)$' % url_helper.BLOG_ROOT, views.blog_sub_path),
-
-
-    # ---------------
-    # CONTENT URLS
-    # ---------------
-    url(r'^documentation$', views.content_home, name='documentation'),
-    url(r'^api$', views.content_home, name='api'),
-    url(r'^book$', views.content_home, name='book'),
-    url(r'^models$', views.content_home, name='models'),
-    url(r'^mobile$', views.content_home, name='mobile'),
-
-    # url(r'^docs/(?P<version>[^/]+)/(?P<path>[^./]+)/?$', views.content_root_path, name=url_helper.URL_NAME_CONTENT_ROOT),
-    # url(r'^docs/(?P<version>[^/]+)/(?P<path>.*)$', views.content_sub_path, name=url_helper.URL_NAME_CONTENT),
 
     # ---------------
     # ACTION URLS
@@ -73,16 +60,14 @@ urlpatterns = [
     url(r'^get-menu$', views.get_menu, name='get_menu'),
     url(r'^save-menu$', views.save_menu, name='save_menu'),
 
-    url(r'^download_latest_doc_workspace$', views.download_latest_doc_workspace, name='download_latest_doc_workspace'),
-
-    # -------------------
-    # OTHER ARBITRARY URLS
-    # -------------------
-    url(r'^docs/(?P<version>.*)/other/(?P<path>.*)$', views.other_path, name=url_helper.URL_NAME_OTHER),
-    url(r'^docs/(?P<version>.*)/flush$', views.flush_other_page, name='flush_other_page'),
-    # url(r'^book$', views.book_home, name=url_helper.URL_NAME_BOOK_ROOT),
-
-
+    # ---------------
+    # CONTENT URLS
+    # ---------------
+    url(r'^documentation$', views.content_home, name='documentation'),
+    url(r'^api$', views.content_home, name='api'),
+    url(r'^book$', views.content_home, name='book'),
+    url(r'^models$', views.content_home, name='models'),
+    url(r'^mobile$', views.content_home, name='mobile'),
     # HAS TO BE KEPT IN THE END, because it picks everything.
     url(r'^(?P<path>.*)$', views.content_sub_path, name='content'),
 ]
