+++ scripts/deploy/legacy/generate_book_docs.sh
