+++ scripts/travis/deploy-image.sh
@@ -10,7 +10,6 @@ then
     export DOCKER_CONTAINER_NAME="paddlepaddle.org"
     export PORT=80
     export ENV=release
-    export CONTENT_DIR="/var/content"
 elif [[ "$TRAVIS_BRANCH" =~ ^release.*$ ]]
 then
     # Staging Deploy
@@ -19,7 +18,6 @@ then
     export DOCKER_CONTAINER_NAME="staging.paddlepaddle.org"
     export PORT=81
     export ENV=release
-    export CONTENT_DIR="/var/content_staging"
 elif [ "$TRAVIS_BRANCH" == "develop" ]
 then
     # Development Deploy
@@ -28,7 +26,6 @@ then
     export DOCKER_CONTAINER_NAME="develop.paddlepaddle.org"
     export PORT=82
     export ENV=development
-    export CONTENT_DIR="/var/content_staging"
 else
     # All other branches should be ignored
     echo "Cannot deploy image, invalid branch: $TRAVIS_BRANCH"
